% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{abs.tensor}
\alias{abs.tensor}
\title{Abs}
\usage{
\method{abs}{tensor}(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns absolute values of tensor elements.
}
\examples{
x <- tensor(c(-1,1))
abs(x)
}
