% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{any.tensor}
\alias{any.tensor}
\title{any}
\usage{
\method{any}{tensor}(x, dim = -1, keepdim = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{dim}{if negative (the default) will reduce to a scalar. Otherwise it will
return TRUE if all elements in each row of the tensor in the given dimension
`dim` are TRUE, FALSE otherwise. **Note** that `dim` is 1 based like in R.}

\item{keepdim}{If keepdim is TRUE, the output tensor is of the same size as
input except in the dimension dim where it is of size 1. Otherwise, dim is
squeezed [squeeze()], resulting in the output tensor having 1 fewer
dimension than input.}

\item{na.rm}{won't be used by the function. Only there to be compatible with
[all] generic.}
}
\description{
any
}
\examples{
x <- tensor(array(c(TRUE, FALSE, TRUE, TRUE), dim = c(2, 2)))
any(x)
any(x, dim = 1)
any(x, dim = 1, keepdim = FALSE)
any(x, dim = 2)
}
