% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{as.array.tensor}
\alias{as.array.tensor}
\title{as.array}
\usage{
\method{as.array}{tensor}(x)
}
\arguments{
\item{x}{torch::Tensor object to be casted to an R array.}
}
\description{
Casts a `torch::Tensor` to an R array
}
\examples{
x <- tensor(array(1:8, dim = c(2, 2, 2)))
as.array(x)

}
