% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.matrix.tensor}
\alias{as.matrix.tensor}
\title{Casts a 2d tensor to a matrix.}
\usage{
\method{as.matrix}{tensor}(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Casts a 2d tensor to a matrix.
}
\examples{
x <- tensor(as.matrix(mtcars))
as.matrix(x)
}
\seealso{
\code{\link[=as.array.tensor]{as.array.tensor()}}
}
