% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{as_strided.tensor}
\alias{as_strided.tensor}
\title{as_strided}
\usage{
\method{as_strided}{tensor}(x, size, stride, storage_offset = NULL)
}
\arguments{
\item{x}{tensor object}

\item{size}{size}

\item{stride}{stride}

\item{storage_offset}{(optional) storage_offset}
}
\description{
TODO: create better docs.
}
\examples{
x <- tensor(array(runif(8), dim = c(2,2,2)))
as_strided(x, 0, 1)
}
