% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{autograd_set_grad_mode}
\alias{autograd_set_grad_mode}
\title{Set grad mode}
\usage{
autograd_set_grad_mode(enabled)
}
\arguments{
\item{enabled}{bool wether to enable or disable the gradient recording.}
}
\description{
Sets or disables gradient history.
}
