% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{backends_mkl_is_available}
\alias{backends_mkl_is_available}
\title{MKL is available}
\usage{
backends_mkl_is_available()
}
\value{
Returns whether LibTorch is built with MKL support.
}
\description{
MKL is available
}
