% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{backends_mkldnn_is_available}
\alias{backends_mkldnn_is_available}
\title{MKLDNN is available}
\usage{
backends_mkldnn_is_available()
}
\value{
Returns whether LibTorch is built with MKL-DNN support.
}
\description{
MKLDNN is available
}
