% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{backends_mps_is_available}
\alias{backends_mps_is_available}
\title{MPS is available}
\usage{
backends_mps_is_available()
}
\value{
Returns whether LibTorch is built with MPS support.
}
\description{
MPS is available
}
