% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mps.R
\name{backends_mps_synchronize}
\alias{backends_mps_synchronize}
\title{Waits for all kernels in all streams on the MPS device to complete.}
\usage{
backends_mps_synchronize()
}
\description{
Waits for all kernels in all streams on the MPS device to complete.
}
