% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backends.R
\name{backends_openmp_is_available}
\alias{backends_openmp_is_available}
\title{OpenMP is available}
\usage{
backends_openmp_is_available()
}
\value{
Returns whether LibTorch is built with OpenMP support.
}
\description{
OpenMP is available
}
