% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{bernoulli.tensor}
\alias{bernoulli.tensor}
\title{bernoulli}
\usage{
\method{bernoulli}{tensor}(x, p = NULL)
}
\arguments{
\item{x}{tensor object}

\item{p}{probability (if null uses tensor values)}
}
\description{
bernoulli
}
\examples{
x <- tensor(runif(10))
bernoulli(x)

x <- tensor(rep(0, 10))
bernoulli(x)
}
