% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{bmm.tensor}
\alias{bmm.tensor}
\title{bmm}
\usage{
\method{bmm}{tensor}(x, mat2)
}
\description{
Performs a batch matrix-matrix product of matrices stored in batch1 and
batch2.
}
\examples{
x <- tensor(array(runif(120), dim = c(10, 3, 4)))
y <- tensor(array(runif(200), dim = c(10, 4, 5)))
bmm(x, y)
}
