% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuda.R
\name{cuda_empty_cache}
\alias{cuda_empty_cache}
\title{Empty cache}
\usage{
cuda_empty_cache()
}
\description{
Releases all unoccupied cached memory currently held by the caching allocator
so that those can be used in other GPU application and visible in \code{nvidia-smi}.
}
\note{
\code{\link[=cuda_empty_cache]{cuda_empty_cache()}} doesn’t increase the amount of GPU memory available
for torch. However, it may help reduce fragmentation of GPU memory in certain
cases. See Memory management article for more details about GPU memory management.
}
