% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuda.R
\name{cuda_get_device_capability}
\alias{cuda_get_device_capability}
\title{Returns the major and minor CUDA capability of \code{device}}
\usage{
cuda_get_device_capability(device)
}
\arguments{
\item{device}{Integer value of the CUDA device to return capabilities of.}
}
\description{
Returns the major and minor CUDA capability of \code{device}
}
