% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cuda.R
\name{cuda_is_available}
\alias{cuda_is_available}
\title{Returns a bool indicating if CUDA is currently available.}
\usage{
cuda_is_available()
}
\description{
Returns a bool indicating if CUDA is currently available.
}
