% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-dataloader.R
\name{dataloader_make_iter}
\alias{dataloader_make_iter}
\title{Creates an iterator from a DataLoader}
\usage{
dataloader_make_iter(dataloader)
}
\arguments{
\item{dataloader}{a dataloader object.}
}
\description{
Creates an iterator from a DataLoader
}
