% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{dataset_subset}
\alias{dataset_subset}
\title{Dataset Subset}
\usage{
dataset_subset(dataset, indices)
}
\arguments{
\item{dataset}{(Dataset): The whole Dataset}

\item{indices}{(sequence): Indices in the whole set selected for subset}
}
\description{
Subset of a dataset at specified indices.
}
