% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtype.R
\name{torch_set_default_dtype}
\alias{torch_set_default_dtype}
\alias{torch_get_default_dtype}
\title{Gets and sets the default floating point dtype.}
\usage{
torch_set_default_dtype(d)

torch_get_default_dtype()
}
\arguments{
\item{d}{The default floating point dtype to set. Initially set to
\code{torch_float()}.}
}
\description{
Gets and sets the default floating point dtype.
}
\concept{tensor-attributes}
