% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-constraints.R
\docType{data}
\name{dependent}
\alias{dependent}
\title{Public interface
TODO: check .GreaterThan and other classes,
which are not instanced}
\format{
An object of class \code{torch_Dependent} (inherits from \code{torch_Constraint}, \code{R6}) of length 4.
}
\usage{
dependent
}
\description{
Public interface
TODO: check .GreaterThan and other classes,
which are not instanced
}
\keyword{datasets}
