% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-enum.R
\name{enum}
\alias{enum}
\title{Dataloader enum}
\usage{
enum(dataloader, max_len = 1e+06)
}
\arguments{
\item{dataloader}{the dataloader to enumerate.}

\item{max_len}{maximum number of iterations.}
}
\description{
Dataloader enum
}
