% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-enum.R
\name{enumerate.dataloader}
\alias{enumerate.dataloader}
\title{Enumerate an iterator}
\usage{
\method{enumerate}{dataloader}(x, max_len = 1e+06, ...)
}
\arguments{
\item{x}{the generator to enumerate.}

\item{max_len}{maximum number of iterations.}

\item{...}{passed to specific methods.}
}
\description{
Enumerate an iterator
}
