% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{==.tensor}
\alias{==.tensor}
\title{Eq}
\usage{
\method{==}{tensor}(x, other)
}
\arguments{
\item{x}{tensor object}

\item{other}{the tensor or value to compare}
}
\description{
Computes element-wise equality
The second argument can be a number or a tensor whose shape is broadcastable
with the first argument.
}
\examples{
x <- tensor(1:10)
y <- tensor(c(1:5, 10:6))
x == y

}
