% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{get_install_libs_url}
\alias{get_install_libs_url}
\title{List of files to download}
\usage{
get_install_libs_url(
  version = "1.10.2",
  type = install_type(version = version)
)
}
\arguments{
\item{version}{The Torch version to install.}

\item{type}{The installation type for Torch. Valid values are \code{"cpu"} or the 'CUDA' version.}
}
\description{
List the Torch and Lantern files to download as local files in order to proceed with install_torch_from_file().
}
