% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_torch}
\alias{install_torch}
\title{Install Torch}
\usage{
install_torch(version = "1.5.0", type = Sys.getenv("CUDA", unset =
  "cpu"), reinstall = FALSE, path = install_path())
}
\arguments{
\item{version}{The Torch version to install.}

\item{type}{The installation type for Torch. Valid values are \code{"cpu"} or the 'CUDA' version.}

\item{reinstall}{Re-install Torch even if its already installed?}

\item{path}{Optional path to install or check for an already existing installation.}
}
\description{
Installs Torch and its dependencies.
}
