% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-dataloader.R
\name{is_dataloader}
\alias{is_dataloader}
\title{Checks if the object is a dataloader}
\usage{
is_dataloader(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Checks if the object is a dataloader
}
