% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{is_grad_enabled}
\alias{is_grad_enabled}
\title{Test if gradient mode is enabled}
\usage{
is_grad_enabled()
}
\value{
Returns \code{TRUE} if grad mode is currently enabled.
}
\description{
Test if gradient mode is enabled
}
\examples{
if (torch_is_installed()) {
with_enable_grad(is_grad_enabled())
with_no_grad(is_grad_enabled())
}
}
