% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{is_nn_parameter}
\alias{is_nn_parameter}
\title{Checks if an object is a nn_parameter}
\usage{
is_nn_parameter(x)
}
\arguments{
\item{x}{the object to check}
}
\description{
Checks if an object is a nn_parameter
}
