% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{is_optimizer}
\alias{is_optimizer}
\title{Checks if the object is a torch optimizer}
\usage{
is_optimizer(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Checks if the object is a torch optimizer
}
