% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{is_sparse}
\alias{is_sparse}
\title{Checks if a tensor is a sparse (e.g. such as created by \code{torch_sparse_coo_tensor()})}
\usage{
is_sparse(x)
}
\arguments{
\item{x}{tensor to check}
}
\description{
Checks if a tensor is a sparse (e.g. such as created by \code{torch_sparse_coo_tensor()})
}
