% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{is_torch_device}
\alias{is_torch_device}
\title{Checks if object is a device}
\usage{
is_torch_device(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Checks if object is a device
}
\concept{tensor-attributes}
