% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtype.R
\name{is_torch_dtype}
\alias{is_torch_dtype}
\title{Check if object is a torch data type}
\usage{
is_torch_dtype(x)
}
\arguments{
\item{x}{object to check.}
}
\description{
Check if object is a torch data type
}
\concept{tensor-attributes}
