% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{is_torch_layout}
\alias{is_torch_layout}
\title{Check if an object is a torch layout.}
\usage{
is_torch_layout(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if an object is a torch layout.
}
