% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_format.R
\name{is_torch_memory_format}
\alias{is_torch_memory_format}
\title{Check if an object is a memory format}
\usage{
is_torch_memory_format(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Check if an object is a memory format
}
