% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{is_torch_tensor}
\alias{is_torch_tensor}
\title{Checks if object is a \link{torch_tensor}}
\usage{
is_torch_tensor(x)
}
\arguments{
\item{x}{object to check}
}
\description{
Checks if object is a \link{torch_tensor}
}
