% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{is_undefined_tensor}
\alias{is_undefined_tensor}
\title{Checks if a tensor is undefined}
\usage{
is_undefined_tensor(x)
}
\arguments{
\item{x}{tensor to check}
}
\description{
Checks if a tensor is undefined
}
