% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{jit_load}
\alias{jit_load}
\title{Loads a \code{script_function} or \code{script_module} previously saved with \code{jit_save}}
\usage{
jit_load(path, ...)
}
\arguments{
\item{path}{a path to a \code{script_function} or \code{script_module} serialized with
\code{\link[=jit_save]{jit_save()}}.}

\item{...}{currently unused.}
}
\description{
Loads a \code{script_function} or \code{script_module} previously saved with \code{jit_save}
}
