% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivalue.R
\name{jit_scalar}
\alias{jit_scalar}
\title{Adds the 'jit_scalar' class to the input}
\usage{
jit_scalar(x)
}
\arguments{
\item{x}{a length 1 R vector.}
}
\description{
Allows disambiguating length 1 vectors from scalars when passing
them to the jit.
}
