% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{jit_serialize}
\alias{jit_serialize}
\title{Serialize a Script Module}
\usage{
jit_serialize(obj)
}
\arguments{
\item{obj}{(\code{script_module})\cr
Model to be serialized.}
}
\value{
\code{raw()}
}
\description{
Serializes a script module and returns it as a raw vector.
You can read the object again using \code{\link{jit_unserialize}}.
}
\examples{
if (torch_is_installed()) {
model <- jit_trace(nn_linear(1, 1), torch_randn(1))
serialized <- jit_serialize(model)
}
}
