% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivalue.R
\name{jit_tuple}
\alias{jit_tuple}
\title{Adds the 'jit_tuple' class to the input}
\usage{
jit_tuple(x)
}
\arguments{
\item{x}{the list object that will be converted to a tuple.}
}
\description{
Allows specifying that an output or input must be considered a jit
tuple and instead of a list or dictionary when tracing.
}
