% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{jit_unserialize}
\alias{jit_unserialize}
\title{Unserialize a Script Module}
\usage{
jit_unserialize(obj)
}
\arguments{
\item{obj}{(\code{raw})\cr
Serialized model.}
}
\value{
\code{script_module}
model <- jit_trace(nn_linear(1, 1), torch_randn(1))
serialized <- jit_serialize(model)
model2 <- jit_unserialize(serialized)
}
\description{
Unserializes a script module from a raw vector (generated with \code{\link{jit_serialize}}`).
}
