% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{linalg_solve_triangular}
\alias{linalg_solve_triangular}
\title{Triangular solve}
\usage{
linalg_solve_triangular(A, B, ..., upper, left = TRUE, unitriangular = FALSE)
}
\arguments{
\item{A}{tensor of shape \verb{(*, n, n)} or \verb{(*, k, k)} if \code{left=TRUE}) where \code{*} is zero or more batch dimensions.}

\item{B}{right-hand side tensor of shape \verb{(*, n, k)}}

\item{...}{Currently ignored.}

\item{upper}{whether A is an upper or lower triangular matrix.}

\item{left}{wheter to solve the system AX=B or XA=B}

\item{unitriangular}{if \code{TRUE}, the diagonal elements of A are assumed to be all equal to 1.}
}
\description{
Triangular solve
}
\seealso{
Other linalg:
\code{\link[=linalg_cholesky]{linalg_cholesky()}},
\code{\link[=linalg_cholesky_ex]{linalg_cholesky_ex()}},
\code{\link[=linalg_det]{linalg_det()}},
\code{\link[=linalg_eig]{linalg_eig()}},
\code{\link[=linalg_eigh]{linalg_eigh()}},
\code{\link[=linalg_eigvals]{linalg_eigvals()}},
\code{\link[=linalg_eigvalsh]{linalg_eigvalsh()}},
\code{\link[=linalg_householder_product]{linalg_householder_product()}},
\code{\link[=linalg_inv]{linalg_inv()}},
\code{\link[=linalg_inv_ex]{linalg_inv_ex()}},
\code{\link[=linalg_lstsq]{linalg_lstsq()}},
\code{\link[=linalg_matrix_norm]{linalg_matrix_norm()}},
\code{\link[=linalg_matrix_power]{linalg_matrix_power()}},
\code{\link[=linalg_matrix_rank]{linalg_matrix_rank()}},
\code{\link[=linalg_multi_dot]{linalg_multi_dot()}},
\code{\link[=linalg_norm]{linalg_norm()}},
\code{\link[=linalg_pinv]{linalg_pinv()}},
\code{\link[=linalg_qr]{linalg_qr()}},
\code{\link[=linalg_slogdet]{linalg_slogdet()}},
\code{\link[=linalg_solve]{linalg_solve()}},
\code{\link[=linalg_svd]{linalg_svd()}},
\code{\link[=linalg_svdvals]{linalg_svdvals()}},
\code{\link[=linalg_tensorinv]{linalg_tensorinv()}},
\code{\link[=linalg_tensorsolve]{linalg_tensorsolve()}},
\code{\link[=linalg_vector_norm]{linalg_vector_norm()}}
}
\concept{linalg}
