% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{log}
\alias{log}
\alias{tch_log}
\alias{tch_log2}
\alias{tch_log10}
\alias{tch_log1p}
\title{log}
\usage{
tch_log(x)

tch_log2(x)

tch_log10(x)

tch_log1p(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the logarithm of the elements of input. The available logarithm functions are:

tch_log: Natural logarithm.

tch_log2: Base 2 logarithm.

tch_log10: Base 10 logarithm.

tch_log1p: Same as tch_log(1 + x). This function is more accurate than tch_log() for small values of x.
}
\examples{
x <- tch_randn(c(2,2))

tch_log(x)

tch_log2(x)

tch_log10(x)

tch_log1p(x)

}
