% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-utils.R
\name{logits_to_probs}
\alias{logits_to_probs}
\title{Converts a tensor of logits into probabilities. Note that for the
binary case, each value denotes log odds, whereas for the
multi-dimensional case, the values along the last dimension denote
the log probabilities (possibly unnormalized) of the events.}
\usage{
logits_to_probs(logits, is_binary = FALSE)
}
\description{
Converts a tensor of logits into probabilities. Note that for the
binary case, each value denotes log odds, whereas for the
multi-dimensional case, the values along the last dimension denote
the log probabilities (possibly unnormalized) of the events.
}
