% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim-lr_scheduler.R
\name{lr_cosine_annealing}
\alias{lr_cosine_annealing}
\title{Set the learning rate of each parameter group using a cosine annealing schedule}
\usage{
lr_cosine_annealing(
  optimizer,
  T_max,
  eta_min = 0,
  last_epoch = -1,
  verbose = FALSE
)
}
\arguments{
\item{optimizer}{(Optimizer): Wrapped optimizer.}

\item{T_max}{Maximum number of iterations}

\item{eta_min}{Minimum learning rate. Default: 0.}

\item{last_epoch}{The index of the last epoch}

\item{verbose}{(bool): If \code{TRUE}, prints a message to stdout for
each update. Default: \code{FALSE}.}
}
\description{
Set the learning rate of each parameter group using a cosine annealing schedule
}
