% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision-utils.R
\name{make_grid}
\alias{make_grid}
\title{A simplified version of torchvision.utils.make_grid.}
\usage{
make_grid(tensor, num_rows = 8, padding = 2, pad_value = 0)
}
\arguments{
\item{tensor}{tensor to normalize}

\item{num_rows}{number of rows making up the grid (default 8)}

\item{padding}{amount of padding between batch images (default 2)}

\item{pad_value}{pixel value to use for padding}
}
\description{
Arranges a batch of (image) tensors in a grid, with optional padding between images.
Expects a 4d mini-batch tensor of shape (B x C x H x W).
}
