% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{mm.tensor}
\alias{mm.tensor}
\title{matrix multiplication}
\usage{
\method{mm}{tensor}(x, mat2)
}
\arguments{
\item{x}{tensor object}

\item{mat2}{second tensor object}
}
\description{
Performs matrix multiplication for 2 tensors.
}
\examples{
x <- tensor(matrix(runif(10), ncol = 5))
y <- tensor(matrix(runif(10), nrow = 5))
mm(x, y)
}
