% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-pooling.R
\name{nn_adaptive_avg_pool1d}
\alias{nn_adaptive_avg_pool1d}
\title{Applies a 1D adaptive average pooling over an input signal composed of several input planes.}
\usage{
nn_adaptive_avg_pool1d(output_size)
}
\arguments{
\item{output_size}{the target output size H}
}
\description{
The output size is H, for any input size.
The number of output features is equal to the number of input planes.
}
\examples{
if (torch_is_installed()) {
# target output size of 5
m <- nn_adaptive_avg_pool1d(5)
input <- torch_randn(1, 64, 8)
output <- m(input)
}
}
