% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nn_buffer}
\alias{nn_buffer}
\title{Creates a nn_buffer}
\usage{
nn_buffer(x, persistent = TRUE)
}
\arguments{
\item{x}{the tensor that will be converted to nn_buffer}

\item{persistent}{whether the buffer should be persistent or not.}
}
\description{
Indicates that a tensor is a buffer in a nn_module
}
