% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-activation.R
\name{nn_contrib_sparsemax}
\alias{nn_contrib_sparsemax}
\title{Sparsemax activation}
\usage{
nn_contrib_sparsemax(dim = -1)
}
\arguments{
\item{dim}{The dimension over which to apply the sparsemax function. (-1)}
}
\description{
Sparsemax activation module.
}
\details{
The SparseMax activation is described in
\href{https://arxiv.org/abs/1602.02068}{'From Softmax to Sparsemax: A Sparse Model of Attention and Multi-Label Classification'}
The implementation is based on \href{https://github.com/aced125/sparsemax/tree/master/sparsemax}{aced125/sparsemax}
}
