% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-linear.R
\name{nn_identity}
\alias{nn_identity}
\title{Identity module}
\usage{
nn_identity(...)
}
\arguments{
\item{...}{any arguments (unused)}
}
\description{
A placeholder identity operator that is argument-insensitive.
}
\examples{
if (torch_is_installed()) {
m <- nn_identity(54, unused_argument1 = 0.1, unused_argument2 = FALSE)
input <- torch_randn(128, 20)
output <- m(input)
print(output$size())
}
}
