% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-init.R
\name{nn_init_calculate_gain}
\alias{nn_init_calculate_gain}
\title{Calculate gain}
\usage{
nn_init_calculate_gain(nonlinearity, param = NULL)
}
\arguments{
\item{nonlinearity}{the non-linear function}

\item{param}{optional parameter for the non-linear function}
}
\description{
Return the recommended gain value for the given nonlinearity function.
}
