% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-init.R
\name{nn_init_constant_}
\alias{nn_init_constant_}
\title{Constant initialization}
\usage{
nn_init_constant_(tensor, val)
}
\arguments{
\item{tensor}{an n-dimensional \code{Tensor}}

\item{val}{the value to fill the tensor with}
}
\description{
Fills the input Tensor with the value \code{val}.
}
\examples{
if (torch_is_installed()) {
w <- torch_empty(3, 5)
nn_init_constant_(w, 0.3)
}
}
