% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-init.R
\name{nn_init_normal_}
\alias{nn_init_normal_}
\title{Normal initialization}
\usage{
nn_init_normal_(tensor, mean = 0, std = 1)
}
\arguments{
\item{tensor}{an n-dimensional Tensor}

\item{mean}{the mean of the normal distribution}

\item{std}{the standard deviation of the normal distribution}
}
\description{
Fills the input Tensor with values drawn from the normal distribution
}
\examples{
if (torch_is_installed()) {
w <- torch_empty(3, 5)
nn_init_normal_(w)
}
}
