% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-init.R
\name{nn_init_ones_}
\alias{nn_init_ones_}
\title{Ones initialization}
\usage{
nn_init_ones_(tensor)
}
\arguments{
\item{tensor}{an n-dimensional \code{Tensor}}
}
\description{
Fills the input Tensor with the scalar value \code{1}
}
\examples{
if (torch_is_installed()) {
w <- torch_empty(3, 5)
nn_init_ones_(w)
}
}
