% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-init.R
\name{nn_init_uniform_}
\alias{nn_init_uniform_}
\title{Uniform initialization}
\usage{
nn_init_uniform_(tensor, a = 0, b = 1)
}
\arguments{
\item{tensor}{an n-dimensional Tensor}

\item{a}{the lower bound of the uniform distribution}

\item{b}{the upper bound of the uniform distribution}
}
\description{
Fills the input Tensor with values drawn from the uniform distribution
}
\examples{
if (torch_is_installed()) {
w <- torch_empty(3, 5)
nn_init_uniform_(w)
}
}
