% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-init.R
\name{nn_init_zeros_}
\alias{nn_init_zeros_}
\title{Zeros initialization}
\usage{
nn_init_zeros_(tensor)
}
\arguments{
\item{tensor}{an n-dimensional tensor}
}
\description{
Fills the input Tensor with the scalar value \code{0}
}
\examples{
if (torch_is_installed()) {
w <- torch_empty(3, 5)
nn_init_zeros_(w)
}
}
