% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_linear.R
\name{nn_linear}
\alias{nn_linear}
\title{Linear module}
\usage{
nn_linear(in_features, out_features, bias = TRUE)
}
\arguments{
\item{in_features}{size of each input sample}

\item{out_features}{size of each output sample}

\item{bias}{If set to \code{FALSE}, the layer will not learn an additive bias.
Default: \code{TRUE}}
}
\description{
Applies a linear transformation to the incoming data: \code{y = xA^T + b}
}
