% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nn_parameter}
\alias{nn_parameter}
\title{Creates an \code{nn_parameter}}
\usage{
nn_parameter(x, requires_grad = TRUE)
}
\arguments{
\item{x}{the tensor that you want to indicate as parameter}

\item{requires_grad}{whether this parameter should  have
\code{requires_grad = TRUE}}
}
\description{
Indicates to nn_module that \code{x} is a parameter
}
