% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-activation.R
\name{nn_sigmoid}
\alias{nn_sigmoid}
\title{Sigmoid module}
\usage{
nn_sigmoid()
}
\description{
Applies the element-wise function:
}
\details{
\deqn{
  \mbox{Sigmoid}(x) = \sigma(x) = \frac{1}{1 + \exp(-x)}
}
}
\section{Shape}{

\itemize{
\item Input: \eqn{(N, *)} where \code{*} means, any number of additional
dimensions
\item Output: \eqn{(N, *)}, same shape as the input
}
}

\examples{
if (torch_is_installed()) {
m <- nn_sigmoid()
input <- torch_randn(2)
output <- m(input)
}
}
