% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-activation.R
\name{nn_tanhshrink}
\alias{nn_tanhshrink}
\title{Tanhshrink module}
\usage{
nn_tanhshrink()
}
\description{
Applies the element-wise function:
}
\details{
\deqn{
  \mbox{Tanhshrink}(x) = x - \tanh(x)
}
}
\section{Shape}{

\itemize{
\item Input: \eqn{(N, *)} where \code{*} means, any number of additional
dimensions
\item Output: \eqn{(N, *)}, same shape as the input
}
}

\examples{
if (torch_is_installed()) {
m <- nn_tanhshrink()
input <- torch_randn(2)
output <- m(input)
}
}
