% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-utils-clip-grad.R
\name{nn_utils_clip_grad_value_}
\alias{nn_utils_clip_grad_value_}
\title{Clips gradient of an iterable of parameters at specified value.}
\usage{
nn_utils_clip_grad_value_(parameters, clip_value)
}
\arguments{
\item{parameters}{(Iterable\link{Tensor} or Tensor): an iterable of Tensors or a
single Tensor that will have gradients normalized}

\item{clip_value}{(float or int): maximum allowed value of the gradients.}
}
\description{
Gradients are modified in-place.
}
\details{
The gradients are clipped in the range
\eqn{\left[\mbox{-clip\_value}, \mbox{clip\_value}\right]}
}
