% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-pooling.R
\name{nnf_adaptive_avg_pool2d}
\alias{nnf_adaptive_avg_pool2d}
\title{Adaptive_avg_pool2d}
\usage{
nnf_adaptive_avg_pool2d(input, output_size)
}
\arguments{
\item{input}{input tensor (minibatch, in_channels , iH , iW)}

\item{output_size}{the target output size (single integer or double-integer tuple)}
}
\description{
Applies a 2D adaptive average pooling over an input signal composed of
several input planes.
}
