% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_adaptive_avg_pool3d}
\alias{nnf_adaptive_avg_pool3d}
\title{Adaptive_avg_pool3d}
\arguments{
\item{output_size}{NA the target output size (single integer or        triple-integer tuple)}
}
\description{
Applies a 3D adaptive average pooling over an input signal composed of
several input planes.
}
\details{
\preformatted{See `~torch.nn.AdaptiveAvgPool3d` for details and output shape.
}
}
\examples{


}
