% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-dropout.R
\name{nnf_alpha_dropout}
\alias{nnf_alpha_dropout}
\title{Alpha_dropout}
\usage{
nnf_alpha_dropout(input, p = 0.5, training = FALSE, inplace = FALSE)
}
\arguments{
\item{input}{the input tensor}

\item{p}{probability of an element to be zeroed. Default: 0.5}

\item{training}{apply dropout if is \code{TRUE}. Default: \code{TRUE}}

\item{inplace}{If set to \code{TRUE}, will do this operation in-place.
Default: \code{FALSE}}
}
\description{
Applies alpha dropout to the input.
}
