% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_avg_pool1d}
\alias{nnf_avg_pool1d}
\title{Avg_pool1d}
\arguments{
\item{input}{NA input tensor of shape \eqn{(\text{minibatch} , \text{in\_channels} , iW)}}

\item{kernel_size}{NA the size of the window. Can be a single number or a      tuple \verb{(kW,)}}

\item{stride}{NA the stride of the window. Can be a single number or a tuple      \verb{(sW,)}. Default: \code{kernel_size}}

\item{padding}{NA implicit zero paddings on both sides of the input. Can be a      single number or a tuple \verb{(padW,)}. Default: 0}

\item{ceil_mode}{NA when True, will use \code{ceil} instead of \code{floor} to compute the        output shape. Default: \code{False}}

\item{count_include_pad}{NA when True, will include the zero-padding in the        averaging calculation. Default: \code{True}}
}
\description{
Avg_pool1d
}
\section{avg_pool1d(input, kernel_size, stride=None, padding=0, ceil_mode=False, count_include_pad=True) -> Tensor }{


Applies a 1D average pooling over an input signal composed of several
input planes.

See \code{~torch.nn.AvgPool1d} for details and output shape.
}

\examples{

# pool of square window of size=3, stride=2
input = torch_tensor(c([[1, 2, 3, 4, 5, 6, 7]]), dtype=torch_float32())
F$avg_pool1d(input, kernel_size=3, stride=2)
}
