% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_avg_pool2d}
\alias{nnf_avg_pool2d}
\title{Avg_pool2d}
\arguments{
\item{input}{NA input tensor \eqn{(\text{minibatch} , \text{in\_channels} , iH , iW)}}

\item{kernel_size}{NA size of the pooling region. Can be a single number or a      tuple \verb{(kH, kW)}}

\item{stride}{NA stride of the pooling operation. Can be a single number or a      tuple \verb{(sH, sW)}. Default: \code{kernel_size}}

\item{padding}{NA implicit zero paddings on both sides of the input. Can be a      single number or a tuple \verb{(padH, padW)}. Default: 0}

\item{ceil_mode}{NA when True, will use \code{ceil} instead of \code{floor} in the formula        to compute the output shape. Default: \code{False}}

\item{count_include_pad}{NA when True, will include the zero-padding in the        averaging calculation. Default: \code{True}}

\item{divisor_override}{NA if specified, it will be used as divisor, otherwise         size of the pooling region will be used. Default: None}
}
\description{
Avg_pool2d
}
\section{avg_pool2d(input, kernel_size, stride=None, padding=0, ceil_mode=False, count_include_pad=True, divisor_override=None) -> Tensor }{


Applies 2D average-pooling operation in \eqn{kH \times kW} regions by step size
\eqn{sH \times sW} steps. The number of output features is equal to the number of
input planes.

See \code{~torch.nn.AvgPool2d} for details and output shape.
}

\examples{


}
