% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_batch_norm}
\alias{nnf_batch_norm}
\title{Batch_norm}
\description{
Applies Batch Normalization for each channel across a batch of data.
}
\details{
\preformatted{See `~torch.nn.BatchNorm1d`, `~torch.nn.BatchNorm2d`,
`~torch.nn.BatchNorm3d` for details.
}
}
\examples{


}
