% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_bilinear}
\alias{nnf_bilinear}
\title{Bilinear}
\description{
Applies a bilinear transformation to the incoming data:
\eqn{y = x_1 A x_2 + b}
}
\details{
\preformatted{Shape:

    - input1: \eqn{(N, *, H_{in1})} where \eqn{H_{in1}=\text{in1\_features}}
      and \eqn{*} means any number of additional dimensions.
      All but the last dimension of the inputs should be the same.
    - input2: \eqn{(N, *, H_{in2})} where \eqn{H_{in2}=\text{in2\_features}}
    - weight: \eqn{(\text{out\_features}, \text{in1\_features},
          \text{in2\_features})}
    - bias: \eqn{(\text{out\_features})}
    - output: \eqn{(N, *, H_{out})} where \eqn{H_{out}=\text{out\_features}}
      and all but the last dimension are the same shape as the input.
}
}
\examples{


}
