% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-loss.R
\name{nnf_binary_cross_entropy}
\alias{nnf_binary_cross_entropy}
\title{Binary_cross_entropy}
\usage{
nnf_binary_cross_entropy(
  input,
  target,
  weight = NULL,
  size_average = NULL,
  reduction = c("mean", "sum", "none")
)
}
\arguments{
\item{input}{tensor (N,∗) where *∗ means, any number of additional dimensions}

\item{target}{tensor (N,∗) , same shape as the input}

\item{reduction}{(string, optional) – Specifies the reduction to apply to the
output: 'none' | 'mean' | 'sum'. 'none': no reduction will be applied, 'mean':
the sum of the output will be divided by the number of elements in the output,
'sum': the output will be summed. Default: 'mean'}
}
\description{
Function that measures the Binary Cross Entropy
between the target and the output.
}
\examples{

input = torch_randn(list(3, 2), requires_grad=TRUE)
target = torch_rand(list(3, 2), requires_grad=FALSE)
loss = F$binary_cross_entropy(F$sigmoid(input), target)
loss$backward()
}
