% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_celu}
\alias{nnf_celu}
\title{Celu}
\description{
Celu
}
\section{celu(input, alpha=1., inplace=False) -> Tensor }{


Applies element-wise,
\eqn{\text{CELU}(x) = \max(0,x) + \min(0, \alpha * (\exp(x/\alpha) - 1))}.\preformatted{See `~torch.nn.CELU` for more details.
}
}

\examples{


}
