% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_celu_}
\alias{nnf_celu_}
\title{Celu_}
\description{
Celu_
}
\section{celu_(input, alpha=1.) -> Tensor }{


In-place version of \code{\link{~celu}}.
}

\examples{


}
