% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-conv.R
\name{nnf_conv1d}
\alias{nnf_conv1d}
\title{Conv1d}
\usage{
nnf_conv1d(
  input,
  weight,
  bias = NULL,
  stride = 1,
  padding = 0,
  dilation = 1,
  groups = 1
)
}
\arguments{
\item{input}{input tensor of shape (minibatch, in_channels , iW)}

\item{weight}{filters of shape (out_channels, in_channels/groups , kW)}

\item{bias}{optional bias of shape (out_channels). Default: \code{NULL}}

\item{stride}{the stride of the convolving kernel. Can be a single number or
a one-element tuple \verb{(sW,)}. Default: 1}

\item{padding}{implicit paddings on both sides of the input. Can be a
single number or a one-element tuple \verb{(padW,)}. Default: 0}

\item{dilation}{the spacing between kernel elements. Can be a single number or
a one-element tuple \verb{(dW,)}. Default: 1}

\item{groups}{split input into groups, \code{in_channels} should be divisible by
the number of groups. Default: 1}
}
\description{
Applies a 1D convolution over an input signal composed of several input
planes.
}
\examples{

filters = torch_randn(c(33, 16, 3))
inputs = torch_randn(c(20, 16, 50))
F$conv1d(inputs, filters)
}
