% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-conv.R
\name{nnf_conv2d}
\alias{nnf_conv2d}
\title{Conv2d}
\usage{
nnf_conv2d(
  input,
  weight,
  bias = NULL,
  stride = 1,
  padding = 0,
  dilation = 1,
  groups = 1
)
}
\arguments{
\item{input}{input tensor of shape (minibatch, in_channels, iH , iW)}

\item{weight}{filters of shape (out_channels , in_channels/groups, kH , kW)}

\item{bias}{optional bias tensor of shape (out_channels). Default: \code{NULL}}

\item{stride}{the stride of the convolving kernel. Can be a single number or a
tuple \verb{(sH, sW)}. Default: 1}

\item{padding}{implicit paddings on both sides of the input. Can be a
single number or a tuple \verb{(padH, padW)}. Default: 0}

\item{dilation}{the spacing between kernel elements. Can be a single number or
a tuple \verb{(dH, dW)}. Default: 1}

\item{groups}{split input into groups, \code{in_channels} should be divisible by the
number of groups. Default: 1}
}
\description{
Applies a 2D convolution over an input image composed of several input
planes.
}
\examples{

# With square kernels and equal stride
filters = torch_randn(c(8,4,3,3))
inputs = torch_randn(c(1,4,5,5))
F$conv2d(inputs, filters, padding=1)
}
