% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-conv.R
\name{nnf_conv_tbc}
\alias{nnf_conv_tbc}
\title{Conv_tbc}
\usage{
nnf_conv_tbc(input, weight, bias, pad = 0)
}
\arguments{
\item{input}{input tensor of shape \eqn{(\text{sequence length} \times 
batch \times \text{in\_channels})}}

\item{weight}{filter of shape (\eqn{\text{kernel width} \times \text{in\_channels} 
\times \text{out\_channels}})}

\item{bias}{bias of shape (\eqn{\text{out\_channels}})}

\item{pad}{number of timesteps to pad. Default: 0}
}
\description{
Applies a 1-dimensional sequence convolution over an input sequence.
Input and output dimensions are (Time, Batch, Channels) - hence TBC.
}
