% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_cosine_embedding_loss}
\alias{nnf_cosine_embedding_loss}
\title{Cosine_embedding_loss}
\description{
Cosine_embedding_loss
}
\section{cosine_embedding_loss(input1, input2, target, margin=0, size_average=None, reduce=None, reduction='mean') -> Tensor }{


See \code{~torch.nn.CosineEmbeddingLoss} for details.
}

\examples{


}
