% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-dropout.R
\name{nnf_dropout}
\alias{nnf_dropout}
\title{Dropout}
\usage{
nnf_dropout(input, p = 0.5, training = TRUE, inplace = FALSE)
}
\arguments{
\item{input}{the input tensor}

\item{p}{probability of an element to be zeroed. Default: 0.5}

\item{training}{apply dropout if is \code{TRUE}. Default: \code{TRUE}}

\item{inplace}{If set to \code{TRUE}, will do this operation in-place.
Default: \code{FALSE}}
}
\description{
During training, randomly zeroes some of the elements of the input
tensor with probability \code{p} using samples from a Bernoulli
distribution.
}
\examples{


}
