% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_dropout2d}
\alias{nnf_dropout2d}
\title{Dropout2d}
\arguments{
\item{p}{NA probability of a channel to be zeroed. Default: 0.5}

\item{training}{NA apply dropout if is \code{True}. Default: \code{True}}

\item{inplace}{NA If set to \code{True}, will do this operation in-place. Default: \code{False}}
}
\description{
Dropout2d
}
\section{Randomly zero out entire channels (a channel is a 2D feature map, }{


Randomly zero out entire channels (a channel is a 2D feature map,
e.g., the \eqn{j}-th channel of the \eqn{i}-th sample in the
batched input is a 2D tensor \eqn{\text{input}[i, j]}) of the input tensor).
Each channel will be zeroed out independently on every forward call with
probability \code{p} using samples from a Bernoulli distribution.\preformatted{See `~torch.nn.Dropout2d` for details.
}
}

\examples{


}
