% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_elu}
\alias{nnf_elu}
\title{Elu}
\description{
Applies element-wise,
\eqn{\text{ELU}(x) = \max(0,x) + \min(0, \alpha * (\exp(x) - 1))}.
}
\details{
\preformatted{See `~torch.nn.ELU` for more details.
}
}
\examples{


}
