% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_elu_}
\alias{nnf_elu_}
\title{Elu_}
\description{
Elu_
}
\section{elu_(input, alpha=1.) -> Tensor }{


In-place version of \code{\link{~elu}}.
}

\examples{


}
