% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_glu}
\alias{nnf_glu}
\title{Glu}
\arguments{
\item{input}{(Tensor) input tensor}

\item{dim}{(int) dimension on which to split the input. Default: -1}
}
\description{
Glu
}
\section{glu(input, dim=-1) -> Tensor }{


The gated linear unit. Computes:\preformatted{.. math ::
    \\text\{GLU\}(a, b) = a \\otimes \\sigma(b)

where `input` is split in half along `dim` to form `a` and `b`, \eqn{\sigma}
is the sigmoid function and \eqn{\otimes} is the element-wise product between matrices.

See `Language Modeling with Gated Convolutional Networks <https://arxiv.org/abs/1612.08083>`_.
}
}

\examples{


}
