% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_group_norm}
\alias{nnf_group_norm}
\title{Group_norm}
\description{
Applies Group Normalization for last certain number of dimensions.
}
\details{
\preformatted{See `~torch.nn.GroupNorm` for details.
}
}
\examples{


}
