% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R,
%   R/nnf-activation.R
\name{nnf_gumbel_softmax}
\alias{nnf_gumbel_softmax}
\title{Gumbel_softmax}
\usage{
nnf_gumbel_softmax(logits, tau = 1, hard = FALSE, dim = -1)
}
\arguments{
\item{logits}{\verb{[..., num_features]} unnormalized log probabilities}

\item{tau}{non-negative scalar temperature}

\item{hard}{if \code{True}, the returned samples will be discretized as one-hot vectors,        but will be differentiated as if it is the soft sample in autograd}

\item{dim}{(int) A dimension along which softmax will be computed. Default: -1.}
}
\description{
Samples from the Gumbel-Softmax distribution and
optionally discretizes.
}
\examples{

logits = torch_randn(c(20, 32))
# Sample soft categorical using reparametrization trick:
F$gumbel_softmax(logits, tau=1, hard=FALSE)
# Sample hard categorical using "Straight-through" trick:
F$gumbel_softmax(logits, tau=1, hard=TRUE)
}
