% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_handle_torch_function}
\alias{nnf_handle_torch_function}
\title{Handle_torch_function}
\description{
Implement a function with checks for \strong{torch_function} overrides.
}
\details{
\preformatted{See torch::autograd::handle_torch_function for the equivalent of this
function in the C++ implementation.

Arguments
---------
public_api : function
    Function exposed by the public torch API originally called like
    ``public_api(*args, **kwargs)`` on which arguments are now being
    checked.
relevant_args : iterable
    Iterable of arguments to check for __torch_function__ methods.
args : tuple
    Arbitrary positional arguments originally passed into ``public_api``.
kwargs : tuple
    Arbitrary keyword arguments originally passed into ``public_api``.

Returns
-------
Result from calling `implementation()` or an `__torch_function__`
method, as appropriate.

Raises
------
TypeError : if no implementation is found.
}
}
\examples{


}
