% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_hardshrink}
\alias{nnf_hardshrink}
\title{Hardshrink}
\description{
Hardshrink
}
\section{hardshrink(input, lambd=0.5) -> Tensor }{


Applies the hard shrinkage function element-wise\preformatted{See `~torch.nn.Hardshrink` for more details.
}
}

\examples{


}
