% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_hardsigmoid}
\alias{nnf_hardsigmoid}
\title{Hardsigmoid}
\usage{
nnf_hardsigmoid(input, inplace = FALSE)
}
\arguments{
\item{input}{(N,∗) tensor, where * means, any number of additional
dimensions}

\item{inplace}{NA If set to \code{True}, will do this operation in-place. Default: \code{False}}
}
\description{
Applies the element-wise function \eqn{\text{Hardsigmoid}(x) = \frac{ReLU6(x + 3)}{6}}
}
