% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_hardtanh}
\alias{nnf_hardtanh}
\title{Hardtanh}
\description{
Hardtanh
}
\section{hardtanh(input, min_val=-1., max_val=1., inplace=False) -> Tensor }{


Applies the HardTanh function element-wise. See \code{~torch.nn.Hardtanh} for more
details.
}

\examples{


}
