% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_hardtanh_}
\alias{nnf_hardtanh_}
\title{Hardtanh_}
\description{
Hardtanh_
}
\section{hardtanh_(input, min_val=-1., max_val=1.) -> Tensor }{


In-place version of \code{\link{~hardtanh}}.
}

\examples{


}
