% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_has_torch_function}
\alias{nnf_has_torch_function}
\title{Has_torch_function}
\description{
Check for \strong{torch_function} implementations in the elements of an iterable
}
\details{
\preformatted{Arguments
---------
relevant_args : iterable
    Iterable or aguments to check for __torch_function__ methods.

Returns
-------
True if any of the elements of relevant_args have __torch_function__
implementations, False otherwise.
}
}
\examples{


}
