% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-loss.R
\name{nnf_hinge_embedding_loss}
\alias{nnf_hinge_embedding_loss}
\title{Hinge_embedding_loss}
\usage{
nnf_hinge_embedding_loss(input, target, margin = 1, reduction = "mean")
}
\arguments{
\item{input}{tensor (N,∗) where *∗ means, any number of additional dimensions}

\item{target}{tensor (N,∗) , same shape as the input}

\item{margin}{Has a default value of 1.}

\item{reduction}{(string, optional) – Specifies the reduction to apply to the
output: 'none' | 'mean' | 'sum'. 'none': no reduction will be applied, 'mean':
the sum of the output will be divided by the number of elements in the output,
'sum': the output will be summed. Default: 'mean'}
}
\description{
Measures the loss given an input tensor xx and a labels tensor yy (containing 1 or -1).
This is usually used for measuring whether two inputs are similar or dissimilar, e.g.
using the L1 pairwise distance as xx , and is typically used for learning nonlinear
embeddings or semi-supervised learning.
}
