% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-instancenorm.R
\name{nnf_instance_norm}
\alias{nnf_instance_norm}
\title{Instance_norm}
\usage{
nnf_instance_norm(
  input,
  running_mean = NULL,
  running_var = NULL,
  weight = NULL,
  bias = NULL,
  use_input_stats = TRUE,
  momentum = 0.1,
  eps = 1e-05
)
}
\arguments{
\item{input}{the input tensor}

\item{running_mean}{the running_mean tensor}

\item{running_var}{the running var tensor}

\item{weight}{the weight tensor}

\item{bias}{the bias tensor}

\item{use_input_stats}{whether to use input stats}

\item{momentum}{a double for the momentum}

\item{eps}{an eps double for numerical stability}
}
\description{
Applies Instance Normalization for each channel in each data sample in a
batch.
}
