% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-loss.R
\name{nnf_kl_div}
\alias{nnf_kl_div}
\title{Kl_div}
\usage{
nnf_kl_div(input, target, reduction = "mean")
}
\arguments{
\item{input}{tensor (N,*) where ** means, any number of additional dimensions}

\item{target}{tensor (N,*) , same shape as the input}

\item{reduction}{(string, optional) – Specifies the reduction to apply to the
output: 'none' | 'mean' | 'sum'. 'none': no reduction will be applied, 'mean':
the sum of the output will be divided by the number of elements in the output,
'sum': the output will be summed. Default: 'mean'}
}
\description{
The Kullback-Leibler divergence Loss.
}
