% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_layer_norm}
\alias{nnf_layer_norm}
\title{Layer_norm}
\description{
Applies Layer Normalization for last certain number of dimensions.
}
\details{
\preformatted{See `~torch.nn.LayerNorm` for details.
}
}
\examples{


}
