% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_leaky_relu}
\alias{nnf_leaky_relu}
\title{Leaky_relu}
\description{
Leaky_relu
}
\section{leaky_relu(input, negative_slope=0.01, inplace=False) -> Tensor }{


Applies element-wise,
\eqn{\text{LeakyReLU}(x) = \max(0, x) + \text{negative\_slope} * \min(0, x)}\preformatted{See `~torch.nn.LeakyReLU` for more details.
}
}

\examples{


}
