% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_leaky_relu_}
\alias{nnf_leaky_relu_}
\title{Leaky_relu_}
\description{
Leaky_relu_
}
\section{leaky_relu_(input, negative_slope=0.01) -> Tensor }{


In-place version of \code{\link{~leaky_relu}}.
}

\examples{


}
