% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-linear.R
\name{nnf_linear}
\alias{nnf_linear}
\title{Linear}
\usage{
nnf_linear(input, weight, bias = NULL)
}
\arguments{
\item{input}{\eqn{(N, *, in\_features)} where \code{*} means any number of
additional dimensions}

\item{weight}{\eqn{(out\_features, in\_features)} the weights tensor.}

\item{bias}{optional tensor \eqn{(out\_features)}}
}
\description{
Applies a linear transformation to the incoming data: \eqn{y = xA^T + b}.
}
\examples{


}
