% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_local_response_norm}
\alias{nnf_local_response_norm}
\title{Local_response_norm}
\description{
Applies local response normalization over an input signal composed of
several input planes, where channels occupy the second dimension.
Applies normalization across channels.
}
\details{
\preformatted{See `~torch.nn.LocalResponseNorm` for details.
}
}
\examples{


}
