% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_logsigmoid}
\alias{nnf_logsigmoid}
\title{Logsigmoid}
\description{
Logsigmoid
}
\section{logsigmoid(input) -> Tensor }{


Applies element-wise \eqn{\text{LogSigmoid}(x_i) = \log \left(\frac{1}{1 + \exp(-x_i)}\right)}

See \code{~torch.nn.LogSigmoid} for more details.
}

\examples{


}
