% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_lp_pool1d}
\alias{nnf_lp_pool1d}
\title{Lp_pool1d}
\description{
Applies a 1D power-average pooling over an input signal composed of
several input planes. If the sum of all inputs to the power of \code{p} is
zero, the gradient is set to zero as well.
}
\details{
\preformatted{See `~torch.nn.LPPool1d` for details.
}
}
\examples{


}
