% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_margin_ranking_loss}
\alias{nnf_margin_ranking_loss}
\title{Margin_ranking_loss}
\description{
Margin_ranking_loss
}
\section{margin_ranking_loss(input1, input2, target, margin=0, size_average=None, reduce=None, reduction='mean') -> Tensor }{


See \code{~torch.nn.MarginRankingLoss} for details.
}

\examples{


}
