% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_math}
\alias{nnf_math}
\title{Math}
\description{
This module provides access to the mathematical functions
defined by the C standard.
}
\examples{


}
