% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-pooling.R
\name{nnf_max_pool1d}
\alias{nnf_max_pool1d}
\title{Max_pool1d}
\usage{
nnf_max_pool1d(
  input,
  kernel_size,
  stride = NULL,
  padding = 0,
  dilation = 1,
  ceil_mode = FALSE,
  return_indices = FALSE
)
}
\arguments{
\item{input}{input tensor of shape (minibatch , in_channels , iW)}

\item{kernel_size}{the size of the window. Can be a single number or a
tuple \verb{(kW,)}.}

\item{stride}{the stride of the window. Can be a single number or a tuple
\verb{(sW,)}. Default: \code{kernel_size}}

\item{padding}{implicit zero paddings on both sides of the input. Can be a
single number or a tuple \verb{(padW,)}. Default: 0}

\item{ceil_mode}{when True, will use \code{ceil} instead of \code{floor} to compute the
output shape. Default: \code{FALSE}}

\item{return_indices}{whether to return the indices where the max occurs.}
}
\description{
Applies a 1D max pooling over an input signal composed of several input
planes.
}
\examples{


}
