% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_max_unpool1d}
\alias{nnf_max_unpool1d}
\title{Max_unpool1d}
\description{
Computes a partial inverse of \code{MaxPool1d}.
}
\details{
\preformatted{See `~torch.nn.MaxUnpool1d` for details.
}
}
\examples{


}
