% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_multi_margin_loss}
\alias{nnf_multi_margin_loss}
\title{Multi_margin_loss}
\description{
Multi_margin_loss
}
\section{multi_margin_loss(input, target, p=1, margin=1, weight=None, size_average=None, }{


reduce=None, reduction='mean') -> Tensor\preformatted{See `~torch.nn.MultiMarginLoss` for details.
}
}

\examples{


}
