% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_multilabel_soft_margin_loss}
\alias{nnf_multilabel_soft_margin_loss}
\title{Multilabel_soft_margin_loss}
\description{
Multilabel_soft_margin_loss
}
\section{multilabel_soft_margin_loss(input, target, weight=None, size_average=None) -> Tensor }{


See \code{~torch.nn.MultiLabelSoftMarginLoss} for details.
}

\examples{


}
