% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_prelu}
\alias{nnf_prelu}
\title{Prelu}
\description{
Prelu
}
\section{prelu(input, weight) -> Tensor }{


Applies element-wise the function
\eqn{\text{PReLU}(x) = \max(0,x) + \text{weight} * \min(0,x)} where weight is a
learnable parameter.\preformatted{See `~torch.nn.PReLU` for more details.
}
}

\examples{


}
