% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_relu}
\alias{nnf_relu}
\alias{nnf_relu_}
\title{Relu}
\usage{
nnf_relu(input, inplace = FALSE)

nnf_relu_(input)
}
\arguments{
\item{input}{(N,*) tensor, where * means, any number of additional
dimensions}

\item{inplace}{can optionally do the operation in-place. Default: FALSE}
}
\description{
Applies the rectified linear unit function element-wise.
}
