% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_relu6}
\alias{nnf_relu6}
\title{Relu6}
\usage{
nnf_relu6(input, inplace = FALSE)
}
\arguments{
\item{input}{(N,∗) tensor, where * means, any number of additional
dimensions}

\item{inplace}{can optionally do the operation in-place. Default: FALSE}
}
\description{
Applies the element-wise function \eqn{ReLU6(x) = min(max(0,x), 6)}.
}
