% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_relu_}
\alias{nnf_relu_}
\title{Relu_}
\description{
Relu_
}
\section{relu_(input) -> Tensor }{


In-place version of \code{\link{~relu}}.
}

\examples{


}
