% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_rrelu}
\alias{nnf_rrelu}
\title{Rrelu}
\description{
Rrelu
}
\section{rrelu(input, lower=1./8, upper=1./3, training=False, inplace=False) -> Tensor }{


Randomized leaky ReLU.\preformatted{See `~torch.nn.RReLU` for more details.
}
}

\examples{


}
