% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_rrelu_}
\alias{nnf_rrelu_}
\title{Rrelu_}
\description{
Rrelu_
}
\section{rrelu_(input, lower=1./8, upper=1./3, training=False) -> Tensor }{


In-place version of \code{\link{~rrelu}}.
}

\examples{


}
