% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_selu}
\alias{nnf_selu}
\title{Selu}
\description{
Selu
}
\section{selu(input, inplace=False) -> Tensor }{


Applies element-wise,
\eqn{\text{SELU}(x) = scale * (\max(0,x) + \min(0, \alpha * (\exp(x) - 1)))},
with \eqn{\alpha=1.6732632423543772848170429916717} and
\eqn{scale=1.0507009873554804934193349852946}.\preformatted{See `~torch.nn.SELU` for more details.
}
}

\examples{


}
