% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_selu_}
\alias{nnf_selu_}
\title{Selu_}
\description{
Selu_
}
\section{selu_(input) -> Tensor }{


In-place version of \code{\link{~selu}}.
}

\examples{


}
