% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_sigmoid}
\alias{nnf_sigmoid}
\title{Sigmoid}
\usage{
nnf_sigmoid(input)
}
\arguments{
\item{input}{(N,*) tensor, where * means, any number of additional
dimensions}
}
\description{
Applies element-wise \eqn{Sigmoid(x_i) = \frac{1}{1 + exp(-x_i)}}
}
