% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_silu}
\alias{nnf_silu}
\title{Applies the Sigmoid Linear Unit (SiLU) function, element-wise.
See \code{\link[=nn_silu]{nn_silu()}} for more information.}
\usage{
nnf_silu(input, inplace = FALSE)
}
\arguments{
\item{input}{(N,*) tensor, where * means, any number of additional
dimensions}

\item{inplace}{can optionally do the operation in-place. Default: FALSE}
}
\description{
Applies the Sigmoid Linear Unit (SiLU) function, element-wise.
See \code{\link[=nn_silu]{nn_silu()}} for more information.
}
\seealso{
\code{\link[=nn_silu]{nn_silu()}}.
}
