% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_smooth_l1_loss}
\alias{nnf_smooth_l1_loss}
\title{Smooth_l1_loss}
\description{
Function that uses a squared term if the absolute
element-wise error falls below 1 and an L1 term otherwise.
}
\details{
\preformatted{See `~torch.nn.SmoothL1Loss` for details.
}
}
\examples{


}
