% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_soft_margin_loss}
\alias{nnf_soft_margin_loss}
\title{Soft_margin_loss}
\description{
Soft_margin_loss
}
\section{soft_margin_loss(input, target, size_average=None, reduce=None, reduction='mean') -> Tensor }{


See \code{~torch.nn.SoftMarginLoss} for details.
}

\examples{


}
