% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_softmax}
\alias{nnf_softmax}
\title{Softmax}
\arguments{
\item{input}{(Tensor) input}

\item{dim}{(int) A dimension along which softmax will be computed.}

\item{dtype}{(\code{torch.dtype}, optional) the desired data type of returned tensor.      If specified, the input tensor is casted to \code{dtype} before the operation      is performed. This is useful for preventing data type overflows. Default: None.}
}
\description{
Applies a softmax function.
}
\details{
\preformatted{Softmax is defined as:

\eqn{\text{Softmax}(x_{i}) = \frac{exp(x_i)}{\sum_j exp(x_j)}}

It is applied to all slices along dim, and will re-scale them so that the elements
lie in the range `[0, 1]` and sum to 1.

See `~torch.nn.Softmax` for more details.
}
}
\examples{


}
