% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_softmin}
\alias{nnf_softmin}
\title{Softmin}
\arguments{
\item{input}{(Tensor) input}

\item{dim}{(int) A dimension along which softmin will be computed (so every slice        along dim will sum to 1).}

\item{dtype}{(\code{torch.dtype}, optional) the desired data type of returned tensor.      If specified, the input tensor is casted to \code{dtype} before the operation      is performed. This is useful for preventing data type overflows. Default: None.}
}
\description{
Applies a softmin function.
}
\details{
\preformatted{Note that \eqn{\text{Softmin}(x) = \text{Softmax}(-x)}. See softmax definition for mathematical formula.

See `~torch.nn.Softmin` for more details.
}
}
\examples{


}
