% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_softplus}
\alias{nnf_softplus}
\title{Softplus}
\description{
Softplus
}
\section{softplus(input, beta=1, threshold=20) -> Tensor }{


Applies element-wise, the function \eqn{\text{Softplus}(x) = \frac{1}{\beta} * \log(1 + \exp(\beta * x))}.

For numerical stability the implementation reverts to the linear function
when \eqn{input \times \beta > threshold}.

See \code{~torch.nn.Softplus} for more details.
}

\examples{


}
