% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_softshrink}
\alias{nnf_softshrink}
\title{Softshrink}
\description{
Softshrink
}
\section{softshrink(input, lambd=0.5) -> Tensor }{


Applies the soft shrinkage function elementwise

See \code{~torch.nn.Softshrink} for more details.
}

\examples{


}
