% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_softsign}
\alias{nnf_softsign}
\title{Softsign}
\description{
Softsign
}
\section{softsign(input) -> Tensor }{


Applies element-wise, the function \eqn{\text{SoftSign}(x) = \frac{x}{1 + |x|}}\preformatted{See `~torch.nn.Softsign` for more details.
}
}

\examples{


}
