% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_tanh}
\alias{nnf_tanh}
\title{Tanh}
\description{
Tanh
}
\section{tanh(input) -> Tensor }{


Applies element-wise,
\eqn{\text{Tanh}(x) = \tanh(x) = \frac{\exp(x) - \exp(-x)}{\exp(x) + \exp(-x)}}\preformatted{See `~torch.nn.Tanh` for more details.
}
}

\examples{


}
