% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_tanhshrink}
\alias{nnf_tanhshrink}
\title{Tanhshrink}
\description{
Tanhshrink
}
\section{tanhshrink(input) -> Tensor }{


Applies element-wise, \eqn{\text{Tanhshrink}(x) = x - \text{Tanh}(x)}\preformatted{See `~torch.nn.Tanhshrink` for more details.
}
}

\examples{


}
