% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R,
%   R/nnf-activation.R
\name{nnf_threshold}
\alias{nnf_threshold}
\alias{nnf_threshold_}
\title{Threshold}
\usage{
nnf_threshold(input, threshold, value, inplace = FALSE)

nnf_threshold_(input, threshold, value)
}
\arguments{
\item{input}{(N,∗) tensor, where * means, any number of additional
dimensions}

\item{threshold}{The value to threshold at}

\item{value}{The value to replace with}

\item{inplace}{can optionally do the operation in-place. Default: FALSE}
}
\description{
Thresholds each element of the input Tensor.
}
\examples{


}
