% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_triplet_margin_loss}
\alias{nnf_triplet_margin_loss}
\title{Triplet_margin_loss}
\description{
See \code{~torch.nn.TripletMarginLoss} for details
}
\examples{


}
