% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_unfold}
\alias{nnf_unfold}
\title{Unfold}
\description{
Extracts sliding local blocks from an batched input tensor.
}
\details{
\preformatted{.. warning::
    Currently, only 4-D input tensors (batched image-like tensors) are
    supported.

.. warning::

    More than one element of the unfolded tensor may refer to a single
    memory location. As a result, in-place operations (especially ones that
    are vectorized) may result in incorrect behavior. If you need to write
    to the tensor, please clone it first.


See `torch_nn.Unfold` for details
}
}
\examples{


}
