% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_upsample_bilinear}
\alias{nnf_upsample_bilinear}
\title{Upsample_bilinear}
\arguments{
\item{input}{(Tensor) input}

\item{size}{(int or Tuple\link{int, int}) output spatial size.}

\item{scale_factor}{(int or Tuple\link{int, int}) multiplier for spatial size}
}
\description{
Upsamples the input, using bilinear upsampling.
}
\details{
\preformatted{.. warning::
    This function is deprecated in favor of [`torch_nn.functional.interpolate`].
    This is equivalent with
    ``nn.functional.interpolate(..., mode='bilinear', align_corners=True)``.

Expected inputs are spatial (4 dimensional). Use `upsample_trilinear` fo
volumetric (5 dimensional) inputs.
}
}
\examples{


}
