% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_upsample_nearest}
\alias{nnf_upsample_nearest}
\title{Upsample_nearest}
\arguments{
\item{input}{(Tensor) input}

\item{size}{(int or Tuple\link{int, int} or Tuple\link{int, int, int}) output spatia        size.}

\item{scale_factor}{(int) multiplier for spatial size. Has to be an integer.}
}
\description{
Upsamples the input, using nearest neighbours' pixel values.
}
\details{
\preformatted{.. warning::
    This function is deprecated in favor of [`torch_nn.functional.interpolate`].
    This is equivalent with ``nn.functional.interpolate(..., mode='nearest')``.

Currently spatial and volumetric upsampling are supported (i.e. expected
inputs are 4 or 5 dimensional).
}
}
\examples{


}
