% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vision-utils.R
\name{normalize}
\alias{normalize}
\title{Normalizes a tensor to values between 0 and 1.}
\usage{
normalize(x)
}
\arguments{
\item{x}{tensor to normalize}
}
\description{
Normalizes a tensor to values between 0 and 1.
}
