% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignite.R
\name{optim_ignite_adam}
\alias{optim_ignite_adam}
\title{LibTorch implementation of Adam}
\usage{
optim_ignite_adam(
  params,
  lr = 0.001,
  betas = c(0.9, 0.999),
  eps = 1e-08,
  weight_decay = 0.01,
  amsgrad = FALSE
)
}
\description{
It has been proposed in \href{https://arxiv.org/abs/1412.6980}{Adam: A Method for Stochastic Optimization}.
}
\section{Methods}{

TODO:
}

\section{Fields}{

}

\examples{
if (torch_is_installed()) {
\dontrun{
optimizer <- optim_ignite_adam(model$parameters(), lr = 0.1)
optimizer$zero_grad()
loss_fn(model(input), target)$backward()
optimizer$step()
}
}
}
