% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ignite.R
\name{optimizer_ignite}
\alias{optimizer_ignite}
\title{Abstract Base Class for LibTorch Optimizers}
\usage{
optimizer_ignite(
  name = NULL,
  ...,
  private = NULL,
  active = NULL,
  parent_env = parent.frame()
)
}
\description{
Abstract base class for wrapping LibTorch C++ optimizers.
}
