% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{+.tensor}
\alias{+.tensor}
\title{elementwise addition}
\usage{
\method{+}{tensor}(x, y)
}
\arguments{
\item{x}{tensor object}

\item{y}{tensor object}
}
\description{
elementwise addition
}
\examples{
x <- tensor(1)
y <- tensor(2)
x + y
}
