% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{^.tensor}
\alias{^.tensor}
\title{pow}
\usage{
\method{^}{tensor}(x, y)
}
\arguments{
\item{x}{tensor object}

\item{y}{exponent (a tensor)}
}
\description{
pow
}
\examples{
x <- tensor(c(1,2,3,4))
y <- tensor(2)
x^y
x^3
}
