% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-utils.R
\name{probs_to_logits}
\alias{probs_to_logits}
\title{Converts a tensor of probabilities into logits. For the binary case,
this denotes the probability of occurrence of the event indexed by \code{1}.
For the multi-dimensional case, the values along the last dimension
denote the probabilities of occurrence of each of the events.}
\usage{
probs_to_logits(probs, is_binary = FALSE)
}
\description{
Converts a tensor of probabilities into logits. For the binary case,
this denotes the probability of occurrence of the event indexed by \code{1}.
For the multi-dimensional case, the values along the last dimension
denote the probabilities of occurrence of each of the events.
}
