% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-sampler.R
\name{sampler}
\alias{sampler}
\title{Creates a new Sampler}
\usage{
sampler(
  name = NULL,
  inherit = Sampler,
  ...,
  private = NULL,
  active = NULL,
  parent_env = parent.frame()
)
}
\description{
Samplers can be used with \code{\link[=dataloader]{dataloader()}} when creating batches from a torch
\code{\link[=dataset]{dataset()}}.
}
\details{
A sampler must implement the \code{.iter} and \code{.lenght()} methods.
\itemize{
\item \code{initialize} takes in a \code{data_source}. In general this is a \code{\link[=dataset]{dataset()}}.
\item \code{.iter} returns a function that returns a dataset index everytime it's called.
\item \code{.length} returns the maximum number of samples that can be retrieved from
that sampler.
}
}
