% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing.R
\name{slc}
\alias{slc}
\title{Creates a slice}
\usage{
slc(start, end, step = 1)
}
\arguments{
\item{start}{(integer) starting index.}

\item{end}{(integer) the last selected index.}

\item{step}{(integer) the step between indexes.}
}
\description{
Creates a slice object that can be used when indexing torch tensors.
}
\examples{
if (torch_is_installed()) {
x <- torch_randn(10)
x[slc(start = 1, end = 5, step = 2)]
}
}
