% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{sum.tensor}
\alias{sum.tensor}
\title{sum}
\usage{
\method{sum}{tensor}(x, dim = NULL, keepdim = NULL, dtype = NULL,
  na.rm = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{dim}{dimension in which to sum}

\item{keepdim}{wether to keep or not the dim}

\item{dtype}{optionaly cast the sum result}
}
\description{
sum
}
\examples{
x <- tensor(1:10)
sum(x)
}
