% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_any}
\alias{tch_any}
\title{any}
\usage{
tch_any(x, dim = NULL, keepdim = FALSE, na.rm = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{dim}{if NULL (the default) will reduce to a scalar. Otherwise it will
return TRUE if all elements in each row of the tensor in the given dimension
\code{dim} are TRUE, FALSE otherwise.}

\item{keepdim}{If keepdim is TRUE, the output tensor is of the same size as
input except in the dimension dim where it is of size 1. Otherwise, dim is
squeezed \code{\link[=squeeze]{squeeze()}}, resulting in the output tensor having 1 fewer
dimension than input.}

\item{na.rm}{won't be used by the function. Only there to be compatible with
\link{all} generic.}
}
\description{
any
}
\examples{
x <- tensor(array(c(TRUE, FALSE, TRUE, TRUE), dim = c(2, 2)))
tch_any(x)
tch_any(x, dim = 0)
tch_any(x, dim = 1, keepdim = FALSE)
}
