% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_atan}
\alias{tch_atan}
\title{atan}
\usage{
tch_atan(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the arctangent of the elements of input.
}
\examples{
x <- tensor(array(runif(8), dim = c(2,2,2)))
tch_atan(x)
}
