% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_baddbmm}
\alias{tch_baddbmm}
\title{baddbmm}
\usage{
tch_baddbmm(x, batch1, batch2, beta = 1, alpha = 1)
}
\arguments{
\item{x}{tensor object}

\item{batch1}{the first batch of matrices to be multiplied}

\item{batch2}{the second batch of matrices to be multiplied}

\item{beta}{multiplier for x}

\item{alpha}{multiplier for batch1 * batch2}
}
\description{
Performs a batch matrix-matrix product of matrices in batch1 and batch2.
x is added to the final result.
}
\examples{
x <- tensor(array(runif(45), dim = c(3, 3, 5)))
batch1 <- tensor(array(runif(36), dim = c(3, 3, 4)))
batch2 <- tensor(array(runif(60), dim = c(3, 4, 5)))
tch_baddbmm(x, batch1, batch2)
}
