% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_bernoulli}
\alias{tch_bernoulli}
\title{bernoulli}
\usage{
tch_bernoulli(x, p = NULL)
}
\arguments{
\item{x}{tensor object}

\item{p}{probability (if null uses tensor values)}
}
\description{
bernoulli
}
\examples{
x <- tensor(runif(10))
tch_bernoulli(x)

x <- tensor(rep(0, 10))
tch_bernoulli(x)
}
