% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_bincount}
\alias{tch_bincount}
\title{bincount}
\usage{
tch_bincount(x, weights = NULL, minlength = 0)
}
\arguments{
\item{x}{tensor object}

\item{weights}{optional, weight for each value in the input tensor. Should be of same size as input tensor.}

\item{minlength}{optional, minimum number of bins. Should be non-negative.}
}
\description{
Count the frequency of each value in an array of non-negative ints.
}
\examples{
x <- tensor(sample(0:9, 50, replace = TRUE))
tch_bincount(x)
}
