% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_btrifact}
\alias{tch_btrifact}
\title{Batch LU factorization}
\usage{
tch_btrifact(x, pivot = TRUE)
}
\arguments{
\item{x}{tensor object}

\item{pivot}{controls whether pivoting is done}
}
\description{
Returns a tuple containing the LU factorization and pivots.
Pivoting is done if pivot is set.
}
\examples{
x <- tensor(array(runif(18), dim = c(2, 3, 3)))
tch_btrifact(x)
}
