% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_btrisolve}
\alias{tch_btrisolve}
\title{Batch LU solve}
\usage{
tch_btrisolve(x, LU_data, LU_pivots)
}
\arguments{
\item{x}{tensor object}

\item{LU_data}{the pivoted LU factorization of A from \code{\link[=tch_btrifact]{tch_btrifact()}}}

\item{LU_pivots}{the pivots of the LU factorization}
}
\description{
Returns the LU solve of the linear system $Ax = b$.
}
\examples{
A <- tensor(array(runif(18), dim = c(2,3,3)))
b <- tensor(matrix(runif(6), ncol = 3))
A_LU <- tch_btrifact(A)
tch_btrisolve(b, A_LU[[1]], A_LU[[2]])
}
