% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_chunk}
\alias{tch_chunk}
\title{Chunk}
\usage{
tch_chunk(x, chunks, dim)
}
\arguments{
\item{x}{tensor object}

\item{chunks}{number of chunks to return}

\item{dim}{dimension along which to split the tensor}
}
\description{
Splits a tensor into a specific number of chunks.
Last chunk will be smaller if the tensor size along the given dimension
\code{dim} is not divisible by chunks.
}
\examples{
x <- tensor(array(runif(100), dim = c(4, 5, 5)))
tch_chunk(x, 2, 0)
}
