% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_clamp}
\alias{tch_clamp}
\title{Clamp}
\usage{
tch_clamp(x, min, max)
}
\arguments{
\item{x}{tensor object}

\item{min}{lower-bound of the range to be clamped to}

\item{max}{upper-bound of the range to be clamped to}
}
\description{
Clamp all elements in x into the range min, max and return a
resulting tensor.
}
\examples{
x <- tensor(1:10)
tch_clamp(x, 5, 7)
}
