% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_clamp_max}
\alias{tch_clamp_max}
\title{Clamp max}
\usage{
tch_clamp_max(x, max)
}
\arguments{
\item{x}{tensor object}

\item{max}{upper-bound of the range to be clamped to}
}
\description{
Clamp all elements in x into the range -Inf, max and return a
resulting tensor.
}
\examples{
x <- tensor(1:10)
tch_clamp_max(x, 5)
}
