% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_clamp_min}
\alias{tch_clamp_min}
\title{Clamp min}
\usage{
tch_clamp_min(x, min)
}
\arguments{
\item{x}{tensor object}

\item{min}{lower-bound of the range to be clamped to}
}
\description{
Clamp all elements in x into the range min, Inf and return a
resulting tensor.
}
\examples{
x <- tensor(1:10)
tch_clamp_min(x, 5)
}
