% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_cos}
\alias{tch_cos}
\title{Cosinus}
\usage{
tch_cos(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the cosine of the elements of input
}
\examples{
tch_cos(tensor(pi))
}
