% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_cosh}
\alias{tch_cosh}
\title{Hyperbolic Cosine}
\usage{
tch_cosh(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the hyperbolic cosine of the elements of input.
}
\examples{
tch_cosh(tensor(pi))
}
