% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_cross}
\alias{tch_cross}
\title{Cross}
\usage{
tch_cross(x, other, dim = -1)
}
\arguments{
\item{x}{tensor object}

\item{other}{the second input tensor}

\item{dim}{the dimension to take the cross-product in.}
}
\description{
Returns the cross product of vectors in dimension \code{dim} of \code{x} and \code{other}.
\code{x} and \code{other} must have the same size, and the size of their \code{dim} dimension
should be 3.
}
\details{
If dim is not given, it defaults to the first dimension found with the size 3.
}
\examples{

a <- tensor(matrix(runif(12), ncol = 3))
b <- tensor(matrix(runif(12), ncol = 3))
tch_cross(a, b, dim=1)
tch_cross(a, b)
}
