% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_cumprod}
\alias{tch_cumprod}
\title{Cumprod}
\usage{
tch_cumprod(x, dim)
}
\arguments{
\item{x}{tensor object}

\item{dim}{the dimension to do the operation over}
}
\description{
Returns the cumulative product of elements of input in the dimension dim.
}
\examples{
x <- tensor(1:10)
tch_cumprod(x)
}
