% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_cumsum}
\alias{tch_cumsum}
\title{Cumsum}
\usage{
tch_cumsum(x, dim)
}
\arguments{
\item{x}{tensor object}

\item{dim}{the dimension to do the operation over}
}
\description{
Returns the cumulative sum of elements of input in the dimension dim.
}
\examples{
x <- tensor(1:10)
tch_cumsum(x)
}
