% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_det}
\alias{tch_det}
\title{Determinant}
\usage{
tch_det(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Calculates determinant of a 2D square tensor.
}
\examples{
x <- tensor(matrix(runif(36), ncol = 6))
tch_det(x)
}
