% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/device.R
\name{tch_device}
\alias{tch_device}
\title{Create a Device}
\usage{
tch_device(type, index = NULL)
}
\arguments{
\item{type}{a device type, 'cuda' or 'cpu'.}

\item{index}{an index for the device (starting from 0). only used for 'cuda'
devices.}
}
\description{
Create a Device
}
