% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_diag}
\alias{tch_diag}
\title{Diag}
\usage{
tch_diag(x, diagonal = 0)
}
\arguments{
\item{x}{tensor object}

\item{diagonal}{the diagonal to consider}
}
\description{
If input is a vector (1-D tensor), then returns a 2-D square tensor with the
elements of input as the diagonal.
}
\details{
If input is a matrix (2-D tensor), then returns a 1-D tensor with the
diagonal elements of input.
\itemize{
\item If diagonal = 0, it is the main diagonal.
\item If diagonal > 0, it is above the main diagonal.
\item If diagonal < 0, it is below the main diagonal.
}
}
\examples{
tch_diag(tensor(1:10))
tch_diag(tensor(matrix(4, 2, 2)))
}
\seealso{
\code{\link[=tch_diagflat]{tch_diagflat()}} \code{\link[=tch_diagonal]{tch_diagonal()}}
}
