% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_diagflat}
\alias{tch_diagflat}
\title{Diagflat}
\usage{
tch_diagflat(x, offset = 0)
}
\arguments{
\item{x}{tensor object}

\item{offset}{the offset to consider}
}
\description{
If input is a vector (1-D tensor), then returns a 2-D square tensor with the
elements of input as the diagonal.
If input is a tensor with more than one dimension, then returns a 2-D tensor
with diagonal elements equal to a flattened input.
The argument offset controls which diagonal to consider:
}
\details{
\itemize{
\item If offset = 0, it is the main diagonal.
\item If offset > 0, it is above the main diagonal.
\item If offset < 0, it is below the main diagonal.
}
}
\examples{
tch_diagflat(tensor(1:10))
}
\seealso{
\code{\link[=tch_diag]{tch_diag()}} \code{\link[=tch_diagonal]{tch_diagonal()}}
}
