% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_diagonal}
\alias{tch_diagonal}
\title{Diagonal}
\usage{
tch_diagonal(x, offset = 0, dim1 = 0, dim2 = 1)
}
\arguments{
\item{x}{tensor object}

\item{offset}{which diagonal to consider}

\item{dim1}{first dimension with respect to which to take diagonal}

\item{dim2}{second dimension with respect to which to take diagonal}
}
\description{
Returns a partial view of input with the its diagonal elements with respect
to dim1 and dim2 appended as a dimension at the end of the shape.
}
\details{
The argument offset controls which diagonal to consider:

If offset = 0, it is the main diagonal.
If offset > 0, it is above the main diagonal.
If offset < 0, it is below the main diagonal.
}
\examples{
tch_diagonal(tensor(matrix(1:4, 2, 2)))
}
\seealso{
\code{\link[=tch_diag]{tch_diag()}} \code{\link[=tch_diagonal]{tch_diagonal()}}
}
