% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_digamma}
\alias{tch_digamma}
\title{Digamma}
\usage{
tch_digamma(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Computes the logarithmic derivative of the gamma function on input.
}
\examples{
tch_digamma(tensor(c(1, 0.5)))
}
