% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_dist}
\alias{tch_dist}
\title{Distance}
\usage{
tch_dist(x, other, p = 2)
}
\arguments{
\item{x}{tensor object}

\item{other}{the Right-hand-side input tensor}

\item{p}{the norm to be computed}
}
\description{
Returns the p-norm of (input - other)
The shapes of input and other must be broadcastable.
}
\examples{
x <- tensor(as.numeric(1:10))
y <- tensor(as.numeric(10:1))
tch_dist(x, y)
}
