% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_dot}
\alias{tch_dot}
\title{Dot product}
\usage{
tch_dot(x, tensor)
}
\arguments{
\item{x}{tensor object}

\item{tensor}{tensor object}
}
\description{
Computes the dot product (inner product) of two tensors.
}
\examples{
tch_dot(tensor(c(2,3)), tensor(c(2,1)))
}
