% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_eig}
\alias{tch_eig}
\title{Eigenvalues and Eigenvectors}
\usage{
tch_eig(x, eigenvectors = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{eigenvectors}{True to compute both eigenvalues and eigenvectors;
otherwise, only eigenvalues will be computed.}
}
\description{
Computes the eigenvalues and eigenvectors of a real square matrix.
}
\examples{
x <- tensor(matrix(runif(25), 5, 5))
tch_eig(x)

}
