% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_equal}
\alias{tch_equal}
\title{Equal}
\usage{
tch_equal(x, other)
}
\arguments{
\item{x}{tensor object}

\item{other}{tensor object to compare}
}
\description{
\code{TRUE} if two tensors have the same size and elements, \code{FALSE} otherwise.
}
\examples{
x <- tensor(c(1,2))
y <- tensor(c(1,2))
tch_equal(x, y)

}
