% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_eye}
\alias{tch_eye}
\title{Eye matrix (identity matrix)}
\usage{
tch_eye(n, m = n, dtype = NULL, layout = NULL, device = NULL,
  requires_grad = FALSE)
}
\arguments{
\item{n}{integer. The number of rows.}

\item{m}{(optional) integer. The number of columns with default being n.}

\item{dtype}{the desired data type of returned tensor. Default: if \code{NULL}, infers
data type from \code{x}.}

\item{device}{the desired device of returned tensor. Default: if \code{NULL}, uses
the current device for the default tensor type (see \code{tch_set_default_tensor_type()}).
device will be the CPU for CPU tensor types and the current CUDA device for
CUDA tensor types.}

\item{requires_grad}{If autograd should record operations on the
returned tensor. Default: \code{FALSE}.}

\item{the}{desired layout of returned Tensor. Default: 'strided'}
}
\description{
Returns a 2-D tensor with ones on the diagonal and zeros elsewhere.
}
\examples{
tch_eye(3)
tch_eye(2,4)


}
