% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_gels}
\alias{tch_gels}
\title{Gels}
\usage{
tch_gels(x, A)
}
\arguments{
\item{x}{tensor object}

\item{A}{the m,b by nn matrix A}
}
\description{
Computes the solution to the least squares and least norm problems for a full
rank matrix A of size (m,n) and a matrix B of size (m,k).
}
\examples{
A <- tensor(matrix(runif(100), ncol = 10))
y <- tensor(matrix(runif(10), ncol = 1))
tch_gels(y, A)
}
