% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_lerp}
\alias{tch_lerp}
\title{lerp}
\usage{
tch_lerp(start, end, weight)
}
\arguments{
\item{start}{the tensor with the starting points.}

\item{end}{the tensor with the ending points.}

\item{weight}{the weight for the interpolation formula.}
}
\description{
Does a linear interpolation of two tensors start and end based on a scalar weight and returns the resulting out tensor.
}
\details{
out = start - weight * (end - start)
}
\examples{
start <- tch_arange(1, 5)
end <- tch_empty(4)$fill_(10)
tch_lerp(start, end, 0.5)
}
