% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_logspace}
\alias{tch_logspace}
\title{Loglinear spaced tensor}
\usage{
tch_logspace(start, end, steps = 100, dtype = NULL, layout = NULL,
  device = NULL, requires_grad = FALSE)
}
\arguments{
\item{start}{the starting value for the set of points.}

\item{end}{the ending value for the set of points.}

\item{steps}{number of points to sample between start and end. Default: 100.}

\item{dtype}{the desired data type of returned tensor. Default: if \code{NULL}, infers
data type from \code{x}.}

\item{device}{the desired device of returned tensor. Default: if \code{NULL}, uses
the current device for the default tensor type (see \code{tch_set_default_tensor_type()}).
device will be the CPU for CPU tensor types and the current CUDA device for
CUDA tensor types.}

\item{requires_grad}{If autograd should record operations on the
returned tensor. Default: \code{FALSE}.}

\item{fill_value}{the number to fill the output tensor with.}

\item{the}{desired layout of returned Tensor. Default: 'strided'}
}
\description{
Returns a one-dimensional tensor of steps points logarithmically spaced between 10^start and 10^end.
}
\details{
The output tensor is 1-D of size steps.
}
\examples{
tch_linspace(3, 10, steps = 5)
tch_linspace(-10, 10, steps = 5)
tch_linspace(start = -10, end = 10, steps = 5)
tch_linspace(0, 1)

}
