% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_logsumexp}
\alias{tch_logsumexp}
\title{logsumexp}
\usage{
tch_logsumexp(x, dim = NULL, keepdim = NULL, dtype = NULL,
  na.rm = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{dim}{the dimension to reduce}

\item{keepdim}{wether to keep or not the dim}

\item{dtype}{optionaly cast the sum result}
}
\description{
Returns the log(sum(exp(x))) of all elements in the x tensor.
}
\examples{
x <- tensor(runif(10))
tch_logsumexp(x)
}
