% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_mean}
\alias{tch_mean}
\title{mean}
\usage{
tch_mean(x, dim = NULL, keepdim = FALSE)

tch_mean(x, dim = NULL, keepdim = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{dim}{the dimension to reduce}

\item{keepdim}{whether the output tensor has dim retained or not. (ignored if dim is \code{NULL})}

\item{x}{tensor object}

\item{dim}{the dimension to reduce}

\item{keepdim}{whether the output tensor has dim retained or not. (ignored if dim is \code{NULL})}
}
\description{
Returns the mean value of all elements in the x tensor.

Returns the mean value of all elements in the x tensor.
}
\examples{
x <- tensor(runif(100))
tch_mean(x)
x <- tensor(runif(100))
tch_min(x)
}
