% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_median}
\alias{tch_median}
\title{median}
\usage{
tch_median(x, dim = -1, values = NULL, indices = NULL,
  keepdim = NULL, dtype = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{dim}{the dimension to reduce}

\item{values}{the output tensor}

\item{indices}{the output index tensor}

\item{keepdim}{wether to keep or not the dim}

\item{dtype}{optionaly cast the sum result}
}
\description{
Returns the median value of each row of the input tensor in the given dimension dim.
Also returns the index location of the median value as a LongTensor.
By default, dim is the last dimension of the input tensor.
}
\examples{
x <- tch_randn(4, 5)
tch_median(x)
}
