% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_min}
\alias{tch_min}
\title{min}
\usage{
tch_min(x, dim = NULL, keepdim = FALSE, other = NULL)
}
\arguments{
\item{x}{tensor object}

\item{dim}{the dimension to reduce}

\item{keepdim}{wether to keep or not the dim}

\item{other}{a tensor.}
}
\description{
Returns the minimum value of all elements in the x tensor.
}
\examples{
x <- tensor(runif(100))
tch_min(x)
}
