% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_permute}
\alias{tch_permute}
\title{permute}
\usage{
tch_permute(x, dims)
}
\arguments{
\item{x}{tensor object}

\item{dims}{the desired ordering of dimensions (0 based).}
}
\description{
Permute the dimensions of this tensor.
}
\examples{
x <- tensor(array(1:10, dim = c(2,5,1)))
x
tch_permute(x, c(2,1,0))
}
