% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_prod}
\alias{tch_prod}
\title{prod}
\usage{
tch_prod(x, dim = NULL, keepdim = NULL, dtype = NULL)
}
\arguments{
\item{x}{tensor object}

\item{dim}{the dimension to reduce}

\item{keepdim}{wether to keep or not the dim}

\item{dtype}{optionaly cast the sum result}
}
\description{
Returns the product of all elements in the x tensor.
}
\examples{
x <- tensor(runif(10))
tch_prod(x)
}
