% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_qr}
\alias{tch_qr}
\title{QR decomposition}
\usage{
tch_qr(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Computes the QR decomposition of a matrix input, and returns matrices Q and R
such that $input = QR$, with QQ being an orthogonal matrix and RR being an
upper triangular matrix.
}
\details{
This returns the thin (reduced) QR factorization.
}
\examples{
x <- tensor(matrix(runif(16), ncol = 4))
tch_qr(x)
}
