% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_randint}
\alias{tch_randint}
\title{Random discrete uniform}
\usage{
tch_randint(low = 0, high = NULL, sizes = NULL, dtype = NULL,
  layout = NULL, device = NULL, requires_grad = FALSE)
}
\arguments{
\item{low}{(optional) Lowest integer to be drawn from the distribution. Default: 0.}

\item{sizes}{a sequence of integers defining the shape of the output tensor.}

\item{dtype}{the desired data type of returned tensor. Default: if \code{NULL}, infers
data type from \code{x}.}

\item{device}{the desired device of returned tensor. Default: if \code{NULL}, uses
the current device for the default tensor type (see \code{tch_set_default_tensor_type()}).
device will be the CPU for CPU tensor types and the current CUDA device for
CUDA tensor types.}

\item{requires_grad}{If autograd should record operations on the
returned tensor. Default: \code{FALSE}.}

\item{hogh}{One above the highest integer to be drawn from the distribution.}

\item{the}{desired layout of returned Tensor. Default: 'strided'}
}
\description{
Returns a tensor filled with random integers generated uniformly between low (inclusive) and high (exclusive).
}
\details{
The shape of the tensor is defined by the variable argument size.
}
\examples{
tch_randint(3, 5, (3,))
tch_randint(10, (2, 2))
tch_randint(3, 10, (2, 2))

}
