% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_round}
\alias{tch_round}
\title{round}
\usage{
tch_round(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with each of the elements of x rounded to the closest integer.
}
\examples{
x <- tensor(matrix(runif(6), nrow = 3))
tch_round(x)
}
