% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_rsqrt}
\alias{tch_rsqrt}
\title{rsqrt}
\usage{
tch_rsqrt(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the reciprocal of the square-root of each of the elements of x: 1/sqrt(x).
}
\examples{
x <- tensor(matrix(runif(6), nrow = 3))
tch_rsqrt(x)
}
