% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_sigmoid}
\alias{tch_sigmoid}
\title{sigmoid}
\usage{
tch_sigmoid(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the sigmoid of the elements of each of the elements of x: 1/(1 + exp(-x)).
}
\examples{
x <- tensor(matrix(rnorm(6), nrow = 3))
tch_sigmoid(x)
}
