% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_sign}
\alias{tch_sign}
\title{sign}
\usage{
tch_sign(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the sign of the elements of x.
}
\examples{
x <- tensor(matrix(rnorm(6), nrow = 3))
tch_sign(x)
}
