% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_sin}
\alias{tch_sin}
\title{sin}
\usage{
tch_sin(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the sine of the elements of input.
}
\examples{
x <- tensor(array(runif(8), dim = c(2,2,2)))
tch_sin(x)
}
