% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_sinh}
\alias{tch_sinh}
\title{sinh}
\usage{
tch_sinh(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the hyperbolic sine of the elements of input.
}
\examples{
x <- tensor(array(runif(8), dim = c(2,2,2)))
tch_asin(x)
}
