% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_sqrt}
\alias{tch_sqrt}
\title{sqrt}
\usage{
tch_sqrt(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the square-root of the elements of x
}
\examples{
x <- tensor(matrix(rnorm(6), nrow = 3))
tch_sqrt(x)
}
