% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_tan}
\alias{tch_tan}
\title{tan}
\usage{
tch_tan(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the tangent of the elements of input.
}
\examples{
x <- tensor(array(runif(8), dim = c(2,2,2)))
tch_tan(x)
}
