% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_tanh}
\alias{tch_tanh}
\title{tanh}
\usage{
tch_tanh(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the hyperbolic tangent of the elements of input.
}
\examples{
x <- tensor(array(runif(8), dim = c(2,2,2)))
tch_tanh(x)
}
