% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_triu}
\alias{tch_triu}
\title{Upper triangle}
\usage{
tch_triu(x, diagonal = 0)
}
\arguments{
\item{x}{tensor object}

\item{diagonal}{integer. the diagonal to consider}
}
\description{
Returns the upper triangular part of the matrix (2-D tensor) x, the other elements of the result tensor out are set to 0.
The upper triangular part of the matrix is defined as the elements on and above the diagonal.
}
\details{
The argument diagonal controls which diagonal to consider. If diagonal = 0, all elements on and above the main diagonal are retained.
A positive value excludes just as many diagonals above the main diagonal, and similarly a negative value includes just as many diagonals below the main diagonal.
}
\examples{
x <- array(1:20, c(4, 5))
tch_triu(tensor(x))
tch_triu(tensor(x), 1)
tch_triu(tensor(x), 2)
tch_triu(tensor(x), -1)
tch_triu(tensor(x), -2)
}
