% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_trunc}
\alias{tch_trunc}
\title{trunc}
\usage{
tch_trunc(x)
}
\arguments{
\item{x}{tensor object}
}
\description{
Returns a new tensor with the truncated integer values of the elements of x.
}
\examples{
x <- tensor(array(runif(8), dim = c(2,2,2)))
tch_trunc(x)
}
