% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_var}
\alias{tch_var}
\title{var}
\usage{
tch_var(x, unbiased = TRUE, dim = NULL, keepdim = NULL,
  dtype = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{unbiased}{whether to use the unbiased estimation or not}

\item{dim}{the dimension to reduce}

\item{keepdim}{wether to keep or not the dim}

\item{dtype}{optionaly cast the sum result}
}
\description{
Returns the variance of all elements in the x tensor.
If unbiased is FALSE, then the variance will be calculated via the biased estimator. Otherwise, Bessel’s correction will be used.
}
\examples{
x <- tensor(runif(100))
tch_var(x)
}
