% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensor}
\alias{tensor}
\title{Creates a torch tensor.}
\usage{
tensor(x, ...)
}
\arguments{
\item{x}{an R object or a torch tensor.}

\item{dtype}{a string with torch types. Allowed types are double, float and int.}

\item{device}{a device type}

\item{requires_grad}{boolean indicating if tensor requires grad.}
}
\description{
Creates a torch tensor.
}
\examples{
x <- tensor(1:10)
x

y <- tensor(x, dtype = "double")
y
}
