% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensor_from_buffer}
\alias{tensor_from_buffer}
\title{Creates a tensor from a buffer of memory}
\usage{
tensor_from_buffer(buffer, shape, dtype = "float")
}
\arguments{
\item{buffer}{An R atomic object containing the data in a contiguous array.}

\item{shape}{The shape of the resulting tensor.}

\item{dtype}{A torch data type for the tresulting tensor.}
}
\description{
It creates a tensor without taking ownership of the memory it points to.
You must call \code{clone} if you want to copy the memory over a new tensor.
}
