% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{tensor_from_r}
\alias{tensor_from_r}
\title{Create torch Tensor from R object}
\usage{
tensor_from_r(x, dtype = NULL, device = NULL, requires_grad = FALSE)
}
\arguments{
\item{x}{an R vector, matrix or array.}

\item{dtype}{dtype}

\item{device}{device}

\item{requires_grad}{requires_grad}
}
\description{
Create torch Tensor from R object
}
\note{
it uses the R type when creating the tensor.
}
\examples{
tensor_from_r(1:10)
tensor_from_r(array(runif(8), dim = c(2, 2, 2)))
tensor_from_r(matrix(c(TRUE, FALSE), nrow = 3, ncol = 4))
}
