% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{threads}
\alias{threads}
\alias{torch_set_num_threads}
\alias{torch_set_num_interop_threads}
\alias{torch_get_num_interop_threads}
\alias{torch_get_num_threads}
\title{Number of threads}
\usage{
torch_set_num_threads(num_threads)

torch_set_num_interop_threads(num_threads)

torch_get_num_interop_threads()

torch_get_num_threads()
}
\arguments{
\item{num_threads}{number of threads to set.}
}
\description{
Get and set the numbers used by torch computations.
}
\details{
For details see the \href{https://pytorch.org/docs/stable/notes/cpu_threading_torchscript_inference.html?highlight=set_num_threads}{CPU threading article}
in the PyTorch documentation.
}
\note{
torch_set_threads do not work on macOS system as it must be 1.
}
