% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{*.tensor}
\alias{*.tensor}
\title{elementwise multiplciation}
\usage{
\method{*}{tensor}(x, y)
}
\arguments{
\item{x}{tensor object}

\item{other}{tensor object}
}
\description{
elementwise multiplciation
}
\examples{
x <- tensor(2)
y <- tensor(3)
x * y
}
