% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_absolute}
\alias{torch_absolute}
\title{Absolute}
\usage{
torch_absolute(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Absolute
}
\section{absolute(input, *, out=None) -> Tensor}{


Alias for \code{\link[=torch_abs]{torch_abs()}}
}

