% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_acos}
\alias{torch_acos}
\title{Acos}
\usage{
torch_acos(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Acos
}
\section{acos(input) -> Tensor }{


Returns a new tensor with the arccosine  of the elements of \code{input}.

\deqn{
    \mbox{out}_{i} = \cos^{-1}(\mbox{input}_{i})
}
}

\examples{

a = torch_randn(c(4))
a
torch_acos(a)
}
