% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_adaptive_avg_pool1d}
\alias{torch_adaptive_avg_pool1d}
\title{Adaptive_avg_pool1d}
\arguments{
\item{output_size}{NA the target output size (single integer)}
}
\description{
adaptive_avg_pool1d(input, output_size) -> Tensor
}
\details{
Applies a 1D adaptive average pooling over an input signal composed of
several input planes.

See :class:\code{~torch.nn.AdaptiveAvgPool1d} for details and output shape.
}
