% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_addcdiv}
\alias{torch_addcdiv}
\title{Addcdiv}
\arguments{
\item{input}{(Tensor) the tensor to be added}

\item{value}{(Number, optional) multiplier for \eqn{\text{tensor1} / \text{tensor2}}}

\item{tensor1}{(Tensor) the numerator tensor}

\item{tensor2}{(Tensor) the denominator tensor}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Addcdiv
}
\section{addcdiv(input, value=1, tensor1, tensor2, out=None) -> Tensor }{


Performs the element-wise division of \code{tensor1} by \code{tensor2},
multiply the result by the scalar \code{value} and add it to \code{input}.

\deqn{
    \text{out}_i = \text{input}_i + \text{value} \times \frac{\text{tensor1}_i}{\text{tensor2}_i}
}
The shapes of \code{input}, \code{tensor1}, and \code{tensor2} must be
broadcastable .

For inputs of type \code{FloatTensor} or \code{DoubleTensor}, \code{value} must be
a real number, otherwise an integer.
}

\examples{

t = torch_randn(c(1, 3))
t1 = torch_randn(c(3, 1))
t2 = torch_randn(c(1, 3))
torch_addcdiv(t, t1, t2, 0.1)
}
