% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_addcmul}
\alias{torch_addcmul}
\title{Addcmul}
\arguments{
\item{input}{(Tensor) the tensor to be added}

\item{value}{(Number, optional) multiplier for :math:\verb{tensor1 .* tensor2}}

\item{tensor1}{(Tensor) the tensor to be multiplied}

\item{tensor2}{(Tensor) the tensor to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
addcmul(input, value=1, tensor1, tensor2, out=None) -> Tensor
}
\details{
Performs the element-wise multiplication of :attr:\code{tensor1}
by :attr:\code{tensor2}, multiply the result by the scalar :attr:\code{value}
and add it to :attr:\code{input}.

.. math::
\text{out}_i = \text{input}_i + \text{value} \times \text{tensor1}_i \times \text{tensor2}_i

The shapes of :attr:\code{tensor}, :attr:\code{tensor1}, and :attr:\code{tensor2} must be
:ref:\verb{broadcastable <broadcasting-semantics>}.

For inputs of type \code{FloatTensor} or \code{DoubleTensor}, :attr:\code{value} must be
a real number, otherwise an integer.
}
\examples{
\dontrun{
t = torch_randn(1, 3)
t1 = torch_randn(3, 1)
t2 = torch_randn(1, 3)
torch_addcmul(t, 0.1, t1, t2)
}

}
