% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_addmm}
\alias{torch_addmm}
\title{Addmm}
\arguments{
\item{beta}{(Number, optional) multiplier for :attr:\code{input} (:math:\verb{\\beta})}

\item{input}{(Tensor) matrix to be added}

\item{alpha}{(Number, optional) multiplier for :math:\code{mat1 @ mat2} (:math:\verb{\\alpha})}

\item{mat1}{(Tensor) the first matrix to be multiplied}

\item{mat2}{(Tensor) the second matrix to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
addmm(beta=1, input, alpha=1, mat1, mat2, out=None) -> Tensor
}
\details{
Performs a matrix multiplication of the matrices :attr:\code{mat1} and :attr:\code{mat2}.
The matrix :attr:\code{input} is added to the final result.

If :attr:\code{mat1} is a :math:\verb{(n \\times m)} tensor, :attr:\code{mat2} is a
:math:\verb{(m \\times p)} tensor, then :attr:\code{input} must be
:ref:\verb{broadcastable <broadcasting-semantics>} with a :math:\verb{(n \\times p)} tensor
and :attr:\code{out} will be a :math:\verb{(n \\times p)} tensor.

:attr:\code{alpha} and :attr:\code{beta} are scaling factors on matrix-vector product between
:attr:\code{mat1} and :attr:\code{mat2} and the added matrix :attr:\code{input} respectively.

.. math::
\text{out} = \beta\ \text{input} + \alpha\ (\text{mat1}_i \mathbin{@} \text{mat2}_i)

For inputs of type \code{FloatTensor} or \code{DoubleTensor}, arguments :attr:\code{beta} and
:attr:\code{alpha} must be real numbers, otherwise they should be integers.
}
