% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_addmv}
\alias{torch_addmv}
\title{Addmv}
\arguments{
\item{beta}{(Number, optional) multiplier for :attr:\code{input} (:math:\verb{\\beta})}

\item{input}{(Tensor) vector to be added}

\item{alpha}{(Number, optional) multiplier for :math:\code{mat @ vec} (:math:\verb{\\alpha})}

\item{mat}{(Tensor) matrix to be multiplied}

\item{vec}{(Tensor) vector to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
addmv(beta=1, input, alpha=1, mat, vec, out=None) -> Tensor
}
\details{
Performs a matrix-vector product of the matrix :attr:\code{mat} and
the vector :attr:\code{vec}.
The vector :attr:\code{input} is added to the final result.

If :attr:\code{mat} is a :math:\verb{(n \\times m)} tensor, :attr:\code{vec} is a 1-D tensor of
size \code{m}, then :attr:\code{input} must be
:ref:\verb{broadcastable <broadcasting-semantics>} with a 1-D tensor of size \code{n} and
:attr:\code{out} will be 1-D tensor of size \code{n}.

:attr:\code{alpha} and :attr:\code{beta} are scaling factors on matrix-vector product between
:attr:\code{mat} and :attr:\code{vec} and the added tensor :attr:\code{input} respectively.

.. math::
\text{out} = \beta\ \text{input} + \alpha\ (\text{mat} \mathbin{@} \text{vec})

For inputs of type \code{FloatTensor} or \code{DoubleTensor}, arguments :attr:\code{beta} and
:attr:\code{alpha} must be real numbers, otherwise they should be integers
}
\examples{
\dontrun{
M = torch_randn(2)
mat = torch_randn(2, 3)
vec = torch_randn(3)
torch_addmv(M, mat, vec)
}

}
