% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_arange}
\alias{torch_arange}
\title{Arange}
\arguments{
\item{start}{(Number) the starting value for the set of points. Default: \code{0}.}

\item{end}{(Number) the ending value for the set of points}

\item{step}{(Number) the gap between each pair of adjacent points. Default: \code{1}.}

\item{out}{(Tensor, optional) the output tensor.}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{arguments.}{NA}

\item{`dtype`}{NA}

\item{be}{NA}

\item{layout}{NA class:\code{torch.layout}, optional): the desired layout of returned Tensor.}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if \code{None}, uses the current device for the default tensor type}

\item{NA}{func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU}

\item{for}{NA}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}
}
\description{
arange(start=0, end, step=1, out=None, dtype=None, layout=torch.strided, device=None, requires_grad=False) -> Tensor
}
\details{
Returns a 1-D tensor of size :math:\verb{\\left\\lceil \\frac\{\\text\{end\} - \\text\{start\}\}\{\\text\{step\}\} \\right\\rceil}
with values from the interval \verb{[start, end)} taken with common difference
:attr:\code{step} beginning from \code{start}.

Note that non-integer :attr:\code{step} is subject to floating point rounding errors when
comparing against :attr:\code{end}; to avoid inconsistency, we advise adding a small epsilon to :attr:\code{end}
in such cases.

.. math::
\text{out}\emph{{{i+1}} = \text{out}}{i} + \text{step}
}
