% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_arccos}
\alias{torch_arccos}
\title{Arccos}
\usage{
torch_arccos(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Arccos
}
\section{arccos(input, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_acos]{torch_acos()}}.
}

