% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_arcsin}
\alias{torch_arcsin}
\title{Arcsin}
\usage{
torch_arcsin(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Arcsin
}
\section{arcsin(input, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_asin]{torch_asin()}}.
}

