% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-new-examples.R, R/gen-namespace.R
\name{torch_arcsinh}
\alias{torch_arcsinh}
\title{Arcsinh}
\usage{
torch_arcsinh(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Arcsinh
}
\section{arcsinh(input, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_asinh]{torch_asinh()}}.
}

