% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_arctan}
\alias{torch_arctan}
\title{Arctan}
\usage{
torch_arctan(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Arctan
}
\section{arctan(input, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_atan]{torch_atan()}}.
}

