% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_arctanh}
\alias{torch_arctanh}
\title{Arctanh}
\usage{
torch_arctanh(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Arctanh
}
\section{arctanh(input, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_atanh]{torch_atanh()}}.
}

