% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_argmin}
\alias{torch_argmin}
\title{Argmin}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int) the dimension to reduce. If \code{None}, the argmin of the flattened input is returned.}

\item{keepdim}{(bool) whether the output tensor has :attr:\code{dim} retained or not. Ignored if \code{dim=None}.}
}
\description{
Returns the indices of the minimum value of all elements in the \code{input} tensor.

Returns the indices of the minimum values of a tensor across a dimension.
}
\details{
This is the second value returned by :meth:\code{torch_min}. See its
documentation for the exact semantics of this method.

This is the second value returned by :meth:\code{torch_min}. See its
documentation for the exact semantics of this method.
}
\section{Signatures}{


argmin(input) -> LongTensor



argmin(input, dim, keepdim=False, out=None) -> LongTensor
}

\examples{
\dontrun{
a = torch_randn(4, 4)
a
torch_argmin(a)
}

\dontrun{
a = torch_randn(4, 4)
a
torch_argmin(a, dim=1)
}

}
