% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_argsort}
\alias{torch_argsort}
\title{Argsort}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int, optional) the dimension to sort along}

\item{descending}{(bool, optional) controls the sorting order (ascending or descending)}
}
\description{
Returns the indices that sort a tensor along a given dimension in ascending
order by value.
}
\details{
This is the second value returned by :meth:\code{torch_sort}.  See its documentation
for the exact semantics of this method.
}
\section{Signatures}{


argsort(input, dim=-1, descending=False, out=None) -> LongTensor
}

\examples{
\dontrun{
a = torch_randn(4, 4)
a
torch_argsort(a, dim=1)
}

}
