% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_asin}
\alias{torch_asin}
\title{Asin}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
asin(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the arcsine  of the elements of :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \sin^{-1}(\text{input}}{i})
}
