% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_asinh}
\alias{torch_asinh}
\title{Asinh}
\usage{
torch_asinh(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Asinh
}
\section{asinh(input, *, out=None) -> Tensor }{


Returns a new tensor with the inverse hyperbolic sine of the elements of \code{input}.

\deqn{
    \mbox{out}_{i} = \sinh^{-1}(\mbox{input}_{i})
}
}

\examples{

a <- torch_randn(c(4))
a
torch_asinh(a)
}
