% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_atan}
\alias{torch_atan}
\title{Atan}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the arctangent  of the elements of \code{input}.
}
\details{
\deqn{
    \text{out}_{i} = \tan^{-1}(\text{input}_{i})
}
}
\section{Signatures}{


atan(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(4)
a
torch_atan(a)
}

}
