% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_atan2}
\alias{torch_atan2}
\title{Atan2}
\arguments{
\item{input}{(Tensor) the first input tensor}

\item{other}{(Tensor) the second input tensor}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Element-wise arctangent of :math:\verb{\\text\{input\}_\{i\} / \\text\{other\}_\{i\}}
with consideration of the quadrant. Returns a new tensor with the signed angles
in radians between vector :math:\verb{(\\text\{other\}_\{i\}, \\text\{input\}_\{i\})}
and vector :math:\verb{(1, 0)}. (Note that :math:\verb{\\text\{other\}_\{i\}}, the second
parameter, is the x-coordinate, while :math:\verb{\\text\{input\}_\{i\}}, the first
parameter, is the y-coordinate.)
}
\details{
The shapes of \code{input} and \code{other} must be
:ref:\verb{broadcastable <broadcasting-semantics>}.
}
\section{Signatures}{


atan2(input, other, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(4)
a
torch_atan2(a, torch_randn(4))
}

}
