% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R
\name{torch_atleast_2d}
\alias{torch_atleast_2d}
\title{Atleast_2d}
\arguments{
\item{self}{(Tensor or list of Tensors)}
}
\description{
Returns a 2-dimensional view of each each input tensor with zero dimensions.
Input tensors with two or more dimensions are returned as-is.
}
