% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_atleast_3d}
\alias{torch_atleast_3d}
\title{Atleast_3d}
\usage{
torch_atleast_3d(self, tensors)
}
\arguments{
\item{input}{(Tensor or list of Tensors)}
}
\description{
Returns a 3-dimensional view of each each input tensor with zero dimensions.
Input tensors with three or more dimensions are returned as-is.
}
