% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_avg_pool1d}
\alias{torch_avg_pool1d}
\title{Avg_pool1d}
\arguments{
\item{input}{NA input tensor of shape :math:\verb{(\\text\{minibatch\} , \\text\{in\\_channels\} , iW)}}

\item{kernel_size}{NA the size of the window. Can be a single number or a}

\item{tuple}{(kW,)}

\item{stride}{NA the stride of the window. Can be a single number or a tuple}

\item{`}{(sW,) :attr:\code{kernel_size}}

\item{padding}{NA implicit zero paddings on both sides of the input. Can be a}

\item{single}{(padW,) 0}

\item{ceil_mode}{NA when True, will use \code{ceil} instead of \code{floor} to compute the}

\item{output}{NA \code{False}}

\item{count_include_pad}{NA when True, will include the zero-padding in the}

\item{averaging}{NA \code{True}}
}
\description{
avg_pool1d(input, kernel_size, stride=None, padding=0, ceil_mode=False, count_include_pad=True) -> Tensor
}
\details{
Applies a 1D average pooling over an input signal composed of several
input planes.

See :class:\code{~torch.nn.AvgPool1d} for details and output shape.
}
