% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_bitwise_not}
\alias{torch_bitwise_not}
\title{Bitwise_not}
\usage{
torch_bitwise_not(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Bitwise_not
}
\section{bitwise_not(input, out=NULL) -> Tensor }{


Computes the bitwise NOT of the given input tensor. The input tensor must be of
integral or Boolean types. For bool tensors, it computes the logical NOT.
}

