% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_bitwise_xor}
\alias{torch_bitwise_xor}
\title{Bitwise_xor}
\usage{
torch_bitwise_xor(self, other)
}
\arguments{
\item{self}{NA the first input tensor}

\item{other}{NA the second input tensor}
}
\description{
Bitwise_xor
}
\section{bitwise_xor(input, other, out=NULL) -> Tensor }{


Computes the bitwise XOR of \code{input} and \code{other}. The input tensor must be of
integral or Boolean types. For bool tensors, it computes the logical XOR.
}

