% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_block_diag}
\alias{torch_block_diag}
\title{Block_diag}
\usage{
torch_block_diag(tensors)
}
\arguments{
\item{tensors}{(list of tensors) One or more tensors with 0, 1, or 2
dimensions.}
}
\description{
Create a block diagonal matrix from provided tensors.
}
