% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_broadcast_tensors}
\alias{torch_broadcast_tensors}
\title{Broadcast_tensors}
\description{
broadcast_tensors(*tensors) -> List of Tensors
}
\details{
\preformatted{Broadcasts the given tensors according to :ref:`broadcasting-semantics`.

Args:
    *tensors: any number of tensors of the same type
}
}
