% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_bucketize}
\alias{torch_bucketize}
\title{Bucketize}
\usage{
torch_bucketize(self, boundaries, out_int32 = FALSE, right = FALSE)
}
\arguments{
\item{self}{(Tensor or Scalar) N-D tensor or a Scalar containing the search value(s).}

\item{boundaries}{(Tensor) 1-D tensor, must contain a monotonically increasing sequence.}
}
\description{
Bucketize
}
\section{bucketize(input, boundaries, *, out_int32=FALSE, right=FALSE, out=None) -> Tensor }{


Returns the indices of the buckets to which each value in the \code{input} belongs, where the
boundaries of the buckets are set by \code{boundaries}. Return a new tensor with the same size
as \code{input}. If \code{right} is FALSE (default), then the left boundary is closed.
}

