% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_can_cast}
\alias{torch_can_cast}
\title{Can_cast}
\arguments{
\item{from}{(dtype) The original :class:\code{torch.dtype}.}

\item{to}{(dtype) The target :class:\code{torch.dtype}.}
}
\description{
can_cast(from, to) -> bool
}
\details{
Determines if a type conversion is allowed under PyTorch casting rules
described in the type promotion :ref:\verb{documentation <type-promotion-doc>}.
}
\examples{
\dontrun{
torch_can_cast(torch_double, torch_float)
torch_can_cast(torch_float, torch_int)
}

}
