% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_cdist}
\alias{torch_cdist}
\title{Cdist}
\description{
Computes batched the p-norm distance between each pair of the two collections of row vectors.
}
\details{
\preformatted{Args:
    x1 (Tensor): input tensor of shape :math:`B \\times P \\times M`.
    x2 (Tensor): input tensor of shape :math:`B \\times R \\times M`.
    p: p value for the p-norm distance to calculate between each vector pair
}
}
