% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_ceil}
\alias{torch_ceil}
\title{Ceil}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Ceil
}
\section{ceil(input, out=None) -> Tensor }{


Returns a new tensor with the ceil of the elements of \code{input},
the smallest integer greater than or equal to each element.

\deqn{
    \mbox{out}_{i} = \left\lceil \mbox{input}_{i} \right\rceil = \left\lfloor \mbox{input}_{i} \right\rfloor + 1
}
}

\examples{
if (torch_is_installed()) {

a = torch_randn(c(4))
a
torch_ceil(a)
}
}
