% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_celu}
\alias{torch_celu}
\title{Celu}
\usage{
torch_celu(self, alpha = 1L)
}
\arguments{
\item{self}{the input tensor}

\item{alpha}{the alpha value for the CELU formulation. Default: 1.0}
}
\description{
Celu
}
\section{celu(input, alpha=1.) -> Tensor }{


See \code{\link[=nnf_celu]{nnf_celu()}} for more info.
}

