% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_celu_}
\alias{torch_celu_}
\title{Celu_}
\description{
Celu_
}
\section{celu_(input, alpha=1.) -> Tensor }{


In-place version of \code{\link[=torch_celu]{torch_celu()}}.
}

