% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_chain_matmul}
\alias{torch_chain_matmul}
\title{Chain_matmul}
\description{
Returns the matrix product of the :math:\code{N} 2-D tensors. This product is efficiently computed
using the matrix chain order algorithm which selects the order in which incurs the lowest cost in terms
of arithmetic operations (\verb{[CLRS]}_). Note that since this is a function to compute the product, :math:\code{N}
needs to be greater than or equal to 2; if equal to 2 then a trivial matrix-matrix product is returned.
If :math:\code{N} is 1, then this is a no-op - the original matrix is returned as is.
}
