% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_cholesky_inverse}
\alias{torch_cholesky_inverse}
\title{Cholesky_inverse}
\arguments{
\item{input}{(Tensor) the input 2-D tensor :math:\code{u}, a upper or lower triangular           Cholesky factor}

\item{upper}{(bool, optional) whether to return a lower (default) or upper triangular matrix}

\item{out}{(Tensor, optional) the output tensor for \code{inv}}
}
\description{
cholesky_inverse(input, upper=False, out=None) -> Tensor
}
\details{
Computes the inverse of a symmetric positive-definite matrix :math:\code{A} using its
Cholesky factor :math:\code{u}: returns matrix \code{inv}. The inverse is computed using
LAPACK routines \code{dpotri} and \code{spotri} (and the corresponding MAGMA routines).

If :attr:\code{upper} is \code{False}, :math:\code{u} is lower triangular
such that the returned tensor is

.. math::
inv = (uu^{{T}})^{{-1}}

If :attr:\code{upper} is \code{True} or not provided, :math:\code{u} is upper
triangular such that the returned tensor is

.. math::
inv = (u^T u)^{{-1}}
}
