% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_clip}
\alias{torch_clip}
\title{Clip}
\usage{
torch_clip(self, min = NULL, max = NULL)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{min}{(Number) lower-bound of the range to be clamped to}

\item{max}{(Number) upper-bound of the range to be clamped to}
}
\description{
Clip
}
\section{clip(input, min, max, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_clamp]{torch_clamp()}}.
}

