% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_conj}
\alias{torch_conj}
\title{Conj}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Computes the element-wise conjugate of the given \code{input} tensor.
}
\details{
\deqn{
    \text{out}_{i} = conj(\text{input}_{i})
}
}
\section{Signatures}{


conj(input, out=None) -> Tensor
}

\examples{
\dontrun{
torch_conj(torch_tensor([-1 + 1j, -2 + 2j, 3 - 3j]))
}

}
