% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_conv_tbc}
\alias{torch_conv_tbc}
\title{Conv_tbc}
\arguments{
\item{input}{NA input tensor of shape \eqn{(\text{sequence length} \times batch \times \text{in\_channels})}}

\item{weight}{NA filter of shape (\eqn{\text{kernel width} \times \text{in\_channels} \times \text{out\_channels}})}

\item{bias}{NA bias of shape (\eqn{\text{out\_channels}})}

\item{pad}{NA number of timesteps to pad. Default: 0}
}
\description{
Conv_tbc
}
\section{TEST }{


Applies a 1-dimensional sequence convolution over an input sequence.
Input and output dimensions are (Time, Batch, Channels) - hence TBC.
}

