% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_conv_transpose2d}
\alias{torch_conv_transpose2d}
\title{Conv_transpose2d}
\arguments{
\item{input}{NA input tensor of shape :math:\verb{(\\text\{minibatch\} , \\text\{in\\_channels\} , iH , iW)}}

\item{weight}{NA filters of shape :math:\verb{(\\text\{in\\_channels\} , \\frac\{\\text\{out\\_channels\}\}\{\\text\{groups\}\} , kH , kW)}}

\item{bias}{NA optional bias of shape :math:\verb{(\\text\{out\\_channels\})}. Default: None}

\item{stride}{NA the stride of the convolving kernel. Can be a single number or a}

\item{tuple}{(sH, sW) 1}

\item{padding}{NA \code{dilation * (kernel_size - 1) - padding} zero-padding will be added to both}

\item{sides}{NA}

\item{``}{(padH, padW) 0}

\item{output_padding}{NA additional size added to one side of each dimension in the}

\item{output}{(out_padH, out_padW)}

\item{Default}{NA 0}

\item{groups}{NA split input into groups, :math:\verb{\\text\{in\\_channels\}} should be divisible by the}

\item{number}{NA 1}

\item{dilation}{NA the spacing between kernel elements. Can be a single number or}

\item{a}{(dH, dW) 1}
}
\description{
conv_transpose2d(input, weight, bias=None, stride=1, padding=0, output_padding=0, groups=1, dilation=1) -> Tensor
}
\details{
Applies a 2D transposed convolution operator over an input image
composed of several input planes, sometimes also called "deconvolution".

See :class:\code{~torch.nn.ConvTranspose2d} for details and output shape.

.. include:: cudnn_deterministic.rst
}
