% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_cos}
\alias{torch_cos}
\title{Cos}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the cosine  of the elements of \code{input}.
}
\details{
\deqn{
    \text{out}_{i} = \cos(\text{input}_{i})
}
}
\section{Signatures}{


cos(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(4)
a
torch_cos(a)
}

}
