% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_cross}
\alias{torch_cross}
\title{Cross}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{other}{(Tensor) the second input tensor}

\item{dim}{(int, optional) the dimension to take the cross-product in.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
cross(input, other, dim=-1, out=None) -> Tensor
}
\details{
Returns the cross product of vectors in dimension :attr:\code{dim} of :attr:\code{input}
and :attr:\code{other}.

:attr:\code{input} and :attr:\code{other} must have the same size, and the size of their
:attr:\code{dim} dimension should be 3.

If :attr:\code{dim} is not given, it defaults to the first dimension found with the
size 3.
}
\examples{
\dontrun{
a = torch_randn(4, 3)
a
b = torch_randn(4, 3)
b
torch_cross(a, b, dim=1)
torch_cross(a, b)
}

}
