% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_cumsum}
\alias{torch_cumsum}
\title{Cumsum}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int) the dimension to do the operation over}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{If}{NA attr:\code{dtype} before the operation}

\item{is}{NA None.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
cumsum(input, dim, out=None, dtype=None) -> Tensor
}
\details{
Returns the cumulative sum of elements of :attr:\code{input} in the dimension
:attr:\code{dim}.

For example, if :attr:\code{input} is a vector of size N, the result will also be
a vector of size N, with elements.

.. math::
y_i = x_1 + x_2 + x_3 + \dots + x_i
}
