% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_dequantize}
\alias{torch_dequantize}
\title{Dequantize}
\usage{
torch_dequantize(self, tensors)
}
\arguments{
\item{tensors}{(sequence of Tensors) A list of quantized Tensors}

\item{tensor}{(Tensor) A quantized Tensor}
}
\description{
Dequantize
}
\section{dequantize(tensor) -> Tensor }{


Returns an fp32 Tensor by dequantizing a quantized Tensor
}

\section{dequantize(tensors) -> sequence of Tensors }{


Given a list of quantized Tensors, dequantize them and return a list of fp32 Tensors
}

