% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_det}
\alias{torch_det}
\title{Det}
\arguments{
\item{input}{(Tensor) the input tensor of size \verb{(*, n, n)} where \code{*} is zero or more                batch dimensions.}
}
\description{
Calculates determinant of a square matrix or batches of square matrices.
}
\details{
.. note::
Backward through :meth:\code{det} internally uses SVD results when \code{input} is
not invertible. In this case, double backward through :meth:\code{det} will be
unstable in when \code{input} doesn't have distinct singular values. See
:meth:\code{~torch.svd} for details.
}
\section{Signatures}{


det(input) -> Tensor
}

\examples{
\dontrun{
A = torch_randn(3, 3)
torch_det(A)
A = torch_randn(3, 2, 2)
A
A.det()
}

}
