% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_diag}
\alias{torch_diag}
\title{Diag}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{diagonal}{(int, optional) the diagonal to consider}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
diag(input, diagonal=0, out=None) -> Tensor
}
\details{
\itemize{
\item If :attr:\code{input} is a vector (1-D tensor), then returns a 2-D square tensor
with the elements of :attr:\code{input} as the diagonal.
\item If :attr:\code{input} is a matrix (2-D tensor), then returns a 1-D tensor with
the diagonal elements of :attr:\code{input}.
}

The argument :attr:\code{diagonal} controls which diagonal to consider:
\itemize{
\item If :attr:\code{diagonal} = 0, it is the main diagonal.
\item If :attr:\code{diagonal} > 0, it is above the main diagonal.
\item If :attr:\code{diagonal} < 0, it is below the main diagonal.
}
}
