% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_diag_embed}
\alias{torch_diag_embed}
\title{Diag_embed}
\arguments{
\item{input}{(Tensor) the input tensor. Must be at least 1-dimensional.}

\item{offset}{(int, optional) which diagonal to consider. Default: 0        (main diagonal).}

\item{dim1}{(int, optional) first dimension with respect to which to        take diagonal. Default: -2.}

\item{dim2}{(int, optional) second dimension with respect to which to        take diagonal. Default: -1.}
}
\description{
diag_embed(input, offset=0, dim1=-2, dim2=-1) -> Tensor
}
\details{
Creates a tensor whose diagonals of certain 2D planes (specified by
:attr:\code{dim1} and :attr:\code{dim2}) are filled by :attr:\code{input}.
To facilitate creating batched diagonal matrices, the 2D planes formed by
the last two dimensions of the returned tensor are chosen by default.

The argument :attr:\code{offset} controls which diagonal to consider:
\itemize{
\item If :attr:\code{offset} = 0, it is the main diagonal.
\item If :attr:\code{offset} > 0, it is above the main diagonal.
\item If :attr:\code{offset} < 0, it is below the main diagonal.
}

The size of the new matrix will be calculated to make the specified diagonal
of the size of the last input dimension.
Note that for :attr:\code{offset} other than :math:\code{0}, the order of :attr:\code{dim1}
and :attr:\code{dim2} matters. Exchanging them is equivalent to changing the
sign of :attr:\code{offset}.

Applying :meth:\code{torch.diagonal} to the output of this function with
the same arguments yields a matrix identical to input. However,
:meth:\code{torch.diagonal} has different default dimensions, so those
need to be explicitly specified.
}
