% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_diagflat}
\alias{torch_diagflat}
\title{Diagflat}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{offset}{(int, optional) the diagonal to consider. Default: 0 (main        diagonal).}
}
\description{
diagflat(input, offset=0) -> Tensor
}
\details{
\itemize{
\item If :attr:\code{input} is a vector (1-D tensor), then returns a 2-D square tensor
with the elements of :attr:\code{input} as the diagonal.
\item If :attr:\code{input} is a tensor with more than one dimension, then returns a
2-D tensor with diagonal elements equal to a flattened :attr:\code{input}.
}

The argument :attr:\code{offset} controls which diagonal to consider:
\itemize{
\item If :attr:\code{offset} = 0, it is the main diagonal.
\item If :attr:\code{offset} > 0, it is above the main diagonal.
\item If :attr:\code{offset} < 0, it is below the main diagonal.
}
}
