% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_digamma}
\alias{torch_digamma}
\title{Digamma}
\usage{
torch_digamma(self)
}
\arguments{
\item{self}{(Tensor) the tensor to compute the digamma function on}
}
\description{
Digamma
}
\section{digamma(input, out=None) -> Tensor }{


Computes the logarithmic derivative of the gamma function on \code{input}.

\deqn{
    \psi(x) = \frac{d}{dx} \ln\left(\Gamma\left(x\right)\right) = \frac{\Gamma'(x)}{\Gamma(x)}
}
}

\examples{
if (torch_is_installed()) {

a = torch_tensor(c(1, 0.5))
torch_digamma(a)
}
}
