% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_divide}
\alias{torch_divide}
\title{Divide}
\usage{
torch_divide(self, other, rounding_mode)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{other}{(Number) the number to be divided to each element of \code{input}}

\item{rounding_mode}{(str, optional) – Type of rounding applied to the result:
\itemize{
\item \code{NULL} - default behavior. Performs no rounding and, if both input and
other are integer types, promotes the inputs to the default scalar type.
Equivalent to true division in Python (the / operator) and NumPy’s
\code{np.true_divide}.
\item "trunc" - rounds the results of the division towards zero. Equivalent to
C-style integer division.
\item "floor" - rounds the results of the division down. Equivalent to floor
division in Python (the // operator) and NumPy’s \code{np.floor_divide}.
}}
}
\description{
Divide
}
\section{divide(input, other, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_div]{torch_div()}}.
}

