% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_dot}
\alias{torch_dot}
\title{Dot}
\description{
dot(input, tensor) -> Tensor
}
\details{
Computes the dot product (inner product) of two tensors.

.. note:: This function does not :ref:\verb{broadcast <broadcasting-semantics>}.

Example::
}
