% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtype.R
\name{torch_dtype}
\alias{torch_dtype}
\alias{torch_float32}
\alias{torch_float}
\alias{torch_float64}
\alias{torch_double}
\alias{torch_cfloat32}
\alias{torch_chalf}
\alias{torch_cfloat}
\alias{torch_cfloat64}
\alias{torch_cdouble}
\alias{torch_cfloat128}
\alias{torch_float16}
\alias{torch_half}
\alias{torch_bfloat16}
\alias{torch_float8_e4m3fn}
\alias{torch_float8_e5m2}
\alias{torch_uint8}
\alias{torch_int8}
\alias{torch_int16}
\alias{torch_short}
\alias{torch_int32}
\alias{torch_int}
\alias{torch_int64}
\alias{torch_long}
\alias{torch_bool}
\alias{torch_quint8}
\alias{torch_qint8}
\alias{torch_qint32}
\title{Torch data types}
\usage{
torch_float32()

torch_float()

torch_float64()

torch_double()

torch_cfloat32()

torch_chalf()

torch_cfloat()

torch_cfloat64()

torch_cdouble()

torch_cfloat128()

torch_float16()

torch_half()

torch_bfloat16()

torch_float8_e4m3fn()

torch_float8_e5m2()

torch_uint8()

torch_int8()

torch_int16()

torch_short()

torch_int32()

torch_int()

torch_int64()

torch_long()

torch_bool()

torch_quint8()

torch_qint8()

torch_qint32()
}
\description{
Returns the correspondent data type.
}
\concept{tensor-attributes}
