% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_einsum}
\alias{torch_einsum}
\title{Einsum}
\arguments{
\item{equation}{(string) The equation is given in terms of lower case letters (indices) to be associated}

\item{with}{NA}

\item{dimensions, }{NA}

\item{sorted}{NA}

\item{The}{NA}

\item{entries.}{NA}

\item{If}{NA}

\item{Ellipses}{NA}

\item{the}{NA}

\item{operands}{(Tensor) The operands to compute the Einstein sum of.}
}
\description{
einsum(equation, *operands) -> Tensor
}
\details{
This function provides a way of computing multilinear expressions (i.e. sums of products) using the
Einstein summation convention.
}
