% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_empty_strided}
\alias{torch_empty_strided}
\title{Empty_strided}
\arguments{
\item{size}{(tuple of ints) the shape of the output tensor}

\item{stride}{(tuple of ints) the strides of the output tensor}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{layout}{NA class:\code{torch.layout}, optional): the desired layout of returned Tensor.}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if \code{None}, uses the current device for the default tensor type}

\item{NA}{func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU}

\item{for}{NA}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}

\item{pin_memory}{(bool, optional) If set, returned tensor would be allocated in}

\item{the}{NA \code{False}.}
}
\description{
empty_strided(size, stride, dtype=None, layout=None, device=None, requires_grad=False, pin_memory=False) -> Tensor
}
\details{
Returns a tensor filled with uninitialized data. The shape and strides of the tensor is
defined by the variable argument :attr:\code{size} and :attr:\code{stride} respectively.
\code{torch.empty_strided(size, stride)} is equivalent to
\verb{torch.empty(size).as_strided(size, stride)}.

.. warning::
More than one element of the created tensor may refer to a single memory
location. As a result, in-place operations (especially ones that are
vectorized) may result in incorrect behavior. If you need to write to
the tensors, please clone them first.
}
