% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_eq}
\alias{torch_eq}
\title{Eq}
\arguments{
\item{input}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}

\item{out}{(Tensor, optional) the output tensor. Must be a \code{ByteTensor}}
}
\description{
Computes element-wise equality
}
\details{
The second argument can be a number or a tensor whose shape is
:ref:\verb{broadcastable <broadcasting-semantics>} with the first argument.
}
\section{Signatures}{


eq(input, other, out=None) -> Tensor
}

\examples{
\dontrun{
torch_eq(torch_tensor([[1, 2], [3, 4]]), torch_tensor([[1, 1], [4, 4]]))
}

}
