% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_erf}
\alias{torch_erf}
\title{Erf}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Erf
}
\section{erf(input, out=None) -> Tensor }{


Computes the error function of each element. The error function is defined as follows:

\deqn{
    \mathrm{erf}(x) = \frac{2}{\sqrt{\pi}} \int_{0}^{x} e^{-t^2} dt
}
}

\examples{

\dontrun{
torch_erf(torch_tensor(c(0, -1., 10.)))
}

}
