% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_erfinv}
\alias{torch_erfinv}
\title{Erfinv}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
erfinv(input, out=None) -> Tensor
}
\details{
Computes the inverse error function of each element of :attr:\code{input}.
The inverse error function is defined in the range :math:\verb{(-1, 1)} as:

.. math::
\mathrm{erfinv}(\mathrm{erf}(x)) = x
}
\examples{
\dontrun{
torch_erfinv(torch_tensor([0, 0.5, -1.]))
}

}
