% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_exp}
\alias{torch_exp}
\title{Exp}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the exponential of the elements
of the input tensor \code{input}.
}
\details{
\deqn{
    y_{i} = e^{x_{i}}
}
}
\section{Signatures}{


exp(input, out=None) -> Tensor
}

\examples{
\dontrun{
torch_exp(torch_tensor([0, math.log(2.)]))
}

}
