% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_expm1}
\alias{torch_expm1}
\title{Expm1}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the exponential of the elements minus 1
of \code{input}.
}
\details{
\deqn{
    y_{i} = e^{x_{i}} - 1
}
}
\section{Signatures}{


expm1(input, out=None) -> Tensor
}

\examples{
\dontrun{
torch_expm1list(torch_tensor([0, math$log(2.)]))
}

}
