% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_eye}
\alias{torch_eye}
\title{Eye}
\arguments{
\item{n}{(int) the number of rows}

\item{m}{(int, optional) the number of columns with default being :attr:\code{n}}

\item{out}{(Tensor, optional) the output tensor.}

\item{dtype}{(:class:\code{torch.dtype}, optional) the desired data type of returned tensor.        Default: if \code{None}, uses a global default (see :func:\code{torch.set_default_tensor_type}).}

\item{layout}{(:class:\code{torch.layout}, optional) the desired layout of returned Tensor.        Default: \code{torch.strided}.}

\item{device}{(:class:\code{torch.device}, optional) the desired device of returned tensor.        Default: if \code{None}, uses the current device for the default tensor type        (see :func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU        for CPU tensor types and the current CUDA device for CUDA tensor types.}

\item{requires_grad}{(bool, optional) If autograd should record operations on the        returned tensor. Default: \code{False}.}
}
\description{
eye(n, m=None, out=None, dtype=None, layout=torch.strided, device=None, requires_grad=False) -> Tensor
}
\details{
Returns a 2-D tensor with ones on the diagonal and zeros elsewhere.
}
