% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-info.R
\name{torch_finfo}
\alias{torch_finfo}
\title{Floating point type info}
\usage{
torch_finfo(dtype)
}
\arguments{
\item{dtype}{dtype to check information}
}
\description{
A list that represents the numerical properties of a
floating point torch.dtype
}
