% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_fix}
\alias{torch_fix}
\title{Fix}
\usage{
torch_fix(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Fix
}
\section{fix(input, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_trunc]{torch_trunc()}}
}

