% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_fliplr}
\alias{torch_fliplr}
\title{Fliplr}
\usage{
torch_fliplr(self)
}
\arguments{
\item{self}{(Tensor) Must be at least 2-dimensional.}
}
\description{
Fliplr
}
\note{
Equivalent to \code{input[,-1]}. Requires the array to be at least 2-D.
}
\section{fliplr(input) -> Tensor }{


Flip array in the left/right direction, returning a new tensor.

Flip the entries in each row in the left/right direction.
Columns are preserved, but appear in a different order than before.
}

