% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_floor}
\alias{torch_floor}
\title{Floor}
\usage{
torch_floor(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Floor
}
\section{floor(input, out=NULL) -> Tensor }{


Returns a new tensor with the floor of the elements of \code{input},
the largest integer less than or equal to each element.

\deqn{
    \mbox{out}_{i} = \left\lfloor \mbox{input}_{i} \right\rfloor
}
}

\examples{

a = torch_randn(c(4))
a
torch_floor(a)
}
