% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_frac}
\alias{torch_frac}
\title{Frac}
\description{
Computes the fractional portion of each element in \code{input}.
}
\details{
\deqn{
    \text{out}_{i} = \text{input}_{i} - \left\lfloor |\text{input}_{i}| \right\rfloor * \operatorname{sgn}(\text{input}_{i})
}
Example::
}
\section{Signatures}{


frac(input, out=None) -> Tensor
}

\examples{
\dontrun{
torch_frac(torch_tensor([1, 2.5, -3.2]))
}

}
