% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{torch_generator}
\alias{torch_generator}
\title{Create a Generator object}
\usage{
torch_generator()
}
\description{
A \code{torch_generator}  is an object which manages the state of the algorithm
that produces pseudo random numbers. Used as a keyword argument in many
In-place random sampling functions.
}
\examples{
if (torch_is_installed()) {

# Via string
generator <- torch_generator()
generator$current_seed()
generator$set_current_seed(1234567L)
generator$current_seed()

}
}
