% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_ger}
\alias{torch_ger}
\title{Ger}
\arguments{
\item{input}{(Tensor) 1-D input vector}

\item{vec2}{(Tensor) 1-D input vector}

\item{out}{(Tensor, optional) optional output matrix}
}
\description{
ger(input, vec2, out=None) -> Tensor
}
\details{
Outer product of :attr:\code{input} and :attr:\code{vec2}.
If :attr:\code{input} is a vector of size :math:\code{n} and :attr:\code{vec2} is a vector of
size :math:\code{m}, then :attr:\code{out} must be a matrix of size :math:\verb{(n \\times m)}.

.. note:: This function does not :ref:\verb{broadcast <broadcasting-semantics>}.
}
