% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_greater_equal}
\alias{torch_greater_equal}
\title{Greater_equal}
\usage{
torch_greater_equal(self, other)
}
\arguments{
\item{self}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}
}
\description{
Greater_equal
}
\section{greater_equal(input, other, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_ge]{torch_ge()}}.
}

