% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_gt}
\alias{torch_gt}
\title{Gt}
\arguments{
\item{input}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}

\item{out}{(Tensor, optional) the output tensor that must be a \code{BoolTensor}}
}
\description{
Gt
}
\section{gt(input, other, out=None) -> Tensor }{


Computes \eqn{\text{input} > \text{other}} element-wise.

The second argument can be a number or a tensor whose shape is
broadcastable  with the first argument.
}

\examples{

torch_gt(torch_tensor(c([1, 2], [3, 4])), torch_tensor(c([1, 1], [4, 4])))
}
