% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_hamming_window}
\alias{torch_hamming_window}
\title{Hamming_window}
\arguments{
\item{window_length}{(int) the size of returned window}

\item{periodic}{(bool, optional) If True, returns a window to be used as periodic}

\item{function.}{NA}

\item{alpha}{(float, optional) The coefficient :math:\verb{\\alpha} in the equation above}

\item{beta}{(float, optional) The coefficient :math:\verb{\\beta} in the equation above}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{layout}{NA class:\code{torch.layout}, optional): the desired layout of returned window tensor. Only}

\item{``torch.strided``}{(dense layout)}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if \code{None}, uses the current device for the default tensor type}

\item{NA}{func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU}

\item{for}{NA}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}
}
\description{
hamming_window(window_length, periodic=True, alpha=0.54, beta=0.46, dtype=None, layout=torch.strided, device=None, requires_grad=False) -> Tensor
}
\details{
Hamming window function.

.. math::
w\link{n} = \alpha - \beta\ \cos \left( \frac{2 \pi n}{N - 1} \right),
}
