% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_hypot}
\alias{torch_hypot}
\title{Hypot}
\usage{
torch_hypot(self, other)
}
\arguments{
\item{other}{(Tensor) the second input tensor}

\item{input}{(Tensor) the first input tensor}
}
\description{
Hypot
}
\section{hypot(input, other, *, out=None) -> Tensor }{


Given the legs of a right triangle, return its hypotenuse.

\deqn{
\mbox{out}_{i} = \sqrt{\mbox{input}_{i}^{2} + \mbox{other}_{i}^{2}}
}

The shapes of \code{input} and \code{other} must be
broadcastable .
}

