% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_i0}
\alias{torch_i0}
\title{I0}
\usage{
torch_i0(self)
}
\arguments{
\item{self}{(Tensor) the input tensor}
}
\description{
I0
}
\section{i0(input, *, out=None) -> Tensor }{


Computes the zeroth order modified Bessel function of the first kind for each element of \code{input}.

\deqn{
\mbox{out}_{i} = I_0(\mbox{input}_{i}) = \sum_{k=0}^{\infty} \frac{(\mbox{input}_{i}^2/4)^k}{(k!)^2}
}
}

\examples{

torch_i0(torch_arange(start = 0, end = 5, dtype = torch_float32()))
}
