% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-info.R
\name{torch_iinfo}
\alias{torch_iinfo}
\title{Integer type info}
\usage{
torch_iinfo(dtype)
}
\arguments{
\item{dtype}{dtype to get information from.}
}
\description{
A list that represents the numerical properties of a integer
type.
}
\concept{tensor-attributes}
