% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_imag}
\alias{torch_imag}
\title{Imag}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Imag
}
\section{imag(input, out=None) -> Tensor }{


Returns the imaginary part of the \code{input} tensor.
}

\section{Warning}{

Not yet implemented.

\deqn{
    \mbox{out}_{i} = imag(\mbox{input}_{i})
}
}

\examples{
\dontrun{
torch_imag(torch_tensor(c(-1 + 1i, -2 + 2i, 3 - 3i)))
}
}
