% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace.R, R/wrapers.R
\name{torch_index}
\alias{torch_index}
\title{Index torch tensors}
\usage{
torch_index(self, indices)
}
\arguments{
\item{self}{(Tensor) Tensor that will be indexed.}

\item{indices}{(\code{List[Tensor]}) List of indices. Indices are torch tensors with
\code{torch_long()} dtype.}
}
\description{
Helper functions to index tensors.
}
