% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace.R, R/wrapers.R
\name{torch_index_put_}
\alias{torch_index_put_}
\title{In-place version of \code{torch_index_put}.}
\usage{
torch_index_put_(self, indices, values, accumulate = FALSE)
}
\arguments{
\item{self}{(Tensor) Tensor that will be indexed.}

\item{indices}{(\code{List[Tensor]}) List of indices. Indices are torch tensors with
\code{torch_long()} dtype.}

\item{values}{(Tensor) values that will be replaced the indexed location. Used
for \code{torch_index_put} and \code{torch_index_put_}.}

\item{accumulate}{(bool) Wether instead of replacing the current values with \code{values},
you want to add them.}
}
\description{
In-place version of \code{torch_index_put}.
}
