% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_inverse}
\alias{torch_inverse}
\title{Inverse}
\arguments{
\item{input}{(Tensor) the input tensor of size :math:\verb{(*, n, n)} where \code{*} is zero or more                    batch dimensions}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
inverse(input, out=None) -> Tensor
}
\details{
Takes the inverse of the square matrix :attr:\code{input}. :attr:\code{input} can be batches
of 2D square tensors, in which case this function would return a tensor composed of
individual inverses.

.. note::\preformatted{Irrespective of the original strides, the returned tensors will be
transposed, i.e. with strides like `input.contiguous().transpose(-2, -1).stride()`
}
}
