% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_is_floating_point}
\alias{torch_is_floating_point}
\title{Is_floating_point}
\arguments{
\item{input}{(Tensor) the PyTorch tensor to test}
}
\description{
Is_floating_point
}
\section{is_floating_point(input) -> (bool) }{


Returns True if the data type of \code{input} is a floating point data type i.e.,
one of \code{torch_float64}, \code{torch.float32} and \code{torch.float16}.
}

\examples{


}
