% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_is_nonzero}
\alias{torch_is_nonzero}
\title{Is_nonzero}
\usage{
torch_is_nonzero(self)
}
\arguments{
\item{input}{(Tensor) the input tensor.}
}
\description{
Is_nonzero
}
\section{is_nonzero(input) -> (bool) }{


Returns TRUE if the \code{input} is a single element tensor which is not equal to zero
after type conversions.
i.e. not equal to \code{torch_tensor(c(0.))} or \code{torch.tensor(c(0))} or
\code{torch_tensor(c(FALSE))}.
Throws a \code{RuntimeError} if \code{torch_numel() != 1} (even in case
of sparse tensors).
}

