% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_isinf}
\alias{torch_isinf}
\title{Isinf}
\arguments{
\item{tensor}{(Tensor) A tensor to check}
}
\description{
Isinf
}
\section{TEST }{


Returns a new tensor with boolean elements representing if each element is \verb{+/-INF} or not.
}

\examples{
\dontrun{

torch_isinf(torch_tensor(c(1, Inf, 2, -Inf, NaN)))
}
}
