% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_isnan}
\alias{torch_isnan}
\title{Isnan}
\arguments{
\item{input}{(Tensor) A tensor to check}
}
\description{
Returns a new tensor with boolean elements representing if each element is \code{NaN} or not.
}
\details{
Arguments:
input (Tensor): A tensor to check

Returns:
Tensor: A \code{torch.BoolTensor} containing a True at each location of \code{NaN} elements.

Example::
}
