% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_isneginf}
\alias{torch_isneginf}
\title{Isneginf}
\usage{
torch_isneginf(self)
}
\arguments{
\item{input}{(Tensor) the input tensor.}
}
\description{
Isneginf
}
\section{isneginf(input, *, out=None) -> Tensor }{


Tests if each element of \code{input} is negative infinity or not.
}

