% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_isposinf}
\alias{torch_isposinf}
\title{Isposinf}
\usage{
torch_isposinf(self)
}
\arguments{
\item{input}{(Tensor) the input tensor.}
}
\description{
Isposinf
}
\section{isposinf(input, *, out=None) -> Tensor }{


Tests if each element of \code{input} is positive infinity or not.
}

