% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_isreal}
\alias{torch_isreal}
\title{Isreal}
\usage{
torch_isreal(self)
}
\arguments{
\item{input}{(Tensor) the input tensor.}
}
\description{
Isreal
}
\section{isreal(input) -> Tensor }{


Returns a new tensor with boolean elements representing if each element of \code{input} is real-valued or not.
All real-valued types are considered real. Complex values are considered real when their imaginary part is 0.
}

