% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_kaiser_window}
\alias{torch_kaiser_window}
\title{Kaiser_window}
\usage{
torch_kaiser_window(window_length, periodic, beta, options = list())
}
\arguments{
\item{window_length}{(int) length of the window.}

\item{periodic}{(bool, optional) If TRUE, returns a periodic window suitable for use in spectral analysis.        If FALSE, returns a symmetric window suitable for use in filter design.}

\item{beta}{(float, optional) shape parameter for the window.}
}
\description{
Kaiser_window
}
\note{
If \code{window_length} is one, then the returned window is a single element
tensor containing a one.
}
\section{kaiser_window(window_length, periodic=TRUE, beta=12.0, *, dtype=None, layout=torch.strided, device=None, requires_grad=FALSE) -> Tensor }{


Computes the Kaiser window with window length \code{window_length} and shape parameter \code{beta}.

Let I_0 be the zeroth order modified Bessel function of the first kind (see \code{\link[=torch_i0]{torch_i0()}}) and
\code{N = L - 1} if \code{periodic} is FALSE and \code{L} if \code{periodic} is TRUE,
where \code{L} is the \code{window_length}. This function computes:

\deqn{
    out_i = I_0 \left( \beta \sqrt{1 - \left( {\frac{i - N/2}{N/2}} \right) ^2 } \right) / I_0( \beta )
}

Calling \code{torch_kaiser_window(L, B, periodic=TRUE)} is equivalent to calling
\verb{torch_kaiser_window(L + 1, B, periodic=FALSE)[:-1])}.
The \code{periodic} argument is intended as a helpful shorthand
to produce a periodic window as input to functions like \code{\link[=torch_stft]{torch_stft()}}.
}

