% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_kron}
\alias{torch_kron}
\title{Kronecker product}
\usage{
torch_kron(self, other)
}
\arguments{
\item{self}{(\code{Tensor}) input Tensor}

\item{other}{(\code{Tensor}) other tensor.}
}
\description{
Computes the Kronecker product of \code{self} and \code{other}.
}
