% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{torch_layout}
\alias{torch_layout}
\alias{torch_strided}
\alias{torch_sparse_coo}
\title{Creates the corresponding layout}
\usage{
torch_strided()

torch_sparse_coo()
}
\description{
Creates the corresponding layout
}
