% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_ldexp}
\alias{torch_ldexp}
\title{Ldexp}
\usage{
torch_ldexp(self, other)
}
\arguments{
\item{self}{(Tensor) the tensor of mantissas}

\item{other}{(Tensor) the tensor of exponents, must be an integer dtype}
}
\description{
Ldexp
}
\section{ldexp(input, other, out=NULL) -> Tensor }{


Multiplies \code{input} by \eqn{2^{other}}.

\deqn{
    \text{out}_i = \text{input}_i * 2^{\text{other}_i}
}

Typically this function is used to construct floating point numbers by multiplying
mantissas in \code{input} with integral powers of two created from the exponents in \code{other}.
}

