% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_le}
\alias{torch_le}
\title{Le}
\arguments{
\item{input}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}

\item{out}{(Tensor, optional) the output tensor that must be a \code{BoolTensor}}
}
\description{
Le
}
\section{le(input, other, out=None) -> Tensor }{


Computes \eqn{\mbox{input} \leq \mbox{other}} element-wise.

The second argument can be a number or a tensor whose shape is
broadcastable  with the first argument.
}

\examples{

torch_le(torch_tensor(matrix(1:4, ncol = 2, byrow=TRUE)), torch_tensor(matrix(c(1,1,4,4), ncol = 2, byrow=TRUE)))
}
