% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_lerp}
\alias{torch_lerp}
\title{Lerp}
\arguments{
\item{input}{(Tensor) the tensor with the starting points}

\item{end}{(Tensor) the tensor with the ending points}

\item{weight}{(float or tensor) the weight for the interpolation formula}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
lerp(input, end, weight, out=None)
}
\details{
Does a linear interpolation of two tensors \code{start} (given by \code{input}) and \code{end} based
on a scalar or tensor \code{weight} and returns the resulting \code{out} tensor.

\deqn{
    \text{out}_i = \text{start}_i + \text{weight}_i \times (\text{end}_i - \text{start}_i)
}
The shapes of \code{start} and \code{end} must be
broadcastable . If \code{weight} is a tensor, then
the shapes of \code{weight}, \code{start}, and \code{end} must be broadcastable .
}
\examples{
\dontrun{
start = torch_arange(1., 5.)
end = torch_empty(4)$fill_list(10)
start
end
torch_lerp(start, end, 0.5)
torch_lerp(start, end, torch_full_like(start, 0.5))
}

}
