% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_less}
\alias{torch_less}
\title{Less}
\usage{
torch_less(self, other)
}
\arguments{
\item{self}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}
}
\description{
Less
}
\section{less(input, other, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_lt]{torch_lt()}}.
}

