% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_less_equal}
\alias{torch_less_equal}
\title{Less_equal}
\usage{
torch_less_equal(self, other)
}
\arguments{
\item{self}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}
}
\description{
Less_equal
}
\section{less_equal(input, other, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_le]{torch_le()}}.
}

