% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_lgamma}
\alias{torch_lgamma}
\title{Lgamma}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
lgamma(input, out=None) -> Tensor
}
\details{
Computes the logarithm of the gamma function on \code{input}.

\deqn{
    \text{out}_{i} = \log \Gamma(\text{input}_{i})
}
}
\examples{
\dontrun{
a = torch_arange(0.5, 2, 0.5)
torch_lgamma(a)
}

}
