% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_linspace}
\alias{torch_linspace}
\title{Linspace}
\arguments{
\item{start}{(float) the starting value for the set of points}

\item{end}{(float) the ending value for the set of points}

\item{steps}{(int) number of points to sample between :attr:\code{start}}

\item{and}{NA attr:\code{end}. Default: \code{100}.}

\item{out}{(Tensor, optional) the output tensor.}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{layout}{NA class:\code{torch.layout}, optional): the desired layout of returned Tensor.}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if \code{None}, uses the current device for the default tensor type}

\item{NA}{func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU}

\item{for}{NA}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}
}
\description{
linspace(start, end, steps=100, out=None, dtype=None, layout=torch.strided, device=None, requires_grad=False) -> Tensor
}
\details{
Returns a one-dimensional tensor of :attr:\code{steps}
equally spaced points between :attr:\code{start} and :attr:\code{end}.

The output tensor is 1-D of size :attr:\code{steps}.
}
