% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_log}
\alias{torch_log}
\title{Log}
\usage{
torch_log(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Log
}
\section{log(input, out=NULL) -> Tensor }{


Returns a new tensor with the natural logarithm of the elements
of \code{input}.

\deqn{
    y_{i} = \log_{e} (x_{i})
}
}

\examples{

a <- torch_randn(c(5))
a
torch_log(a)
}
