% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_log2}
\alias{torch_log2}
\title{Log2}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Log2
}
\section{log2(input, out=None) -> Tensor }{


Returns a new tensor with the logarithm to the base 2 of the elements
of \code{input}.

\deqn{
    y_{i} = \log_{2} (x_{i})
}
}

\examples{

\dontrun{
a = torch_rand(5)
a
torch_log2list(a)
}

}
