% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_logaddexp2}
\alias{torch_logaddexp2}
\title{Logaddexp2}
\usage{
torch_logaddexp2(self, other)
}
\arguments{
\item{other}{(Tensor) the second input tensor}

\item{input}{(Tensor) the input tensor.}
}
\description{
Logaddexp2
}
\section{logaddexp2(input, other, *, out=None) -> Tensor }{


Logarithm of the sum of exponentiations of the inputs in base-2.

Calculates pointwise \eqn{\log_2\left(2^x + 2^y\right)}. See
\code{\link[=torch_logaddexp]{torch_logaddexp()}} for more details.
}

