% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_logdet}
\alias{torch_logdet}
\title{Logdet}
\arguments{
\item{input}{(Tensor) the input tensor of size \verb{(*, n, n)} where \code{*} is zero or more                batch dimensions.}
}
\description{
Calculates log determinant of a square matrix or batches of square matrices.
}
\details{
.. note::
Result is \code{-inf} if \code{input} has zero log determinant, and is \code{nan} if
\code{input} has negative determinant.

.. note::
Backward through :meth:\code{logdet} internally uses SVD results when \code{input}
is not invertible. In this case, double backward through :meth:\code{logdet} will
be unstable in when \code{input} doesn't have distinct singular values. See
:meth:\code{~torch.svd} for details.
}
\section{Signatures}{


logdet(input) -> Tensor
}

\examples{
\dontrun{
A = torch_randn(3, 3)
torch_det(A)
torch_logdet(A)
A
A.det()
A.det().log()
}

}
