% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_logical_xor}
\alias{torch_logical_xor}
\title{Logical_xor}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{other}{(Tensor) the tensor to compute XOR with}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
logical_xor(input, other, out=None) -> Tensor
}
\details{
Computes the element-wise logical XOR of the given input tensors. Zeros are treated as \code{False} and nonzeros are
treated as \code{True}.
}
