% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_lstsq}
\alias{torch_lstsq}
\title{Lstsq}
\arguments{
\item{input}{(Tensor) the matrix :math:\code{B}}

\item{A}{(Tensor) the :math:\code{m} by :math:\code{n} matrix :math:\code{A}}

\item{out}{(tuple, optional) the optional destination tensor}
}
\description{
Computes the solution to the least squares and least norm problems for a full
rank matrix :math:\code{A} of size :math:\verb{(m \\times n)} and a matrix :math:\code{B} of
size :math:\verb{(m \\times k)}.
}
\details{
If :math:\verb{m \\geq n}, \code{\link{lstsq}} solves the least-squares problem:

\deqn{
}
\begin{array}{ll}
\min_X & \|AX-B\|_2.
\end{array}

If :math:\code{m < n}, \code{\link{lstsq}} solves the least-norm problem:

\deqn{
}
\begin{array}{ll}
\min_X & \|X\|_2 & \text{subject to} & AX = B.
\end{array}

Returned tensor :math:\code{X} has shape :math:\verb{(\\max(m, n) \\times k)}. The first :math:\code{n}
rows of :math:\code{X} contains the solution. If :math:\verb{m \\geq n}, the residual sum of squares
for the solution in each column is given by the sum of squares of elements in the
remaining :math:\code{m - n} rows of that column.
}
\note{
\preformatted{The case when :math:`m < n` is not supported on the GPU.
}
}
\section{Signatures}{


lstsq(input, A, out=None) -> Tensor
}

\examples{
\dontrun{
A = torch_tensor([[1., 1, 1],
B = torch_tensor([[-10., -3],
X, _ = torch_lstsq(B, A)
X
}

}
