% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_lu_solve}
\alias{torch_lu_solve}
\title{Lu_solve}
\arguments{
\item{b}{(Tensor) the RHS tensor of size :math:\verb{(*, m, k)}, where :math:\code{*}                is zero or more batch dimensions.}

\item{LU_data}{(Tensor) the pivoted LU factorization of A from :meth:\code{torch.lu} of size :math:\verb{(*, m, m)},                       where :math:\code{*} is zero or more batch dimensions.}

\item{LU_pivots}{(IntTensor) the pivots of the LU factorization from :meth:\code{torch.lu} of size :math:\verb{(*, m)},                           where :math:\code{*} is zero or more batch dimensions.                           The batch dimensions of :attr:\code{LU_pivots} must be equal to the batch dimensions of                           :attr:\code{LU_data}.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns the LU solve of the linear system \eqn{Ax = b} using the partially pivoted
LU factorization of A from :meth:\code{torch_lu}.
}
\section{Signatures}{


lu_solve(input, LU_data, LU_pivots, out=None) -> Tensor
}

\examples{
\dontrun{
A = torch_randn(2, 3, 3)
b = torch_randn(2, 3, 1)
A_LU = torch_lu(A)
x = torch_lu_solve(b, *A_LU)
torch_norm(torch_bmm(A, x) - b)
}

}
