% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_maximum}
\alias{torch_maximum}
\title{Maximum}
\usage{
torch_maximum(self, other)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{other}{(Tensor) the second input tensor}
}
\description{
Maximum
}
\note{
If one of the elements being compared is a NaN, then that element is returned.
\code{\link[=torch_maximum]{torch_maximum()}} is not supported for tensors with complex dtypes.
}
\section{maximum(input, other, *, out=None) -> Tensor }{


Computes the element-wise maximum of \code{input} and \code{other}.
}

\examples{
if (torch_is_installed()) {

a <- torch_tensor(c(1, 2, -1))
b <- torch_tensor(c(3, 0, 4))
torch_maximum(a, b)
}
}
