% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_mean}
\alias{torch_mean}
\title{Mean}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int or tuple of ints) the dimension or dimensions to reduce.}

\item{keepdim}{(bool) whether the output tensor has :attr:\code{dim} retained or not.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns the mean value of all elements in the \code{input} tensor.

Returns the mean value of each row of the \code{input} tensor in the given
dimension \code{dim}. If \code{dim} is a list of dimensions,
reduce over all of them.
}
\details{
If \code{keepdim} is \code{True}, the output tensor is of the same size
as \code{input} except in the dimension(s) \code{dim} where it is of size 1.
Otherwise, \code{dim} is squeezed (see :func:\code{torch_squeeze}), resulting in the
output tensor having 1 (or \code{len(dim)}) fewer dimension(s).
}
\section{Signatures}{


mean(input) -> Tensor



mean(input, dim, keepdim=False, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(1, 3)
a
torch_mean(a)
}

\dontrun{
a = torch_randn(4, 4)
a
torch_mean(a, 1)
torch_mean(a, 1, True)
}

}
