% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_format.R
\name{torch_memory_format}
\alias{torch_memory_format}
\alias{torch_contiguous_format}
\alias{torch_preserve_format}
\alias{torch_channels_last_format}
\title{Memory format}
\usage{
torch_contiguous_format()

torch_preserve_format()

torch_channels_last_format()
}
\description{
Returns the correspondent memory format.
}
