% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_minimum}
\alias{torch_minimum}
\title{Minimum}
\usage{
torch_minimum(self, other)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{other}{(Tensor) the second input tensor}
}
\description{
Minimum
}
\note{
If one of the elements being compared is a NaN, then that element is returned.
\code{\link[=torch_minimum]{torch_minimum()}} is not supported for tensors with complex dtypes.
}
\section{minimum(input, other, *, out=None) -> Tensor }{


Computes the element-wise minimum of \code{input} and \code{other}.
}

