% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_mm}
\alias{torch_mm}
\title{Mm}
\arguments{
\item{input}{(Tensor) the first matrix to be multiplied}

\item{mat2}{(Tensor) the second matrix to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
mm(input, mat2, out=None) -> Tensor
}
\details{
Performs a matrix multiplication of the matrices \code{input} and \code{mat2}.

If \code{input} is a :math:\verb{(n \\times m)} tensor, \code{mat2} is a
:math:\verb{(m \\times p)} tensor, \code{out} will be a :math:\verb{(n \\times p)} tensor.

.. note:: This function does not :ref:\verb{broadcast <broadcasting-semantics>}.
For broadcasting matrix products, see \code{\link{torch_matmul}}.
}
\examples{
\dontrun{
mat1 = torch_randn(2, 3)
mat2 = torch_randn(3, 3)
torch_mm(mat1, mat2)
}

}
