% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_mode}
\alias{torch_mode}
\title{Mode}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int) the dimension to reduce.}

\item{keepdim}{(bool) whether the output tensor has :attr:\code{dim} retained or not.}

\item{values}{(Tensor, optional) the output tensor}

\item{indices}{(Tensor, optional) the output index tensor}
}
\description{
mode(input, dim=-1, keepdim=False, values=None, indices=None) -> (Tensor, LongTensor)
}
\details{
Returns a namedtuple \verb{(values, indices)} where \code{values} is the mode
value of each row of the :attr:\code{input} tensor in the given dimension
:attr:\code{dim}, i.e. a value which appears most often
in that row, and \code{indices} is the index location of each mode value found.

By default, :attr:\code{dim} is the last dimension of the :attr:\code{input} tensor.

If :attr:\code{keepdim} is \code{True}, the output tensors are of the same size as
:attr:\code{input} except in the dimension :attr:\code{dim} where they are of size 1.
Otherwise, :attr:\code{dim} is squeezed (see :func:\code{torch.squeeze}), resulting
in the output tensors having 1 fewer dimension than :attr:\code{input}.

.. note:: This function is not defined for \code{torch.cuda.Tensor} yet.
}
