% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_mul}
\alias{torch_mul}
\title{Mul}
\arguments{
\item{{input}}{NA}

\item{value}{(Number) the number to be multiplied to each element of \code{input}}

\item{{out}}{NA}

\item{input}{(Tensor) the first multiplicand tensor}

\item{other}{(Tensor) the second multiplicand tensor}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Mul
}
\section{mul(input, other, out=None) }{


Multiplies each element of the input \code{input} with the scalar
\code{other} and returns a new resulting tensor.

\deqn{
    \text{out}_i = \text{other} \times \text{input}_i
}
If \code{input} is of type \code{FloatTensor} or \code{DoubleTensor}, \code{other}
should be a real number, otherwise it should be an integer



Each element of the tensor \code{input} is multiplied by the corresponding
element of the Tensor \code{other}. The resulting tensor is returned.

The shapes of \code{input} and \code{other} must be
broadcastable .

\deqn{
    \text{out}_i = \text{input}_i \times \text{other}_i
}
}

\examples{

a = torch_randn(c(3))
a
torch_mul(a, 100)


a = torch_randn(c(4, 1))
a
b = torch_randn(c(1, 4))
b
torch_mul(a, b)
}
