% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-new-examples.R, R/gen-namespace.R
\name{torch_multiply}
\alias{torch_multiply}
\title{Multiply}
\usage{
torch_multiply(self, other)
}
\arguments{
\item{self}{(Tensor) the first multiplicand tensor}

\item{other}{(Tensor) the second multiplicand tensor}
}
\description{
Multiply
}
\section{multiply(input, other, *, out=None) }{


Alias for \code{\link[=torch_mul]{torch_mul()}}.
}

