% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_mv}
\alias{torch_mv}
\title{Mv}
\arguments{
\item{input}{(Tensor) matrix to be multiplied}

\item{vec}{(Tensor) vector to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Mv
}
\note{
This function does not broadcast .
}
\section{mv(input, vec, out=None) -> Tensor }{


Performs a matrix-vector product of the matrix \code{input} and the vector
\code{vec}.

If \code{input} is a \eqn{(n \times m)} tensor, \code{vec} is a 1-D tensor of
size \eqn{m}, \code{out} will be 1-D of size \eqn{n}.
}

\examples{

\dontrun{
mat = torch_randn(c(2, 3))
vec = torch_randn(c(3))
torch_mv(mat, vec)
}

}
