% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_mvlgamma}
\alias{torch_mvlgamma}
\title{Mvlgamma}
\arguments{
\item{input}{(Tensor) the tensor to compute the multivariate log-gamma function}

\item{p}{(int) the number of dimensions}
}
\description{
Mvlgamma
}
\section{mvlgamma(input, p) -> Tensor }{


Computes the multivariate log-gamma function (\verb{[reference]}_) with dimension \eqn{p} element-wise, given by

\deqn{
    \log(\Gamma_{p}(a)) = C + \displaystyle \sum_{i=1}^{p} \log\left(\Gamma\left(a - \frac{i - 1}{2}\right)\right)
}
where \eqn{C = \log(\pi) \times \frac{p (p - 1)}{4}} and \eqn{\Gamma(\cdot)} is the Gamma function.

If any of the elements are less than or equal to \eqn{\frac{p - 1}{2}}, then an error
is thrown.
}

\examples{

a = torch_empty(c(2, 3))$uniform_(1, 2)
a
torch_mvlgamma(a, 2)
}
