% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_narrow}
\alias{torch_narrow}
\title{Narrow}
\arguments{
\item{input}{(Tensor) the tensor to narrow}

\item{dim}{(int) the dimension along which to narrow}

\item{start}{(int) the starting dimension}

\item{length}{(int) the distance to the ending dimension}
}
\description{
Returns a new tensor that is a narrowed version of \code{input} tensor. The
dimension \code{dim} is input from \code{start} to \code{start + length}. The
returned tensor and \code{input} tensor share the same underlying storage.
}
\section{Signatures}{


narrow(input, dim, start, length) -> Tensor
}

\examples{
\dontrun{
x = torch_tensor(c([1, 2, 3], [4, 5, 6], [7, 8, 9]))
torch_narrow(x, 0, 0, 2)
torch_narrow(x, 1, 1, 2)
}

}
