% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_ne}
\alias{torch_ne}
\title{Ne}
\usage{
torch_ne(self, other)
}
\arguments{
\item{self}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}
}
\description{
Ne
}
\section{ne(input, other, out=NULL) -> Tensor }{


Computes \eqn{input \neq other} element-wise.

The second argument can be a number or a tensor whose shape is
broadcastable  with the first argument.
}

\examples{

torch_ne(torch_tensor(matrix(1:4, ncol = 2, byrow=TRUE)), 
         torch_tensor(matrix(rep(c(1,4), each = 2), ncol = 2, byrow=TRUE)))
}
