% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_neg}
\alias{torch_neg}
\title{Neg}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Neg
}
\section{neg(input, out=None) -> Tensor }{


Returns a new tensor with the negative of the elements of \code{input}.

\deqn{
    \mbox{out} = -1 \times \mbox{input}
}
}

\examples{
\dontrun{

a = torch_randn(c(5))
a
torch_neg(a)
}
}
