% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_negative}
\alias{torch_negative}
\title{Negative}
\usage{
torch_negative(self)
}
\description{
Negative
}
\section{negative(input, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_neg]{torch_neg()}}
}

