% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_norm}
\alias{torch_norm}
\title{Norm}
\description{
Returns the matrix norm or vector norm of a given tensor.
}
\details{
\preformatted{Args:
    input (Tensor): the input tensor
    p (int, float, inf, -inf, 'fro', 'nuc', optional): the order of norm. Default: ``'fro'``
        The following norms can be calculated:
}
}
