% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_not_equal}
\alias{torch_not_equal}
\title{Not_equal}
\usage{
torch_not_equal(self, other)
}
\arguments{
\item{self}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}
}
\description{
Not_equal
}
\section{not_equal(input, other, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_ne]{torch_ne()}}.
}

