% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_orgqr}
\alias{torch_orgqr}
\title{Orgqr}
\arguments{
\item{input}{(Tensor) the \code{a} from :func:\code{torch.geqrf}.}

\item{input2}{(Tensor) the \code{tau} from :func:\code{torch.geqrf}.}
}
\description{
Computes the orthogonal matrix \code{Q} of a QR factorization, from the \verb{(input, input2)}
tuple returned by :func:\code{torch_geqrf}.
}
\details{
This directly calls the underlying LAPACK function \code{?orgqr}.
See \verb{LAPACK documentation for orgqr}_ for further details.
}
\section{Signatures}{


orgqr(input, input2) -> Tensor
}

