% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_ormqr}
\alias{torch_ormqr}
\title{Ormqr}
\arguments{
\item{input}{(Tensor) the \code{a} from :func:\code{torch.geqrf}.}

\item{input2}{(Tensor) the \code{tau} from :func:\code{torch.geqrf}.}

\item{input3}{(Tensor) the matrix to be multiplied.}
}
\description{
ormqr(input, input2, input3, left=True, transpose=False) -> Tensor
}
\details{
Multiplies \code{mat} (given by :attr:\code{input3}) by the orthogonal \code{Q} matrix of the QR factorization
formed by :func:\code{torch.geqrf} that is represented by \verb{(a, tau)} (given by (:attr:\code{input}, :attr:\code{input2})).

This directly calls the underlying LAPACK function \code{?ormqr}.
See \verb{LAPACK documentation for ormqr}_ for further details.
}
