% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_pinverse}
\alias{torch_pinverse}
\title{Pinverse}
\arguments{
\item{input}{(Tensor) The input tensor of size \eqn{(*, m, n)} where \eqn{*} is zero or more batch dimensions}

\item{rcond}{(float) A floating point value to determine the cutoff for small singular values.                   Default: 1e-15}
}
\description{
Calculates the pseudo-inverse (also known as the Moore-Penrose inverse) of a 2D tensor.
Please look at \verb{Moore-Penrose inverse}_ for more details
}
\note{
\preformatted{This method is implemented using the Singular Value Decomposition.
}

\preformatted{The pseudo-inverse is not necessarily a continuous function in the elements of the matrix `[1]`_.
Therefore, derivatives are not always existent, and exist for a constant rank only `[2]`_.
However, this method is backprop-able due to the implementation by using SVD results, and
could be unstable. Double-backward will also be unstable due to the usage of SVD internally.
See :meth:`~torch.svd` for more details.
}
}
\section{Signatures}{


pinverse(input, rcond=1e-15) -> Tensor
}

\examples{
\dontrun{
input = torch_randn(3, 5)
input
torch_pinverse(input)
# Batched pinverse example
a = torch_randn(2,6,3)
b = torch_pinverse(a)
torch_matmul(b, a)
}

}
