% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_polygamma}
\alias{torch_polygamma}
\title{Polygamma}
\arguments{
\item{n}{(int) the order of the polygamma function}

\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Computes the \eqn{n^{th}} derivative of the digamma function on \code{input}.
\eqn{n \geq 0} is called the order of the polygamma function.
}
\details{
\deqn{
    \psi^{(n)}(x) = \frac{d^{(n)}}{dx^{(n)}} \psi(x)
}
}
\note{
\preformatted{This function is not implemented for \eqn{n \geq 2}.
}
}
\section{Signatures}{


polygamma(n, input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_tensor([1, 0.5])
torch_polygamma(1, a)
}

}
