% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_pow}
\alias{torch_pow}
\title{Pow}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{self}{(float) the scalar base value for the power operation}

\item{exponent}{(Tensor) the exponent tensor}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
pow(input, exponent, out=None) -> Tensor

pow(self, exponent, out=None) -> Tensor
}
\details{
Takes the power of each element in :attr:\code{input} with :attr:\code{exponent} and
returns a tensor with the result.

:attr:\code{exponent} can be either a single \code{float} number or a \code{Tensor}
with the same number of elements as :attr:\code{input}.

When :attr:\code{exponent} is a scalar value, the operation applied is:

.. math::
\text{out}_i = x_i ^ \text{exponent}

When :attr:\code{exponent} is a tensor, the operation applied is:

.. math::
\text{out}_i = x_i ^ {\text{exponent}_i}

When :attr:\code{exponent} is a tensor, the shapes of :attr:\code{input}
and :attr:\code{exponent} must be :ref:\verb{broadcastable <broadcasting-semantics>}.

:attr:\code{self} is a scalar \code{float} value, and :attr:\code{exponent} is a tensor.
The returned tensor :attr:\code{out} is of the same shape as :attr:\code{exponent}

The operation applied is:

.. math::
\text{out}_i = \text{self} ^ {\text{exponent}_i}
}
