% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_promote_types}
\alias{torch_promote_types}
\title{Promote_types}
\arguments{
\item{type1}{(:class:\code{torch.dtype})}

\item{type2}{(:class:\code{torch.dtype})}
}
\description{
promote_types(type1, type2) -> dtype
}
\details{
Returns the :class:\code{torch.dtype} with the smallest size and scalar kind that is
not smaller nor of lower kind than either \code{type1} or \code{type2}. See type promotion
:ref:\verb{documentation <type-promotion-doc>} for more information on the type
promotion logic.
}
