% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qscheme.R
\name{torch_qscheme}
\alias{torch_qscheme}
\alias{torch_per_channel_affine}
\alias{torch_per_tensor_affine}
\alias{torch_per_channel_symmetric}
\alias{torch_per_tensor_symmetric}
\title{Creates the corresponding Scheme object}
\usage{
torch_per_channel_affine()

torch_per_tensor_affine()

torch_per_channel_symmetric()

torch_per_tensor_symmetric()
}
\description{
Creates the corresponding Scheme object
}
