% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_rad2deg}
\alias{torch_rad2deg}
\title{Rad2deg}
\usage{
torch_rad2deg(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Rad2deg
}
\section{rad2deg(input, *, out=None) -> Tensor }{


Returns a new tensor with each of the elements of \code{input}
converted from angles in radians to degrees.
}

\examples{

a <- torch_tensor(rbind(c(3.142, -3.142), c(6.283, -6.283), c(1.570, -1.570)))
torch_rad2deg(a)
}
