% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_randn}
\alias{torch_randn}
\title{Randn}
\arguments{
\item{size}{(int...) a sequence of integers defining the shape of the output tensor.}

\item{Can}{NA}

\item{out}{(Tensor, optional) the output tensor.}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{layout}{NA class:\code{torch.layout}, optional): the desired layout of returned Tensor.}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if \code{None}, uses the current device for the default tensor type}

\item{NA}{func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU}

\item{for}{NA}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}
}
\description{
randn(*size, out=None, dtype=None, layout=torch.strided, device=None, requires_grad=False) -> Tensor
}
\details{
Returns a tensor filled with random numbers from a normal distribution
with mean \code{0} and variance \code{1} (also called the standard normal
distribution).

.. math::
\text{out}_{i} \sim \mathcal{N}(0, 1)

The shape of the tensor is defined by the variable argument :attr:\code{size}.
}
