% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_randperm}
\alias{torch_randperm}
\title{Randperm}
\arguments{
\item{n}{(int) the upper bound (exclusive)}

\item{out}{(Tensor, optional) the output tensor.}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{layout}{NA class:\code{torch.layout}, optional): the desired layout of returned Tensor.}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if \code{None}, uses the current device for the default tensor type}

\item{NA}{func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU}

\item{for}{NA}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}
}
\description{
randperm(n, out=None, dtype=torch.int64, layout=torch.strided, device=None, requires_grad=False) -> LongTensor
}
\details{
Returns a random permutation of integers from \code{0} to \code{n - 1}.
}
