% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_range}
\alias{torch_range}
\title{Range}
\arguments{
\item{start}{(float) the starting value for the set of points. Default: \code{0}.}

\item{end}{(float) the ending value for the set of points}

\item{step}{(float) the gap between each pair of adjacent points. Default: \code{1}.}

\item{out}{(Tensor, optional) the output tensor.}

\item{dtype}{(:class:\code{torch.dtype}, optional) the desired data type of returned tensor.        Default: if \code{None}, uses a global default (see :func:\code{torch.set_default_tensor_type}). If \code{dtype} is not given, infer the data type from the other input        arguments. If any of \code{start}, \code{end}, or \code{stop} are floating-point, the        \code{dtype} is inferred to be the default dtype, see        :meth:\code{~torch.get_default_dtype}. Otherwise, the \code{dtype} is inferred to        be \code{torch.int64}.}

\item{layout}{(:class:\code{torch.layout}, optional) the desired layout of returned Tensor.        Default: \code{torch.strided}.}

\item{device}{(:class:\code{torch.device}, optional) the desired device of returned tensor.        Default: if \code{None}, uses the current device for the default tensor type        (see :func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU        for CPU tensor types and the current CUDA device for CUDA tensor types.}

\item{requires_grad}{(bool, optional) If autograd should record operations on the        returned tensor. Default: \code{False}.}
}
\description{
Returns a 1-D tensor of size :math:\verb{\\left\\lfloor \\frac\{\\text\{end\} - \\text\{start\}\}\{\\text\{step\}\} \\right\\rfloor + 1}
with values from \code{start} to \code{end} with step \code{step}. Step is
the gap between two values in the tensor.
}
\details{
\deqn{
    \text{out}_{i+1} = \text{out}_i + \text{step}.
}
.. warning::
This function is deprecated in favor of \code{\link{torch_arange}}.
}
\section{Signatures}{


range(start=0, end, step=1, out=None, dtype=None, layout=torch.strided, device=None, requires_grad=False) -> Tensor
}

\examples{
\dontrun{
torch_range(1, 4)
torch_range(1, 4, 0.5)
}

}
