% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_reciprocal}
\alias{torch_reciprocal}
\title{Reciprocal}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Reciprocal
}
\section{reciprocal(input, out=None) -> Tensor }{


Returns a new tensor with the reciprocal of the elements of \code{input}

\deqn{
    \mbox{out}_{i} = \frac{1}{\mbox{input}_{i}}
}
}

\examples{
if (torch_is_installed()) {

a = torch_randn(c(4))
a
torch_reciprocal(a)
}
}
