% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduction.R
\name{torch_reduction}
\alias{torch_reduction}
\alias{torch_reduction_sum}
\alias{torch_reduction_mean}
\alias{torch_reduction_none}
\title{Creates the reduction objet}
\usage{
torch_reduction_sum()

torch_reduction_mean()

torch_reduction_none()
}
\description{
Creates the reduction objet
}
\concept{tensor-attributes}
