% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_relu}
\alias{torch_relu}
\title{Relu}
\usage{
torch_relu(self)
}
\arguments{
\item{self}{the input tensor}
}
\description{
Relu
}
\section{relu(input) -> Tensor }{


Computes the relu tranformation.
}

