% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_remainder}
\alias{torch_remainder}
\title{Remainder}
\arguments{
\item{input}{(Tensor) the dividend}

\item{other}{(Tensor or float) the divisor that may be either a number or a                               Tensor of the same shape as the dividend}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
remainder(input, other, out=None) -> Tensor
}
\details{
Computes the element-wise remainder of division.

The divisor and dividend may contain both for integer and floating point
numbers. The remainder has the same sign as the divisor.

When :attr:\code{other} is a tensor, the shapes of :attr:\code{input} and
:attr:\code{other} must be :ref:\verb{broadcastable <broadcasting-semantics>}.
}
