% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_result_type}
\alias{torch_result_type}
\title{Result_type}
\arguments{
\item{tensor1}{(Tensor or Number) an input tensor or number}

\item{tensor2}{(Tensor or Number) an input tensor or number}
}
\description{
Returns the \code{torch_dtype} that would result from performing an arithmetic
operation on the provided input tensors. See type promotion documentation
for more information on the type promotion logic.
}
\section{Signatures}{


result_type(tensor1, tensor2) -> dtype
}

\examples{
\dontrun{
torch_result_type(torch_tensor(c(1, 2), dtype=torch_int()), 1.0)
torch_result_type(torch_tensor(c(1, 2), dtype=torch_uint8()), torch_tensor(1))
}

}
