% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_roll}
\alias{torch_roll}
\title{Roll}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{shifts}{(int or tuple of ints) The number of places by which the elements}

\item{of}{NA}

\item{the}{NA}

\item{value}{NA}

\item{dims}{(int or tuple of ints) Axis along which to roll}
}
\description{
roll(input, shifts, dims=None) -> Tensor
}
\details{
Roll the tensor along the given dimension(s). Elements that are shifted beyond the
last position are re-introduced at the first position. If a dimension is not
specified, the tensor will be flattened before rolling and then restored
to the original shape.
}
