% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{torch_save}
\alias{torch_save}
\title{Saves an object to a disk file.}
\usage{
torch_save(obj, path, ..., compress = TRUE)
}
\arguments{
\item{obj}{the saved object}

\item{path}{a connection or the name of the file to save.}

\item{...}{not currently used.}

\item{compress}{a logical specifying whether saving to a named file is to use
"gzip" compression, or one of "gzip", "bzip2" or "xz" to indicate the type of
compression to be used. Ignored if file is a connection.}
}
\description{
This function is experimental, don't use for long
term storage.
}
\seealso{
Other torch_save:
\code{\link[=torch_load]{torch_load()}},
\code{\link[=torch_serialize]{torch_serialize()}}
}
\concept{serialization}
\concept{torch_save}
