% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_searchsorted}
\alias{torch_searchsorted}
\title{Searchsorted}
\usage{
torch_searchsorted(sorted_sequence, self, out_int32 = FALSE, right = FALSE)
}
\arguments{
\item{sorted_sequence}{(Tensor) N-D or 1-D tensor, containing monotonically increasing
sequence on the \emph{innermost} dimension.}

\item{values}{(Tensor or Scalar) N-D tensor or a Scalar containing the search value(s).}
}
\description{
Searchsorted
}
\section{searchsorted(sorted_sequence, values, *, out_int32=FALSE, right=FALSE, out=None) -> Tensor }{


Find the indices from the \emph{innermost} dimension of \code{sorted_sequence} such that, if the
corresponding values in \code{values} were inserted before the indices, the order of the
corresponding \emph{innermost} dimension within \code{sorted_sequence} would be preserved.
Return a new tensor with the same size as \code{values}. If \code{right} is FALSE (default),
then the left boundary of \code{sorted_sequence} is closed.
}

