% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R, R/gen-namespace.R
\name{torch_selu}
\alias{torch_selu}
\title{Selu}
\usage{
torch_selu(self)
}
\arguments{
\item{self}{the input tensor}
}
\description{
Selu
}
\section{selu(input) -> Tensor }{


Computes the selu transformation.
}

