% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_selu_}
\alias{torch_selu_}
\title{Selu_}
\description{
Selu_
}
\section{selu_(input) -> Tensor }{


In-place version of \code{\link{~selu}}.
}

\examples{


}
