% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_sigmoid}
\alias{torch_sigmoid}
\title{Sigmoid}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
sigmoid(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the sigmoid of the elements of :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \frac{1}{1 + e^{-\text{input}}{i}}}
}
