% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_sign}
\alias{torch_sign}
\title{Sign}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
sign(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the signs of the elements of :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \operatorname{sgn}(\text{input}}{i})
}
\examples{
\dontrun{
a = torch_tensor([0.7, -1.2, 0., 2.3])
a
torch_sign(a)
}

}
