% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_signbit}
\alias{torch_signbit}
\title{Signbit}
\usage{
torch_signbit(self)
}
\arguments{
\item{input}{(Tensor) the input tensor.}
}
\description{
Signbit
}
\section{signbit(input, *, out=None) -> Tensor }{


Tests if each element of \code{input} has its sign bit set (is less than zero) or not.
}

