% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_sin}
\alias{torch_sin}
\title{Sin}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the sine of the elements of \code{input}.
}
\details{
\deqn{
    \text{out}_{i} = \sin(\text{input}_{i})
}
}
\section{Signatures}{


sin(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(4)
a
torch_sin(a)
}

}
