% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_sparse_coo_tensor}
\alias{torch_sparse_coo_tensor}
\title{Sparse_coo_tensor}
\arguments{
\item{indices}{(array_like) Initial data for the tensor. Can be a list, tuple,}

\item{internally.}{NA}

\item{should}{NA}

\item{the}{NA}

\item{values}{(array_like) Initial values for the tensor. Can be a list, tuple,}

\item{NumPy}{NA}

\item{size}{NA class:\code{torch.Size}, optional): Size of the sparse tensor. If not}

\item{provided}{NA}

\item{elements.}{NA}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if None, uses the current device for the default tensor type}

\item{NA}{func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU}

\item{for}{NA}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}
}
\description{
sparse_coo_tensor(indices, values, size=None, dtype=None, device=None, requires_grad=False) -> Tensor
}
\details{
Constructs a sparse tensors in COO(rdinate) format with non-zero elements at the given :attr:\code{indices}
with the given :attr:\code{values}. A sparse tensor can be \code{uncoalesced}, in that case, there are duplicate
coordinates in the indices, and the value at that index is the sum of all duplicate value entries:
\code{torch.sparse}_.
}
