% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_sparse_sampled_addmm}
\alias{torch_sparse_sampled_addmm}
\title{Sparse_sampled_addmm}
\usage{
torch_sparse_sampled_addmm(self, mat1, mat2, beta = 1L, alpha = 1L)
}
\arguments{
\item{self}{(Tensor) a sparse CSR matrix of size \code{(m, n)} to be added and used to compute
the sampled matrix multiplication}

\item{mat1}{(Tensor) a dense matrix of size \code{(m, k)} to be multiplied}

\item{mat2}{(Tensor) a dense matrix of size \code{(k, n)} to be multiplied}

\item{beta}{(Number, optional) multiplier for \code{self} (\eqn{\beta})}

\item{alpha}{(Number, optional) multiplier for \eqn{mat1 @ mat2} (\eqn{\alpha})}
}
\description{
Sparse_sampled_addmm
}
\section{sparse_sampled_addmm(self, mat1, mat2, *, beta=1, alpha=1) -> Tensor }{


Performs a matrix multiplication of the dense matrices \code{mat1} and \code{mat2} at the locations
specified by the sparsity pattern of \code{self}. The matrix \code{self} is added to the final result.

Mathematically, this performs the following operation:

\deqn{
    \mbox{out} = \alpha\ (\mbox{mat1} \mathbin{@} \mbox{mat2}) * \mbox{spy}(\mbox{self}) + \beta\ \mbox{self}
}

where \eqn{\mbox{spy}(\mbox{self})} is the sparsity pattern matrix of \code{self}, \code{alpha} and
\code{beta} are the scaling factors. \eqn{\mbox{spy}(\mbox{self})} has value 1 at the positions
where \code{self} has non-zero values, and 0 elsewhere.

\code{self} must be a sparse CSR tensor. \code{mat1} and \code{mat2} must be dense tensors.
}

\examples{
if (torch_is_installed()) {

# Create a sparse CSR mask from a COO tensor
i <- torch_tensor(matrix(c(1, 2, 2, 1, 2, 3), nrow = 2, byrow = TRUE),
  dtype = torch_long())
v <- torch_tensor(c(1, 2, 3), dtype = torch_float32())
sparse_mask <- torch_sparse_coo_tensor(i, v, c(2, 3))$to_sparse_csr()

mat1 <- torch_randn(c(2, 3))
mat2 <- torch_randn(c(3, 3))

result <- torch_sparse_sampled_addmm(sparse_mask, mat1, mat2)
result$to_dense()

}
}
