% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_sqrt}
\alias{torch_sqrt}
\title{Sqrt}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the square-root of the elements of \code{input}.
}
\details{
\deqn{
    \text{out}_{i} = \sqrt{\text{input}_{i}}
}
}
\section{Signatures}{


sqrt(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(4)
a
torch_sqrt(a)
}

}
