% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_square}
\alias{torch_square}
\title{Square}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Square
}
\section{square(input, out=None) -> Tensor }{


Returns a new tensor with the square of the elements of \code{input}.
}

\examples{
\dontrun{

a = torch_randn(c(4))
a
torch_square(a)
}
}
