% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_symeig}
\alias{torch_symeig}
\title{Symeig}
\arguments{
\item{input}{(Tensor) the input tensor of size :math:\verb{(*, n, n)} where \code{*} is zero or more                    batch dimensions consisting of symmetric matrices.}

\item{eigenvectors}{(boolean, optional) controls whether eigenvectors have to be computed}

\item{upper}{(boolean, optional) controls whether to consider upper-triangular or lower-triangular region}

\item{out}{(tuple, optional) the output tuple of (Tensor, Tensor)}
}
\description{
symeig(input, eigenvectors=False, upper=True, out=None) -> (Tensor, Tensor)
}
\details{
This function returns eigenvalues and eigenvectors
of a real symmetric matrix :attr:\code{input} or a batch of real symmetric matrices,
represented by a namedtuple (eigenvalues, eigenvectors).

This function calculates all eigenvalues (and vectors) of :attr:\code{input}
such that :math:\verb{\\text\{input\} = V \\text\{diag\}(e) V^T}.

The boolean argument :attr:\code{eigenvectors} defines computation of
both eigenvectors and eigenvalues or eigenvalues only.

If it is \code{False}, only eigenvalues are computed. If it is \code{True},
both eigenvalues and eigenvectors are computed.

Since the input matrix :attr:\code{input} is supposed to be symmetric,
only the upper triangular portion is used by default.

If :attr:\code{upper} is \code{False}, then lower triangular portion is used.

.. note:: The eigenvalues are returned in ascending order. If :attr:\code{input} is a batch of matrices,
then the eigenvalues of each matrix in the batch is returned in ascending order.

.. note:: Irrespective of the original strides, the returned matrix \code{V} will
be transposed, i.e. with strides \verb{V.contiguous().transpose(-1, -2).stride()}.

.. note:: Extra care needs to be taken when backward through outputs. Such
operation is really only stable when all eigenvalues are distinct.
Otherwise, \code{NaN} can appear as the gradients are not properly defined.
}
