% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_t}
\alias{torch_t}
\title{T}
\arguments{
\item{input}{(Tensor) the input tensor.}
}
\description{
Expects \code{input} to be <= 2-D tensor and transposes dimensions 0
and 1.
}
\details{
0-D and 1-D tensors are returned as it is and
2-D tensor can be seen as a short-hand function for \code{transpose(input, 0, 1)}.
}
\section{Signatures}{


t(input) -> Tensor
}

\examples{
\dontrun{
x = torch_randn(())
x
torch_t(x)
x = torch_randn(c(3))
x
torch_t(x)
x = torch_randn(2, 3)
x
torch_t(x)
}

}
