% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_take}
\alias{torch_take}
\title{Take}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{indices}{(LongTensor) the indices into tensor}
}
\description{
Returns a new tensor with the elements of \code{input} at the given indices.
The input tensor is treated as if it were viewed as a 1-D tensor. The result
takes the same shape as the indices.
}
\section{Signatures}{


take(input, index) -> Tensor
}

\examples{
\dontrun{
src = torch_tensor([[4, 3, 5],
torch_take(src, torch_tensor([0, 2, 5]))
}

}
