% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_tensordot}
\alias{torch_tensordot}
\title{Tensordot}
\arguments{
\item{a}{(Tensor) Left tensor to contract}

\item{b}{(Tensor) Right tensor to contract}

\item{dims}{(int or tuple of two lists of integers) number of dimensions to     contract or explicit lists of dimensions for :attr:\code{a} and     :attr:\code{b} respectively}
}
\description{
Returns a contraction of a and b over multiple dimensions.
}
\details{
\preformatted{`tensordot` implements a generalized matrix product.
}
}
