% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_trace}
\alias{torch_trace}
\title{Trace}
\description{
Returns the sum of the elements of the diagonal of the input 2-D matrix.
}
\details{
Example::
}
\section{Signatures}{


trace(input) -> Tensor
}

\examples{
\dontrun{
x = torch_arange(1., 10.).view(3, 3)
x
torch_trace(x)
}

}
