% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_triangular_solve}
\alias{torch_triangular_solve}
\title{Triangular_solve}
\arguments{
\item{input}{(Tensor) multiple right-hand sides of size :math:\verb{(*, m, k)} where}

\item{NA}{math:\code{*} is zero of more batch dimensions (:math:\code{b})}

\item{A}{(Tensor) the input triangular coefficient matrix of size :math:\verb{(*, m, m)}}

\item{where}{NA math:\code{*} is zero or more batch dimensions}

\item{upper}{(bool, optional) whether to solve the upper-triangular system}

\item{of}{(default) \code{True}.}

\item{transpose}{(bool, optional) whether :math:\code{A} should be transposed before}

\item{being}{NA \code{False}.}

\item{unitriangular}{(bool, optional) whether :math:\code{A} is unit triangular.}

\item{If}{NA math:\code{A} are assumed to be}

\item{1}{NA math:\code{A}. Default: \code{False}.}
}
\description{
triangular_solve(input, A, upper=True, transpose=False, unitriangular=False) -> (Tensor, Tensor)
}
\details{
Solves a system of equations with a triangular coefficient matrix :math:\code{A}
and multiple right-hand sides :math:\code{b}.

In particular, solves :math:\code{AX = b} and assumes :math:\code{A} is upper-triangular
with the default keyword arguments.

\code{torch.triangular_solve(b, A)} can take in 2D inputs \verb{b, A} or inputs that are
batches of 2D matrices. If the inputs are batches, then returns
batched outputs \code{X}
}
