% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_tril}
\alias{torch_tril}
\title{Tril}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{diagonal}{(int, optional) the diagonal to consider}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns the lower triangular part of the matrix (2-D tensor) or batch of matrices
\code{input}, the other elements of the result tensor \code{out} are set to 0.
}
\details{
The lower triangular part of the matrix is defined as the elements on and
below the diagonal.

The argument \code{diagonal} controls which diagonal to consider. If
\code{diagonal} = 0, all elements on and below the main diagonal are
retained. A positive value includes just as many diagonals above the main
diagonal, and similarly a negative value excludes just as many diagonals below
the main diagonal. The main diagonal are the set of indices
:math:\verb{\\lbrace (i, i) \\rbrace} for :math:\verb{i \\in [0, \\min\\\{d_\{1\}, d_\{2\}\\\} - 1]} where
:math:\verb{d_\{1\}, d_\{2\}} are the dimensions of the matrix.
}
\section{Signatures}{


tril(input, diagonal=0, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(3, 3)
a
torch_tril(a)
b = torch_randn(4, 6)
b
torch_tril(b, diagonal=1)
torch_tril(b, diagonal=-1)
}

}
