% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_trunc}
\alias{torch_trunc}
\title{Trunc}
\usage{
torch_trunc(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Trunc
}
\section{trunc(input, out=NULL) -> Tensor }{


Returns a new tensor with the truncated integer values of
the elements of \code{input}.
}

\examples{

a <- torch_randn(c(4))
a
torch_trunc(a)
}
