% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_unbind}
\alias{torch_unbind}
\title{Unbind}
\arguments{
\item{input}{(Tensor) the tensor to unbind}

\item{dim}{(int) dimension to remove}
}
\description{
unbind(input, dim=0) -> seq
}
\details{
Removes a tensor dimension.

Returns a tuple of all slices along a given dimension, already without it.
}
\examples{
\dontrun{
torch_unbind(torch_tensor([[1, 2, 3],
                           [4, 5, 6],
                           [7, 8, 9]]))
}

}
