% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_unsafe_split}
\alias{torch_unsafe_split}
\title{Unsafe_split}
\usage{
torch_unsafe_split(self, split_size, dim = 1L)
}
\description{
Unsafe_split
}
\section{unsafe_split(tensor, split_size_or_sections, dim=0) -> List of Tensors }{


Works like \code{\link{torch_split}} but without enforcing the autograd restrictions
on inplace modification of the outputs.
}

\section{Warning}{

This function is safe to use as long as only the input, or only the outputs
are modified inplace after calling this function. It is user's
responsibility to ensure that is the case. If both the input and one or more
of the outputs are modified inplace, gradients computed by autograd will be
silently incorrect.
}

