% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_var}
\alias{torch_var}
\title{Var}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int or tuple of ints) the dimension or dimensions to reduce.}

\item{keepdim}{(bool) whether the output tensor has :attr:\code{dim} retained or not.}

\item{unbiased}{(bool) whether to use the unbiased estimation or not}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
var(input, unbiased=True) -> Tensor

var(input, dim, keepdim=False, unbiased=True, out=None) -> Tensor
}
\details{
Returns the variance of all elements in the :attr:\code{input} tensor.

If :attr:\code{unbiased} is \code{False}, then the variance will be calculated via the
biased estimator. Otherwise, Bessel's correction will be used.

Returns the variance of each row of the :attr:\code{input} tensor in the given
dimension :attr:\code{dim}.

If :attr:\code{keepdim} is \code{True}, the output tensor is of the same size
as :attr:\code{input} except in the dimension(s) :attr:\code{dim} where it is of size 1.
Otherwise, :attr:\code{dim} is squeezed (see :func:\code{torch.squeeze}), resulting in the
output tensor having 1 (or \code{len(dim)}) fewer dimension(s).

If :attr:\code{unbiased} is \code{False}, then the variance will be calculated via the
biased estimator. Otherwise, Bessel's correction will be used.
}
\examples{
\dontrun{
a = torch_randn(1, 3)
a
torch_var(a)
}

\dontrun{
a = torch_randn(4, 4)
a
torch_var(a, 1)
}

}
