% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_where}
\alias{torch_where}
\title{Where}
\arguments{
\item{condition}{(BoolTensor) When True (nonzero), yield x, otherwise yield y}

\item{x}{(Tensor) values selected at indices where :attr:\code{condition} is \code{True}}

\item{y}{(Tensor) values selected at indices where :attr:\code{condition} is \code{False}}
}
\description{
where(condition, x, y) -> Tensor

where(condition) -> tuple of LongTensor
}
\details{
Return a tensor of elements selected from either :attr:\code{x} or :attr:\code{y}, depending on :attr:\code{condition}.

The operation is defined as:

.. math::
\text{out}_i = \begin{cases}
\text{x}_i & \text{if } \text{condition}_i \\
\text{y}_i & \text{otherwise} \\
\end{cases}

.. note::
The tensors :attr:\code{condition}, :attr:\code{x}, :attr:\code{y} must be :ref:\verb{broadcastable <broadcasting-semantics>}.

\code{torch.where(condition)} is identical to
\code{torch.nonzero(condition, as_tuple=True)}.

.. note::
See also :func:\code{torch.nonzero}.
}
