% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_zeros_like}
\alias{torch_zeros_like}
\title{Zeros_like}
\arguments{
\item{input}{(Tensor) the size of :attr:\code{input} will determine size of the output tensor.}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned Tensor.}

\item{layout}{NA class:\code{torch.layout}, optional): the desired layout of returned tensor.}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if \code{None}, defaults to the device of :attr:\code{input}.}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}
}
\description{
zeros_like(input, dtype=None, layout=None, device=None, requires_grad=False) -> Tensor
}
\details{
Returns a tensor filled with the scalar value \code{0}, with the same size as
:attr:\code{input}. \code{torch.zeros_like(input)} is equivalent to
\code{torch.zeros(input.size(), dtype=input.dtype, layout=input.layout, device=input.device)}.

.. warning::
As of 0.4, this function does not support an :attr:\code{out} keyword. As an alternative,
the old \code{torch.zeros_like(input, out=output)} is equivalent to
\code{torch.zeros(input.size(), out=output)}.
}
