% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{utils_dataset_tensor}
\alias{utils_dataset_tensor}
\title{Dataset wrapping tensors.}
\usage{
utils_dataset_tensor(...)
}
\arguments{
\item{...}{tensors that have the same size of the first dimension.}
}
\description{
Each sample will be retrieved by indexing tensors along the first dimension.
}
