% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grad_mode.R
\name{with_no_grad}
\alias{with_no_grad}
\title{With no grad}
\usage{
with_no_grad(code)
}
\arguments{
\item{code}{code to be executed with no gradient recording.}
}
\description{
Temporarily modify gradient recording.
}
\examples{
x <- tensor(runif(5), requires_grad = TRUE)
with_no_grad({x$sub_(tensor(as.numeric(1:5)))})
x

}
