% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autograd.R
\name{with_set_grad_mode}
\alias{with_set_grad_mode}
\title{Enable or disable grad}
\usage{
with_set_grad_mode(enabled, code)
}
\arguments{
\item{enabled}{Whether to enable grad mode (otherwise disable it).}

\item{code}{code to be executed with specified gradient context.}
}
\description{
Context-manager that enables or disables gradient calculation.
If \code{enabled = TRUE} enables gradient calculation, otherwise disables it.
}
\details{
This context manager is thread local; it will not affect computation in
other threads.
}
\examples{
if (torch_is_installed()) {

x <- torch_tensor(1, requires_grad=TRUE)
with_set_grad_mode(FALSE, {
  with_set_grad_mode(TRUE, {
    y = x * 2
  })
})
y$backward()
x$grad

}
}
