% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as.array.tensor}
\alias{as.array.tensor}
\title{Tensor casting}
\usage{
\method{as.array}{tensor}(x)
}
\arguments{
\item{x}{tensor object to be casted to an R array.}
}
\description{
Casts an object with class \link{tensor} to an R atomic vector, matrix or array.
}
\examples{
x <- tensor(array(1:8, dim = c(2, 2, 2)))
as.array(x)
}
\seealso{
\code{\link[=as.matrix.tensor]{as.matrix.tensor()}}
}
