% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-utils.R
\name{broadcast_all}
\alias{broadcast_all}
\title{Given a list of values (possibly containing numbers), returns a list where each
value is broadcasted based on the following rules:}
\usage{
broadcast_all(values)
}
\arguments{
\item{values}{List of:
\itemize{
\item \code{torch.*Tensor} instances are broadcasted as per \verb{_broadcasting-semantics}.
\item \code{numeric} instances (scalars) are upcast to tensors having
the same size and type as the first tensor passed to \code{values}.  If all the
values are scalars, then they are upcasted to scalar Tensors.
values (list of \code{numeric}, \code{torch.*Tensor} or objects implementing \strong{torch_function})
}}
}
\description{
Raises value_error: if any of the values is not a \code{numeric} instance,
a \code{torch.*Tensor} instance, or an instance implementing \strong{torch_function}
TODO: add has_torch_function((v,))
See: https://github.com/pytorch/pytorch/blob/master/torch/distributions/utils.py
}
