% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-dataloader.R
\name{dataloader_next}
\alias{dataloader_next}
\title{Get the next element of a dataloader iterator}
\usage{
dataloader_next(iter)
}
\arguments{
\item{iter}{a DataLoader iter created with \link{dataloader_make_iter}.}
}
\description{
Get the next element of a dataloader iterator
}
