% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions-poisson.R
\name{distr_poisson}
\alias{distr_poisson}
\title{Creates a Poisson distribution parameterized by \code{rate}, the rate parameter.
Samples are nonnegative integers, with a pmf given by
\eqn{\mathrm{rate}^k \frac{e^{-\mathrm{rate}}}{k!}}}
\usage{
distr_poisson(rate, validate_args = NULL)
}
\arguments{
\item{rate}{(numeric, torch_tensor): the rate parameter}
}
\description{
Creates a Poisson distribution parameterized by \code{rate}, the rate parameter.
Samples are nonnegative integers, with a pmf given by
\eqn{\mathrm{rate}^k \frac{e^{-\mathrm{rate}}}{k!}}
}
\examples{
if (torch_is_installed()) {
m <- distr_poisson(torch_tensor(4))
m$sample()
}
}
