% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{linalg_det}
\alias{linalg_det}
\title{Computes the determinant of a square matrix.}
\usage{
linalg_det(A)
}
\arguments{
\item{A}{(Tensor): tensor of shape \verb{(*, n, n)} where \code{*} is zero or more batch dimensions.}
}
\description{
Supports input of float, double, cfloat and cdouble dtypes.
Also supports batches of matrices, and if \code{A} is a batch of matrices then
the output has the same batch dimensions.
}
\examples{
if (torch_is_installed()) {
a <- torch_randn(3,3)
linalg_det(a)

a <- torch_randn(3,3,3)
linalg_det(a)

}
}
\seealso{
Other linalg: 
\code{\link{linalg_matrix_norm}()},
\code{\link{linalg_norm}()},
\code{\link{linalg_vector_norm}()}
}
\concept{linalg}
