% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{linalg_eig}
\alias{linalg_eig}
\title{Computes the eigenvalue decomposition of a square matrix if it exists.}
\usage{
linalg_eig(A)
}
\arguments{
\item{A}{(Tensor): tensor of shape \verb{(*, n, n)} where \code{*} is zero or more batch dimensions
consisting of diagonalizable matrices.}
}
\value{
A list \verb{(eigenvalues, eigenvectors)} which corresponds to \eqn{\Lambda} and \eqn{V} above.
\code{eigenvalues} and \code{eigenvectors} will always be complex-valued, even when \code{A} is real. The eigenvectors
will be given by the columns of \code{eigenvectors}.
}
\description{
Letting \eqn{\mathbb{K}} be \eqn{\mathbb{R}} or \eqn{\mathbb{C}},
the \strong{eigenvalue decomposition} of a square matrix
\eqn{A \in \mathbb{K}^{n \times n}} (if it exists) is defined as
}
\details{
\deqn{
  A = V \operatorname{diag}(\Lambda) V^{-1}\mathrlap{\qquad V \in \mathbb{C}^{n \times n}, \Lambda \in \mathbb{C}^n}
}

This decomposition exists if and only if \eqn{A} is \code{diagonalizable}_.
This is the case when all its eigenvalues are different.
Supports input of float, double, cfloat and cdouble dtypes.
Also supports batches of matrices, and if \code{A} is a batch of matrices then
the output has the same batch dimensions.
}
\note{
The eigenvalues and eigenvectors of a real matrix may be complex.
}
\examples{
if (torch_is_installed()) {
a <- torch_randn(2, 2)
wv = linalg_eig(a)

}
}
\seealso{
\itemize{
\item \code{\link[=linalg_eigvals]{linalg_eigvals()}} computes only the eigenvalues. Unlike \code{\link[=linalg_eig]{linalg_eig()}}, the gradients of
\code{\link[=linalg_eigvals]{linalg_eigvals()}} are always numerically stable.
\item \code{\link[=linalg_eigh]{linalg_eigh()}} for a (faster) function that computes the eigenvalue decomposition
for Hermitian and symmetric matrices.
\item \code{\link[=linalg_svd]{linalg_svd()}} for a function that computes another type of spectral
decomposition that works on matrices of any shape.
\item \code{\link[=linalg_qr]{linalg_qr()}} for another (much faster) decomposition that works on matrices of
any shape.
}

Other linalg: 
\code{\link{linalg_cholesky}()},
\code{\link{linalg_det}()},
\code{\link{linalg_eigh}()},
\code{\link{linalg_eigvals}()},
\code{\link{linalg_matrix_norm}()},
\code{\link{linalg_matrix_rank}()},
\code{\link{linalg_norm}()},
\code{\link{linalg_qr}()},
\code{\link{linalg_slogdet}()},
\code{\link{linalg_vector_norm}()}
}
\concept{linalg}
