% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linalg.R
\name{linalg_eigh}
\alias{linalg_eigh}
\title{Computes the eigenvalue decomposition of a complex Hermitian or real symmetric matrix.}
\usage{
linalg_eigh(A, UPLO = "L")
}
\arguments{
\item{A}{(Tensor): tensor of shape \verb{(*, n, n)} where \code{*} is zero or more batch dimensions
consisting of symmetric or Hermitian matrices.}

\item{UPLO}{('L', 'U', optional): controls whether to use the upper or lower triangular part
of \code{A} in the computations. Default: \code{'L'}.}
}
\value{
A list \verb{(eigenvalues, eigenvectors)} which corresponds to \eqn{\Lambda} and \eqn{Q} above.
\code{eigenvalues} will always be real-valued, even when \code{A} is complex.

It will also be ordered in ascending order.
\code{eigenvectors} will have the same dtype as \code{A} and will contain the eigenvectors as its columns.
}
\description{
Letting \eqn{\mathbb{K}} be \eqn{\mathbb{R}} or \eqn{\mathbb{C}},
the \strong{eigenvalue decomposition} of a complex Hermitian or real symmetric matrix
\eqn{A \in \mathbb{K}^{n \times n}} is defined as
}
\details{
\deqn{
  A = Q \operatorname{diag}(\Lambda) Q^{\mbox{H}}\mathrlap{\qquad Q \in \mathbb{K}^{n \times n}, \Lambda \in \mathbb{R}^n}
}

where \eqn{Q^{\mbox{H}}} is the conjugate transpose when \eqn{Q} is complex, and the transpose when \eqn{Q} is real-valued.
\eqn{Q} is orthogonal in the real case and unitary in the complex case.

Supports input of float, double, cfloat and cdouble dtypes.
Also supports batches of matrices, and if \code{A} is a batch of matrices then
the output has the same batch dimensions.

\code{A} is assumed to be Hermitian (resp. symmetric), but this is not checked internally, instead:
\itemize{
\item If \code{UPLO}\ \verb{= 'L'} (default), only the lower triangular part of the matrix is used in the computation.
\item If \code{UPLO}\ \verb{= 'U'}, only the upper triangular part of the matrix is used.
The eigenvalues are returned in ascending order.
}
}
\note{
The eigenvalues of real symmetric or complex Hermitian matrices are always real.
}
\examples{
if (torch_is_installed()) {
a <- torch_randn(2, 2)
linalg_eigh(a)

}
}
\seealso{
\itemize{
\item \code{\link[=linalg_eigvalsh]{linalg_eigvalsh()}} computes only the eigenvalues values of a Hermitian matrix.
Unlike \code{\link[=linalg_eigh]{linalg_eigh()}}, the gradients of \code{\link[=linalg_eigvalsh]{linalg_eigvalsh()}} are always
numerically stable.
\item \code{\link[=linalg_cholesky]{linalg_cholesky()}} for a different decomposition of a Hermitian matrix.
The Cholesky decomposition gives less information about the matrix but is much faster
to compute than the eigenvalue decomposition.
\item \code{\link[=linalg_eig]{linalg_eig()}} for a (slower) function that computes the eigenvalue decomposition
of a not necessarily Hermitian square matrix.
\item \code{\link[=linalg_svd]{linalg_svd()}} for a (slower) function that computes the more general SVD
decomposition of matrices of any shape.
\item \code{\link[=linalg_qr]{linalg_qr()}} for another (much faster) decomposition that works on general
matrices.
}

Other linalg: 
\code{\link{linalg_cholesky}()},
\code{\link{linalg_det}()},
\code{\link{linalg_eigvals}()},
\code{\link{linalg_eig}()},
\code{\link{linalg_matrix_norm}()},
\code{\link{linalg_matrix_rank}()},
\code{\link{linalg_norm}()},
\code{\link{linalg_qr}()},
\code{\link{linalg_slogdet}()},
\code{\link{linalg_vector_norm}()}
}
\concept{linalg}
