% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-activation.R
\name{nn_hardswish}
\alias{nn_hardswish}
\title{Hardswish module}
\usage{
nn_hardswish()
}
\description{
Applies the hardswish function, element-wise, as described in the paper:
\href{https://arxiv.org/abs/1905.02244}{Searching for MobileNetV3}
\deqn{
  \text{Hardswish}(x) = \begin{cases}
0 & \text{if~} x \le -3, \\
x & \text{if~} x \ge +3, \\
x \cdot (x + 3) /6 & \text{otherwise}
\end{cases}
}
}
\section{Shape}{

\itemize{
\item Input: \eqn{(N, *)} where \code{*} means, any number of additional
dimensions
\item Output: \eqn{(N, *)}, same shape as the input
}
}

\examples{
\dontrun{
m <- nn_hardswish()
input <- torch_randn(2)
output <- m(input)
}

}
