% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nn_prune_head}
\alias{nn_prune_head}
\title{Prune top layer(s) of a network}
\usage{
nn_prune_head(x, head_size)
}
\arguments{
\item{x}{nn_network to prune}

\item{head_size}{number of nn_layers to prune}
}
\value{
a nn_sequential network without the top nn_layer
}
\description{
Prune top layer(s) of a network
}
