% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nn_prune_head}
\alias{nn_prune_head}
\title{Prune top layer(s) of a network}
\usage{
nn_prune_head(x, head_size)
}
\arguments{
\item{x}{nn_network to prune}

\item{head_size}{number of nn_layers to prune}
}
\value{
a nn_sequential network with the top nn_layer removed
}
\description{
Prune \code{head_size} last layers of a nn_module in order to
replace them by your own head, or in order to use the pruned module
as a sequential embedding module.
}
\examples{

# Remove the classification head of a resnet34 pretrained network
if (torch_is_installed()) {
  resnet34 <- torchvision::network_resnet34(pretrained=TRUE)
  resnet34_headless <- nn_prune_head(resnet34)
}
}
