% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn.R
\name{nn_sequential}
\alias{nn_sequential}
\title{A sequential container}
\usage{
nn_sequential(...)
}
\arguments{
\item{...}{sequence of modules to be added}
}
\description{
A sequential container.
Modules will be added to it in the order they are passed in the constructor.
See examples.
}
\examples{
if (torch_is_installed()) {

model <- nn_sequential(
  nn_conv2d(1, 20, 5),
  nn_relu(),
  nn_conv2d(20, 64, 5),
  nn_relu()
)
input <- torch_randn(32, 1, 28, 28)
output <- model(input)
}
}
