% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn-activation.R
\name{nn_softsign}
\alias{nn_softsign}
\title{Softsign module}
\usage{
nn_softsign()
}
\description{
Applies the element-wise function:
\deqn{
  \mbox{SoftSign}(x) = \frac{x}{ 1 + |x|}
}
}
\section{Shape}{

\itemize{
\item Input: \eqn{(N, *)} where \code{*} means, any number of additional
dimensions
\item Output: \eqn{(N, *)}, same shape as the input
}
}

\examples{
if (torch_is_installed()) {
m <- nn_softsign()
input <- torch_randn(2)
output <- m(input)
}
}
