% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_adaptive_avg_pool1d}
\alias{nnf_adaptive_avg_pool1d}
\title{Adaptive_avg_pool1d}
\arguments{
\item{output_size}{NA the target output size (single integer)}
}
\description{
Adaptive_avg_pool1d
}
\section{adaptive_avg_pool1d(input, output_size) -> Tensor }{


Applies a 1D adaptive average pooling over an input signal composed of
several input planes.

See \code{~torch.nn.AdaptiveAvgPool1d} for details and output shape.
}

\examples{


}
