% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-pooling.R
\name{nnf_adaptive_avg_pool3d}
\alias{nnf_adaptive_avg_pool3d}
\title{Adaptive_avg_pool3d}
\usage{
nnf_adaptive_avg_pool3d(input, output_size)
}
\arguments{
\item{input}{input tensor (minibatch, in_channels , iT \times iH , iW)}

\item{output_size}{the target output size (single integer or triple-integer tuple)}
}
\description{
Applies a 3D adaptive average pooling over an input signal composed of
several input planes.
}
\examples{


}
