% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-pooling.R
\name{nnf_adaptive_max_pool2d}
\alias{nnf_adaptive_max_pool2d}
\title{Adaptive_max_pool2d}
\usage{
nnf_adaptive_max_pool2d(input, output_size, return_indices = FALSE)
}
\arguments{
\item{input}{input tensor (minibatch, in_channels , iH , iW)}

\item{output_size}{the target output size (single integer or double-integer tuple)}

\item{return_indices}{whether to return pooling indices. Default: \code{FALSE}}
}
\description{
Applies a 2D adaptive max pooling over an input signal composed of
several input planes.
}
\examples{


}
