% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_alpha_dropout}
\alias{nnf_alpha_dropout}
\title{Alpha_dropout}
\description{
Applies alpha dropout to the input.
}
\details{
\preformatted{See `~torch.nn.AlphaDropout` for details.
}
}
\examples{


}
