% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-pooling.R
\name{nnf_avg_pool1d}
\alias{nnf_avg_pool1d}
\title{Avg_pool1d}
\usage{
nnf_avg_pool1d(
  input,
  kernel_size,
  stride = list(),
  padding = 0,
  ceil_mode = FALSE,
  count_include_pad = TRUE
)
}
\arguments{
\item{input}{input tensor of shape (minibatch , in_channels , iW)}

\item{kernel_size}{the size of the window. Can be a single number or a
tuple \verb{(kW,)}.}

\item{stride}{the stride of the window. Can be a single number or a tuple
\verb{(sW,)}. Default: \code{kernel_size}}

\item{padding}{implicit zero paddings on both sides of the input. Can be a
single number or a tuple \verb{(padW,)}. Default: 0}

\item{ceil_mode}{when True, will use \code{ceil} instead of \code{floor} to compute the
output shape. Default: \code{FALSE}}

\item{count_include_pad}{when True, will include the zero-padding in the
averaging calculation. Default: \code{TRUE}}
}
\description{
Applies a 1D average pooling over an input signal composed of several
input planes.
}
