% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-batchnorm.R
\name{nnf_batch_norm}
\alias{nnf_batch_norm}
\title{Batch_norm}
\usage{
nnf_batch_norm(
  input,
  running_mean,
  running_var,
  weight = NULL,
  bias = NULL,
  training = FALSE,
  momentum = 0.1,
  eps = 1e-05
)
}
\arguments{
\item{input}{input tensor}

\item{running_mean}{the running_mean tensor}

\item{running_var}{the running_var tensor}

\item{weight}{the weight tensor}

\item{bias}{the bias tensor}

\item{training}{bool wether it's training. Default: FALSE}

\item{momentum}{the value used for the \code{running_mean} and \code{running_var} computation.
Can be set to None for cumulative moving average (i.e. simple average). Default: 0.1}

\item{eps}{a value added to the denominator for numerical stability. Default: 1e-5}
}
\description{
Applies Batch Normalization for each channel across a batch of data.
}
\examples{


}
