% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_celu}
\alias{nnf_celu}
\alias{nnf_celu_}
\title{Celu}
\usage{
nnf_celu(input, alpha = 1, inplace = FALSE)

nnf_celu_(input, alpha = 1)
}
\arguments{
\item{input}{(N,*) tensor, where * means, any number of additional
dimensions}

\item{alpha}{the alpha value for the CELU formulation. Default: 1.0}

\item{inplace}{can optionally do the operation in-place. Default: FALSE}
}
\description{
Applies element-wise, \eqn{CELU(x) = max(0,x) + min(0, \alpha * (exp(x \alpha) - 1))}.
}
