% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_dropout}
\alias{nnf_dropout}
\title{Dropout}
\arguments{
\item{p}{NA probability of an element to be zeroed. Default: 0.5}

\item{training}{NA apply dropout if is \code{True}. Default: \code{True}}

\item{inplace}{NA If set to \code{True}, will do this operation in-place. Default: \code{False}}
}
\description{
During training, randomly zeroes some of the elements of the input
tensor with probability \code{p} using samples from a Bernoulli
distribution.
}
\details{
\preformatted{See `~torch.nn.Dropout` for details.
}
}
\examples{


}
