% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-dropout.R
\name{nnf_dropout2d}
\alias{nnf_dropout2d}
\title{Dropout2d}
\usage{
nnf_dropout2d(input, p = 0.5, training = TRUE, inplace = FALSE)
}
\arguments{
\item{input}{the input tensor}

\item{p}{probability of a channel to be zeroed. Default: 0.5}

\item{training}{apply dropout if is \code{TRUE}. Default: \code{TRUE}.}

\item{inplace}{If set to \code{TRUE}, will do this operation in-place.
Default: \code{FALSE}}
}
\description{
Randomly zero out entire channels (a channel is a 2D feature map,
e.g., the \eqn{j}-th channel of the \eqn{i}-th sample in the
batched input is a 2D tensor \eqn{input[i, j]}) of the input tensor).
Each channel will be zeroed out independently on every forward call with
probability \code{p} using samples from a Bernoulli distribution.
}
