% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_fold}
\alias{nnf_fold}
\title{Fold}
\description{
Combines an array of sliding local blocks into a large containing
tensor.
}
\details{
\preformatted{.. warning::
    Currently, only 4-D output tensors (batched image-like tensors) are
    supported.

See `torch_nn.Fold` for details
}
}
\examples{


}
