% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_gelu}
\alias{nnf_gelu}
\title{Gelu}
\description{
Gelu
}
\section{gelu(input) -> Tensor }{


Applies element-wise the function
\eqn{\text{GELU}(x) = x * \Phi(x)}\preformatted{where \eqn{\Phi(x)} is the Cumulative Distribution Function for Gaussian Distribution.

See `Gaussian Error Linear Units (GELUs) <https://arxiv.org/abs/1606.08415>`_.
}
}

\examples{


}
