% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R,
%   R/nnf-normalization.R
\name{nnf_group_norm}
\alias{nnf_group_norm}
\title{Group_norm}
\usage{
nnf_group_norm(input, num_groups, weight = NULL, bias = NULL, eps = 1e-05)
}
\arguments{
\item{input}{the input tensor}

\item{num_groups}{number of groups to separate the channels into}

\item{weight}{the weight tensor}

\item{bias}{the bias tensor}

\item{eps}{a value added to the denominator for numerical stability. Default: 1e-5}
}
\description{
Applies Group Normalization for last certain number of dimensions.
}
\examples{


}
