% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_hardshrink}
\alias{nnf_hardshrink}
\title{Hardshrink}
\usage{
nnf_hardshrink(input, lambd = 0.5)
}
\arguments{
\item{input}{(N,∗) tensor, where * means, any number of additional
dimensions}

\item{lambd}{the lambda value for the Hardshrink formulation. Default: 0.5}
}
\description{
Applies the hard shrinkage function element-wise
}
