% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_hardsigmoid}
\alias{nnf_hardsigmoid}
\title{Hardsigmoid}
\arguments{
\item{inplace}{NA If set to \code{True}, will do this operation in-place. Default: \code{False}}
}
\description{
Hardsigmoid
}
\section{hardsigmoid(input) -> Tensor }{


Applies the element-wise function \eqn{\text{Hardsigmoid}(x) = \frac{ReLU6(x + 3)}{6}}
}

\examples{


}
