% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_hardtanh}
\alias{nnf_hardtanh}
\alias{nnf_hardtanh_}
\title{Hardtanh}
\usage{
nnf_hardtanh(input, min_val = -1, max_val = 1, inplace = FALSE)

nnf_hardtanh_(input, min_val = -1, max_val = 1)
}
\arguments{
\item{input}{(N,∗) tensor, where * means, any number of additional
dimensions}

\item{min_val}{minimum value of the linear region range. Default: -1}

\item{max_val}{maximum value of the linear region range. Default: 1}

\item{inplace}{can optionally do the operation in-place. Default: FALSE}
}
\description{
Applies the HardTanh function element-wise.
}
