% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_hinge_embedding_loss}
\alias{nnf_hinge_embedding_loss}
\title{Hinge_embedding_loss}
\description{
Hinge_embedding_loss
}
\section{hinge_embedding_loss(input, target, margin=1.0, size_average=None, reduce=None, reduction='mean') -> Tensor }{


See \code{~torch.nn.HingeEmbeddingLoss} for details.
}

\examples{


}
