% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_instance_norm}
\alias{nnf_instance_norm}
\title{Instance_norm}
\description{
Applies Instance Normalization for each channel in each data sample in a
batch.
}
\details{
\preformatted{See `~torch.nn.InstanceNorm1d`, `~torch.nn.InstanceNorm2d`,
`~torch.nn.InstanceNorm3d` for details.
}
}
\examples{


}
