% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_kl_div}
\alias{nnf_kl_div}
\title{Kl_div}
\arguments{
\item{input}{NA Tensor of arbitrary shape}

\item{target}{NA Tensor of the same shape as input}

\item{size_average}{(bool, optional) Deprecated (see \code{reduction}). By default,        the losses are averaged over each loss element in the batch. Note that for        some losses, there multiple elements per sample. If the field \code{size_average}        is set to \code{False}, the losses are instead summed for each minibatch. Ignored        when reduce is \code{False}. Default: \code{True}}

\item{reduce}{(bool, optional) Deprecated (see \code{reduction}). By default, the        losses are averaged or summed over observations for each minibatch depending        on \code{size_average}. When \code{reduce} is \code{False}, returns a loss per        batch element instead and ignores \code{size_average}. Default: \code{True}}

\item{reduction}{(string, optional) Specifies the reduction to apply to the output:        \code{'none'} | \code{'batchmean'} | \code{'sum'} | \code{'mean'}.        \code{'none'}: no reduction will be applied        \code{'batchmean'}: the sum of the output will be divided by the batchsize        \code{'sum'}: the output will be summed        \code{'mean'}: the output will be divided by the number of elements in the output        Default: \code{'mean'}}
}
\description{
The \verb{Kullback-Leibler divergence}_ Loss.
}
\details{
\preformatted{See `~torch.nn.KLDivLoss` for details.
}
}
\examples{


}
