% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-normalization.R
\name{nnf_layer_norm}
\alias{nnf_layer_norm}
\title{Layer_norm}
\usage{
nnf_layer_norm(
  input,
  normalized_shape,
  weight = NULL,
  bias = NULL,
  eps = 1e-05
)
}
\arguments{
\item{input}{the input tensor}

\item{normalized_shape}{input shape from an expected input of size. If a single
integer is used, it is treated as a singleton list, and this module will normalize
over the last dimension which is expected to be of that specific size.}

\item{weight}{the weight tensor}

\item{bias}{the bias tensor}

\item{eps}{a value added to the denominator for numerical stability. Default: 1e-5}
}
\description{
Applies Layer Normalization for last certain number of dimensions.
}
