% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_leaky_relu}
\alias{nnf_leaky_relu}
\title{Leaky_relu}
\usage{
nnf_leaky_relu(input, negative_slope = 0.01, inplace = FALSE)
}
\arguments{
\item{input}{(N,*) tensor, where * means, any number of additional
dimensions}

\item{negative_slope}{Controls the angle of the negative slope. Default: 1e-2}

\item{inplace}{can optionally do the operation in-place. Default: FALSE}
}
\description{
Applies element-wise,
\eqn{LeakyReLU(x) = max(0, x) + negative_slope * min(0, x)}
}
