% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_linear}
\alias{nnf_linear}
\title{Linear}
\description{
Applies a linear transformation to the incoming data: \eqn{y = xA^T + b}.
}
\details{
\preformatted{Shape:

    - Input: \eqn{(N, *, in\_features)} where `*` means any number of
      additional dimensions
    - Weight: \eqn{(out\_features, in\_features)}
    - Bias: \eqn{(out\_features)}
    - Output: \eqn{(N, *, out\_features)}
}
}
\examples{


}
