% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_logsigmoid}
\alias{nnf_logsigmoid}
\title{Logsigmoid}
\usage{
nnf_logsigmoid(input)
}
\arguments{
\item{input}{(N,*) tensor, where * means, any number of additional
dimensions}
}
\description{
Applies element-wise \eqn{LogSigmoid(x_i) = log(\frac{1}{1 + exp(-x_i)})}
}
