% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-pooling.R
\name{nnf_lp_pool1d}
\alias{nnf_lp_pool1d}
\title{Lp_pool1d}
\usage{
nnf_lp_pool1d(input, norm_type, kernel_size, stride = NULL, ceil_mode = FALSE)
}
\arguments{
\item{input}{the input tensor}

\item{norm_type}{if inf than one gets max pooling if 0 you get sum pooling (
proportional to the avg pooling)}

\item{kernel_size}{a single int, the size of the window}

\item{stride}{a single int, the stride of the window. Default value is kernel_size}

\item{ceil_mode}{when True, will use ceil instead of floor to compute the output shape}
}
\description{
Applies a 1D power-average pooling over an input signal composed of
several input planes. If the sum of all inputs to the power of \code{p} is
zero, the gradient is set to zero as well.
}
