% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_max_pool1d}
\alias{nnf_max_pool1d}
\title{Max_pool1d}
\description{
Applies a 1D max pooling over an input signal composed of several input
planes.
}
\details{
\preformatted{See `~torch.nn.MaxPool1d` for details.
}
}
\examples{


}
