% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-pooling.R
\name{nnf_max_unpool2d}
\alias{nnf_max_unpool2d}
\title{Max_unpool2d}
\usage{
nnf_max_unpool2d(
  input,
  indices,
  kernel_size,
  stride = list(),
  padding = 0,
  output_size = NULL
)
}
\arguments{
\item{input}{the input Tensor to invert}

\item{indices}{the indices given out by max pool}

\item{kernel_size}{Size of the max pooling window.}

\item{stride}{Stride of the max pooling window. It is set to kernel_size by default.}

\item{padding}{Padding that was added to the input}

\item{output_size}{the targeted output size}
}
\description{
Computes a partial inverse of \code{MaxPool2d}.
}
