% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_multi_head_attention_forward}
\alias{nnf_multi_head_attention_forward}
\title{Multi_head_attention_forward}
\arguments{
\item{query, }{NA map a query and a set of key-value pairs to an output.            See "Attention Is All You Need" for more details.}

\item{embed_dim_to_check}{NA total dimension of the model.}

\item{num_heads}{NA parallel attention heads.}

\item{in_proj_weight, }{NA input projection weight and bias.}

\item{bias_k, }{NA bias of the key and value sequences to be added at dim=0.}

\item{add_zero_attn}{NA add a new batch of zeros to the key and                       value sequences at dim=1.}

\item{dropout_p}{NA probability of an element to be zeroed.}

\item{out_proj_weight, }{NA the output projection weight and bias.}

\item{training}{NA apply dropout if is \code{True}.}

\item{key_padding_mask}{NA if provided, specified padding elements in the key will            be ignored by the attention. This is an binary mask. When the value is True,            the corresponding value on the attention layer will be filled with -inf.}

\item{need_weights}{NA output attn_output_weights.}

\item{attn_mask}{NA 2D or 3D mask that prevents attention to certain positions. This is an additive mask            (i.e. the values will be added to the attention layer). A 2D mask will be broadcasted for all            the batches while a 3D mask allows to specify a different mask for the entries of each batch.}

\item{use_separate_proj_weight}{NA the function accept the proj. weights for query, key,            and value in different forms. If false, in_proj_weight will be used, which is            a combination of q_proj_weight, k_proj_weight, v_proj_weight.}

\item{q_proj_weight, }{NA input projection weight and bias.}

\item{static_k, }{NA static key and value used for attention operators.    Shape:}

\item{Inputs:}{NA}

\item{-}{NA \eqn{(L, N, E)} where L is the target sequence length, N is the batch size, E is          the embedding dimension.}

\item{Outputs:}{NA}
}
\description{
Args:
}
\examples{


}
