% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_multi_head_attention_forward}
\alias{nnf_multi_head_attention_forward}
\title{Multi head attention forward}
\usage{
nnf_multi_head_attention_forward(
  query,
  key,
  value,
  embed_dim_to_check,
  num_heads,
  in_proj_weight,
  in_proj_bias,
  bias_k,
  bias_v,
  add_zero_attn,
  dropout_p,
  out_proj_weight,
  out_proj_bias,
  training = TRUE,
  key_padding_mask = NULL,
  need_weights = TRUE,
  attn_mask = NULL,
  use_separate_proj_weight = FALSE,
  q_proj_weight = NULL,
  k_proj_weight = NULL,
  v_proj_weight = NULL,
  static_k = NULL,
  static_v = NULL
)
}
\arguments{
\item{query}{map a query and a set of key-value pairs to an output.
See "Attention Is All You Need" for more details.}

\item{embed_dim_to_check}{total dimension of the model.}

\item{num_heads}{parallel attention heads.}

\item{in_proj_weight}{input projection weight and bias.}

\item{bias_k}{bias of the key and value sequences to be added at dim=0.}

\item{add_zero_attn}{add a new batch of zeros to the key and
value sequences at dim=1.}

\item{dropout_p}{probability of an element to be zeroed.}

\item{out_proj_weight}{the output projection weight and bias.}

\item{training}{apply dropout if is \code{TRUE}.}

\item{key_padding_mask}{if provided, specified padding elements in the key will
be ignored by the attention. This is an binary mask. When the value is True
the corresponding value on the attention layer will be filled with -inf.}

\item{need_weights}{output attn_output_weights.}

\item{attn_mask}{2D or 3D mask that prevents attention to certain positions.
This is an additive mask (i.e. the values will be added to the attention layer).
A 2D mask will be broadcasted for all the batches while a 3D mask allows to
specify a different mask for the entries of each batch.}

\item{use_separate_proj_weight}{the function accept the proj. weights for
query, key, and value in different forms. If false, in_proj_weight will be used,
which is a combination of q_proj_weight, k_proj_weight, v_proj_weight.}

\item{q_proj_weight}{input projection weight and bias.}

\item{static_k}{static key and value used for attention operators.}
}
\description{
Allows the model to jointly attend to information from different representation
subspaces. See reference: Attention Is All You Need
}
