% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-loss.R
\name{nnf_multilabel_soft_margin_loss}
\alias{nnf_multilabel_soft_margin_loss}
\title{Multilabel_soft_margin_loss}
\usage{
nnf_multilabel_soft_margin_loss(input, target, weight, reduction = "mean")
}
\arguments{
\item{input}{tensor (N,∗) where *∗ means, any number of additional dimensions}

\item{target}{tensor (N,∗) , same shape as the input}

\item{reduction}{(string, optional) – Specifies the reduction to apply to the
output: 'none' | 'mean' | 'sum'. 'none': no reduction will be applied, 'mean':
the sum of the output will be divided by the number of elements in the output,
'sum': the output will be summed. Default: 'mean'}
}
\description{
Creates a criterion that optimizes a multi-label one-versus-all loss based on
max-entropy, between input x and target y of size (N, C).
}
\examples{


}
