% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-distance.R
\name{nnf_pairwise_distance}
\alias{nnf_pairwise_distance}
\title{Pairwise_distance}
\usage{
nnf_pairwise_distance(x1, x2, p = 2, eps = 1e-06, keepdim = FALSE)
}
\arguments{
\item{x1}{(Tensor) First input.}

\item{x2}{(Tensor) Second input (of size matching x1).}

\item{p}{the norm degree. Default: 2}

\item{eps}{(float, optional) Small value to avoid division by zero.
Default: 1e-8}

\item{keepdim}{Determines whether or not to keep the vector dimension. Default: False}
}
\description{
Computes the batchwise pairwise distance between vectors using the p-norm.
}
