% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_pixel_shuffle}
\alias{nnf_pixel_shuffle}
\title{Pixel_shuffle}
\arguments{
\item{input}{(Tensor) the input tensor}

\item{upscale_factor}{(int) factor to increase spatial resolution by}
}
\description{
Pixel_shuffle
}
\section{Rearranges elements in a tensor of shape }{
math:\verb{(*, C \\times r^2, H, W)} to a :

Rearranges elements in a tensor of shape \eqn{(*, C \times r^2, H, W)} to a
tensor of shape \eqn{(*, C, H \times r, W \times r)}.

See \code{~torch.nn.PixelShuffle} for details.
}

\examples{

input = torch_randn(c(1, 9, 4, 4))
output = torch_nn$functional$pixel_shuffle(input, 3)
print(output$size())
}
