% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_poisson_nll_loss}
\alias{nnf_poisson_nll_loss}
\title{Poisson_nll_loss}
\arguments{
\item{input}{NA expectation of underlying Poisson distribution.}

\item{target}{NA random sample \eqn{target \sim \text{Poisson}(input)}.}

\item{log_input}{NA if \code{True} the loss is computed as        \eqn{\exp(\text{input}) - \text{target} * \text{input}}, if \code{False} then loss is        \eqn{\text{input} - \text{target} * \log(\text{input}+\text{eps})}. Default: \code{True}}

\item{full}{NA whether to compute full loss, i. e. to add the Stirling        approximation term. Default: \code{False}        \eqn{\text{target} * \log(\text{target}) - \text{target} + 0.5 * \log(2 * \pi * \text{target})}.}

\item{size_average}{(bool, optional) Deprecated (see \code{reduction}). By default,        the losses are averaged over each loss element in the batch. Note that for        some losses, there multiple elements per sample. If the field \code{size_average}        is set to \code{False}, the losses are instead summed for each minibatch. Ignored        when reduce is \code{False}. Default: \code{True}}

\item{eps}{(float, optional) Small value to avoid evaluation of \eqn{\log(0)} when        \code{log_input}=\code{False}. Default: 1e-8}

\item{reduce}{(bool, optional) Deprecated (see \code{reduction}). By default, the        losses are averaged or summed over observations for each minibatch depending        on \code{size_average}. When \code{reduce} is \code{False}, returns a loss per        batch element instead and ignores \code{size_average}. Default: \code{True}}

\item{reduction}{(string, optional) Specifies the reduction to apply to the output:        \code{'none'} | \code{'mean'} | \code{'sum'}. \code{'none'}: no reduction will be applied,        \code{'mean'}: the sum of the output will be divided by the number of        elements in the output, \code{'sum'}: the output will be summed. Note: \code{size_average}        and \code{reduce} are in the process of being deprecated, and in the meantime,        specifying either of those two args will override \code{reduction}. Default: \code{'mean'}}
}
\description{
Poisson negative log likelihood loss.
}
\details{
\preformatted{See `~torch.nn.PoissonNLLLoss` for details.
}
}
\examples{


}
