% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_prelu}
\alias{nnf_prelu}
\title{Prelu}
\usage{
nnf_prelu(input, weight)
}
\arguments{
\item{input}{(N,*) tensor, where * means, any number of additional
dimensions}

\item{weight}{(Tensor) the learnable weights}
}
\description{
Applies element-wise the function
\eqn{PReLU(x) = max(0,x) + weight * min(0,x)}
where weight is a learnable parameter.
}
