% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_relu}
\alias{nnf_relu}
\title{Relu}
\description{
Relu
}
\section{relu(input, inplace=False) -> Tensor }{


Applies the rectified linear unit function element-wise. See
\code{~torch.nn.ReLU} for more details.
}

\examples{


}
