% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_relu6}
\alias{nnf_relu6}
\title{Relu6}
\description{
Relu6
}
\section{relu6(input, inplace=False) -> Tensor }{


Applies the element-wise function \eqn{\text{ReLU6}(x) = \min(\max(0,x), 6)}.\preformatted{See `~torch.nn.ReLU6` for more details.
}
}

\examples{


}
