% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_rrelu}
\alias{nnf_rrelu}
\alias{nnf_rrelu_}
\title{Rrelu}
\usage{
nnf_rrelu(input, lower = 1/8, upper = 1/3, training = FALSE, inplace = FALSE)

nnf_rrelu_(input, lower = 1/8, upper = 1/3, training = FALSE)
}
\arguments{
\item{input}{(N,*) tensor, where * means, any number of additional
dimensions}

\item{lower}{lower bound of the uniform distribution. Default: 1/8}

\item{upper}{upper bound of the uniform distribution. Default: 1/3}

\item{training}{bool wether it's a training pass. DEfault: FALSE}

\item{inplace}{can optionally do the operation in-place. Default: FALSE}
}
\description{
Randomized leaky ReLU.
}
