% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-docs.R,
%   R/gen-nn-functional-examples.R
\name{nnf_sigmoid}
\alias{nnf_sigmoid}
\title{Sigmoid}
\description{
Sigmoid
}
\section{sigmoid(input) -> Tensor }{


Applies the element-wise function \eqn{\text{Sigmoid}(x) = \frac{1}{1 + \exp(-x)}}\preformatted{See `~torch.nn.Sigmoid` for more details.
}
}

\examples{


}
