% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R, R/nnf-loss.R
\name{nnf_smooth_l1_loss}
\alias{nnf_smooth_l1_loss}
\title{Smooth_l1_loss}
\usage{
nnf_smooth_l1_loss(input, target, reduction = "mean")
}
\arguments{
\item{input}{tensor (N,∗) where *∗ means, any number of additional dimensions}

\item{target}{tensor (N,∗) , same shape as the input}

\item{reduction}{(string, optional) – Specifies the reduction to apply to the
output: 'none' | 'mean' | 'sum'. 'none': no reduction will be applied, 'mean':
the sum of the output will be divided by the number of elements in the output,
'sum': the output will be summed. Default: 'mean'}
}
\description{
Function that uses a squared term if the absolute
element-wise error falls below 1 and an L1 term otherwise.
}
\examples{


}
