% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R,
%   R/nnf-activation.R
\name{nnf_softmin}
\alias{nnf_softmin}
\title{Softmin}
\usage{
nnf_softmin(input, dim, dtype = NULL)
}
\arguments{
\item{input}{(Tensor) input}

\item{dim}{(int) A dimension along which softmin will be computed
(so every slice        along dim will sum to 1).}

\item{dtype}{(\code{torch.dtype}, optional) the desired data type of returned tensor.      If specified, the input tensor is casted to \code{dtype} before the operation      is performed.
This is useful for preventing data type overflows. Default: NULL.}
}
\description{
Applies a softmin function.
}
\details{
Note that

\deqn{Softmin(x) = Softmax(-x)}.

See \link{nnf_softmax} definition for mathematical formula.
}
\examples{


}
