% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnf-activation.R
\name{nnf_softshrink}
\alias{nnf_softshrink}
\title{Softshrink}
\usage{
nnf_softshrink(input, lambd = 0.5)
}
\arguments{
\item{input}{(N,∗) tensor, where * means, any number of additional
dimensions}

\item{lambd}{the lambda (must be no less than zero) value for the Softshrink
formulation. Default: 0.5}
}
\description{
Applies the soft shrinkage function elementwise
}
