% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R,
%   R/nnf-activation.R
\name{nnf_softsign}
\alias{nnf_softsign}
\title{Softsign}
\usage{
nnf_softsign(input)
}
\arguments{
\item{input}{(N,∗) tensor, where * means, any number of additional
dimensions}
}
\description{
Applies element-wise, the function \eqn{SoftSign(x) = x/(1 + |x|}
}
\examples{


}
