% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-nn-functional-examples.R,
%   R/nnf-activation.R
\name{nnf_tanhshrink}
\alias{nnf_tanhshrink}
\title{Tanhshrink}
\usage{
nnf_tanhshrink(input)
}
\arguments{
\item{input}{(N,∗) tensor, where * means, any number of additional
dimensions}
}
\description{
Applies element-wise, \eqn{Tanhshrink(x) = x - Tanh(x)}
}
\examples{


}
