% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data-dataloader.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_iterator}
\alias{loop}
\alias{yield}
\title{Re-exporting the as_iterator function.}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{coro}{\code{\link[coro]{as_iterator}}, \code{\link[coro:collect]{loop}}, \code{\link[coro]{yield}}}
}}

