% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_max}
\alias{tch_max}
\title{maX}
\usage{
tch_max(x, dim = NULL, keepdim = NULL, dtype = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{dim}{the dimension to reduce}

\item{keepdim}{wether to keep or not the dim}

\item{dtype}{optionaly cast the sum result}
}
\description{
Returns the maximum value of all elements in the x tensor.
}
\examples{
x <- tensor(runif(100))
tch_max(x)
}
