% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_mean}
\alias{tch_mean}
\title{mean}
\usage{
tch_mean(x, dim = NULL, keepdim = NULL, dtype = NULL,
  na.rm = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{dim}{dimension in which to sum}

\item{keepdim}{wether to keep or not the dim}

\item{dtype}{optionaly cast the sum result}
}
\description{
mean
}
\examples{
x <- tensor(runif(100))
tch_mean(x)
}
