% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_median}
\alias{tch_median}
\title{median}
\usage{
tch_median(x, dim = -1, keepdim = FALSE)
}
\arguments{
\item{x}{tensor object}

\item{dim}{the dimension to reduce}

\item{keepdim}{wether to keep or not the dim}
}
\description{
Returns the median value of each row of the x tensor in the given dimension dim.
Also returns the index location of the median value as a LongTensor.
}
\details{
By default, dim is the last dimension of the x tensor.
}
\examples{
x <- tensor(array(1:20, c(5,4)))
tch_median(x)
tch_median(x, 0)
tch_median(x, 1)
}
