% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_prod}
\alias{tch_prod}
\title{prod}
\usage{
tch_prod(x, dim = NULL, keepdim = NULL, dtype = NULL)
}
\arguments{
\item{x}{tensor object}

\item{dim}{the dimension to reduce}

\item{keepdim}{wether to keep or not the dim}

\item{dtype}{the desired data type of returned tensor. If specified, the input tensor is casted to dtype before the operation is performed.
This is useful for preventing data type overflows. Default: NULL.}
}
\description{
Returns the product of all elements in the x tensor.
}
\examples{
x <- tch_randn(c(2,2))
tch_prod(x)
tch_prod(x, 0)
tch_prod(x, 1, TRUE)
tch_prod(x, 1, TRUE, "double")
}
