% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tch_randn}
\alias{tch_randn}
\title{Random normal}
\usage{
tch_randn(sizes)
}
\arguments{
\item{sizes}{a sequence of integers defining the shape of the output tensor.}
}
\description{
Returns a tensor filled with random numbers
from a normal distribution with mean 0 and variance 1
(also called the standard normal distribution).
}
\details{
The shape of the tensor is defined by the variable argument sizes.
}
\examples{
tch_randn(c(2,2))

}
