% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor.R
\name{tensor}
\alias{tensor}
\title{Create torch Tensor from R object}
\usage{
tensor(x)
}
\arguments{
\item{x}{an R vector, matrix or array.}
}
\description{
Create torch Tensor from R object
}
\note{
it uses the R type when creating the tensor.
}
\examples{
tensor(1:10)
tensor(array(runif(8), dim = c(2, 2, 2)))
tensor(matrix(c(TRUE, FALSE), nrow = 3, ncol = 4))
}
