% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{tensor}
\alias{tensor}
\title{Creates a torch tensor.}
\usage{
tensor(x, ...)
}
\arguments{
\item{x}{an R object or a torch tensor.}

\item{dtype}{the desired data type of returned tensor. Default: if \code{NULL}, infers
data type from \code{x}.}

\item{device}{the desired device of returned tensor. Default: if \code{NULL}, uses
the current device for the default tensor type (see \code{tch_set_default_tensor_type()}).
device will be the CPU for CPU tensor types and the current CUDA device for
CUDA tensor types.}

\item{requires_grad}{If autograd should record operations on the
returned tensor. Default: \code{FALSE}.}
}
\description{
Creates a torch tensor.
}
\examples{
x <- tensor(1:10)
x

y <- tensor(x, dtype = "double")
y
}
