% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_abs}
\alias{torch_abs}
\title{Abs}
\usage{
torch_abs(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Abs
}
\section{abs(input) -> Tensor }{


Computes the element-wise absolute value of the given \code{input} tensor.

\deqn{
    \mbox{out}_{i} = |\mbox{input}_{i}|
}
}

\examples{

torch_abs(torch_tensor(c(-1, -2, 3)))
}
