% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_abs}
\alias{torch_abs}
\title{Abs}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
abs(input, out=None) -> Tensor
}
\details{
Computes the element-wise absolute value of the given :attr:\code{input} tensor.

.. math::
\text{out}\emph{{i} = |\text{input}}{i}|
}
