% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_abs}
\alias{torch_abs}
\title{Abs}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Abs
}
\section{abs(input, out=None) -> Tensor }{


Computes the element-wise absolute value of the given \code{input} tensor.

\deqn{
    \mbox{out}_{i} = |\mbox{input}_{i}|
}
}

\examples{
if (torch_is_installed()) {

torch_abs(torch_tensor(c(-1, -2, 3)))
}
}
