% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_abs}
\alias{torch_abs}
\title{Abs}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Computes the element-wise absolute value of the given \code{input} tensor.
}
\details{
\deqn{
    \text{out}_{i} = |\text{input}_{i}|
}
}
\section{Signatures}{


abs(input, out=None) -> Tensor
}

\examples{
\dontrun{
torch_abs(torch_tensor(c(-1, -2, 3)))
}

}
