% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_acos}
\alias{torch_acos}
\title{Acos}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Acos
}
\section{acos(input, out=None) -> Tensor }{


Returns a new tensor with the arccosine  of the elements of \code{input}.

\deqn{
    \text{out}_{i} = \cos^{-1}(\text{input}_{i})
}
}

\examples{

a = torch_randn(c(4))
a
torch_acos(a)
}
