% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_acos}
\alias{torch_acos}
\title{Acos}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
acos(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the arccosine  of the elements of :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \cos^{-1}(\text{input}}{i})
}
