% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_add}
\alias{torch_add}
\title{Add}
\arguments{
\item{value}{(Number) the number to be added to each element of \code{input}}

\item{input}{(Tensor) the first input tensor}

\item{alpha}{(Number) the scalar multiplier for \code{other}}

\item{other}{(Tensor) the second input tensor}
}
\description{
add(input, other, out=None)

add(input, alpha=1, other, out=None)
}
\details{
Adds the scalar \code{other} to each element of the input \code{input}
and returns a new resulting tensor.

\deqn{
    \text{out} = \text{input} + \text{other}
}
If \code{input} is of type FloatTensor or DoubleTensor, \code{other} must be
a real number, otherwise it should be an integer.

Each element of the tensor \code{other} is multiplied by the scalar
\code{alpha} and added to each element of the tensor \code{input}.
The resulting tensor is returned.

The shapes of \code{input} and \code{other} must be
broadcastable .

\deqn{
    \text{out} = \text{input} + \text{alpha} \times \text{other}
}
If \code{other} is of type FloatTensor or DoubleTensor, \code{alpha} must be
a real number, otherwise it should be an integer.
}
\examples{
\dontrun{
a = torch_randn(c(4))
a
torch_add(a, 20)
}

\dontrun{
a = torch_randn(c(4))
a
b = torch_randn(c(4, 1))
b
torch_add(a, 10, b)
}

}
