% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_addcdiv}
\alias{torch_addcdiv}
\title{Addcdiv}
\arguments{
\item{input}{(Tensor) the tensor to be added}

\item{value}{(Number, optional) multiplier for :math:\verb{\\text\{tensor1\} / \\text\{tensor2\}}}

\item{tensor1}{(Tensor) the numerator tensor}

\item{tensor2}{(Tensor) the denominator tensor}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
addcdiv(input, value=1, tensor1, tensor2, out=None) -> Tensor
}
\details{
Performs the element-wise division of :attr:\code{tensor1} by :attr:\code{tensor2},
multiply the result by the scalar :attr:\code{value} and add it to :attr:\code{input}.

.. math::
\text{out}_i = \text{input}_i + \text{value} \times \frac{\text{tensor1}_i}{\text{tensor2}_i}

The shapes of :attr:\code{input}, :attr:\code{tensor1}, and :attr:\code{tensor2} must be
:ref:\verb{broadcastable <broadcasting-semantics>}.

For inputs of type \code{FloatTensor} or \code{DoubleTensor}, :attr:\code{value} must be
a real number, otherwise an integer.
}
