% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_addcmul}
\alias{torch_addcmul}
\title{Addcmul}
\arguments{
\item{input}{(Tensor) the tensor to be added}

\item{value}{(Number, optional) multiplier for \eqn{tensor1 .* tensor2}}

\item{tensor1}{(Tensor) the tensor to be multiplied}

\item{tensor2}{(Tensor) the tensor to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Addcmul
}
\section{addcmul(input, value=1, tensor1, tensor2, out=None) -> Tensor }{


Performs the element-wise multiplication of \code{tensor1}
by \code{tensor2}, multiply the result by the scalar \code{value}
and add it to \code{input}.

\deqn{
    \text{out}_i = \text{input}_i + \text{value} \times \text{tensor1}_i \times \text{tensor2}_i
}
The shapes of \code{tensor}, \code{tensor1}, and \code{tensor2} must be
broadcastable .

For inputs of type \code{FloatTensor} or \code{DoubleTensor}, \code{value} must be
a real number, otherwise an integer.
}

\examples{

\dontrun{
t = torch_randn(c(1, 3))
t1 = torch_randn(c(3, 1))
t2 = torch_randn(c(1, 3))
torch_addcmul(t, 0.1, t1, t2)
}

}
