% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_addmv}
\alias{torch_addmv}
\title{Addmv}
\arguments{
\item{beta}{(Number, optional) multiplier for :attr:\code{input} (:math:\verb{\\beta})}

\item{input}{(Tensor) vector to be added}

\item{alpha}{(Number, optional) multiplier for :math:\code{mat @ vec} (:math:\verb{\\alpha})}

\item{mat}{(Tensor) matrix to be multiplied}

\item{vec}{(Tensor) vector to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Performs a matrix-vector product of the matrix \code{mat} and
the vector \code{vec}.
The vector \code{input} is added to the final result.
}
\details{
If \code{mat} is a :math:\verb{(n \\times m)} tensor, \code{vec} is a 1-D tensor of
size \code{m}, then \code{input} must be
:ref:\verb{broadcastable <broadcasting-semantics>} with a 1-D tensor of size \code{n} and
\code{out} will be 1-D tensor of size \code{n}.

\code{alpha} and \code{beta} are scaling factors on matrix-vector product between
\code{mat} and \code{vec} and the added tensor \code{input} respectively.

\deqn{
    \text{out} = \beta\ \text{input} + \alpha\ (\text{mat} \mathbin{@} \text{vec})
}
For inputs of type \code{FloatTensor} or \code{DoubleTensor}, arguments \code{beta} and
\code{alpha} must be real numbers, otherwise they should be integers
}
\section{Signatures}{


addmv(beta=1, input, alpha=1, mat, vec, out=None) -> Tensor
}

\examples{
\dontrun{
M = torch_randn(2)
mat = torch_randn(2, 3)
vec = torch_randn(3)
torch_addmv(M, mat, vec)
}

}
