% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_addr}
\alias{torch_addr}
\title{Addr}
\arguments{
\item{beta}{(Number, optional) multiplier for :attr:\code{input} (:math:\verb{\\beta})}

\item{input}{(Tensor) matrix to be added}

\item{alpha}{(Number, optional) multiplier for :math:\verb{\\text\{vec1\} \\otimes \\text\{vec2\}} (:math:\verb{\\alpha})}

\item{vec1}{(Tensor) the first vector of the outer product}

\item{vec2}{(Tensor) the second vector of the outer product}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
addr(beta=1, input, alpha=1, vec1, vec2, out=None) -> Tensor
}
\details{
Performs the outer-product of vectors :attr:\code{vec1} and :attr:\code{vec2}
and adds it to the matrix :attr:\code{input}.

Optional values :attr:\code{beta} and :attr:\code{alpha} are scaling factors on the
outer product between :attr:\code{vec1} and :attr:\code{vec2} and the added matrix
:attr:\code{input} respectively.

.. math::
\text{out} = \beta\ \text{input} + \alpha\ (\text{vec1} \otimes \text{vec2})

If :attr:\code{vec1} is a vector of size \code{n} and :attr:\code{vec2} is a vector
of size \code{m}, then :attr:\code{input} must be
:ref:\verb{broadcastable <broadcasting-semantics>} with a matrix of size
:math:\verb{(n \\times m)} and :attr:\code{out} will be a matrix of size
:math:\verb{(n \\times m)}.

For inputs of type \code{FloatTensor} or \code{DoubleTensor}, arguments :attr:\code{beta} and
:attr:\code{alpha} must be real numbers, otherwise they should be integers
}
\examples{
\dontrun{
vec1 = torch_arange(1., 4.)
vec2 = torch_arange(1., 3.)
M = torch_zeros(3, 2)
torch_addr(M, vec1, vec2)
}

}
