% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_allclose}
\alias{torch_allclose}
\title{Allclose}
\arguments{
\item{input}{(Tensor) first tensor to compare}

\item{other}{(Tensor) second tensor to compare}

\item{atol}{(float, optional) absolute tolerance. Default: 1e-08}

\item{rtol}{(float, optional) relative tolerance. Default: 1e-05}

\item{equal_nan}{(bool, optional) if \code{True}, then two \code{NaN} s will be compared as equal. Default: \code{False}}
}
\description{
This function checks if all \code{input} and \code{other} satisfy the condition:
}
\details{
\deqn{
    \lvert \text{input} - \text{other} \rvert \leq \texttt{atol} + \texttt{rtol} \times \lvert \text{other} \rvert
}
elementwise, for all elements of \code{input} and \code{other}. The behaviour of this function is analogous to
\verb{numpy.allclose <https://docs.scipy.org/doc/numpy/reference/generated/numpy.allclose.html>}_
}
\section{Signatures}{


allclose(input, other, rtol=1e-05, atol=1e-08, equal_nan=False) -> bool
}

\examples{
\dontrun{
torch_allclose(torch_tensor(c(10000., 1e-07)), torch_tensor(c(10000.1, 1e-08)))
torch_allclose(torch_tensor(c(10000., 1e-08)), torch_tensor(c(10000.1, 1e-09)))
torch_allclose(torch_tensor([1.0, float('nan')]), torch_tensor([1.0, float('nan')]))
torch_allclose(torch_tensor([1.0, float('nan')]), torch_tensor([1.0, float('nan')]), equal_nan=TRUE)
}

}
