% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_angle}
\alias{torch_angle}
\title{Angle}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
angle(input, out=None) -> Tensor
}
\details{
Computes the element-wise angle (in radians) of the given :attr:\code{input} tensor.

.. math::
\text{out}\emph{{i} = angle(\text{input}}{i})
}
