% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_arccosh}
\alias{torch_arccosh}
\title{Arccosh}
\usage{
torch_arccosh(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Arccosh
}
\section{arccosh(input, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_acosh]{torch_acosh()}}.
}

