% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_argmax}
\alias{torch_argmax}
\title{Argmax}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int) the dimension to reduce. If \code{None}, the argmax of the flattened input is returned.}

\item{keepdim}{(bool) whether the output tensor has \code{dim} retained or not. Ignored if \code{dim=None}.}
}
\description{
Returns the indices of the maximum value of all elements in the \code{input} tensor.

Returns the indices of the maximum values of a tensor across a dimension.
}
\details{
This is the second value returned by \code{torch_max}. See its
documentation for the exact semantics of this method.

This is the second value returned by \code{torch_max}. See its
documentation for the exact semantics of this method.
}
\section{Signatures}{


argmax(input) -> LongTensor



argmax(input, dim, keepdim=False) -> LongTensor
}

\examples{
\dontrun{
a = torch_randn(4, 4)
a
torch_argmax(a)
}

\dontrun{
a = torch_randn(4, 4)
a
torch_argmax(a, dim=1)
}

}
