% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_atan}
\alias{torch_atan}
\title{Atan}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
atan(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the arctangent  of the elements of :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \tan^{-1}(\text{input}}{i})
}
