% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_atan}
\alias{torch_atan}
\title{Atan}
\usage{
torch_atan(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Atan
}
\section{atan(input, out=NULL) -> Tensor }{


Returns a new tensor with the arctangent  of the elements of \code{input}.

\deqn{
    \mbox{out}_{i} = \tan^{-1}(\mbox{input}_{i})
}
}

\examples{

a = torch_randn(c(4))
a
torch_atan(a)
}
