% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_atan}
\alias{torch_atan}
\title{Atan}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Atan
}
\section{atan(input, out=None) -> Tensor }{


Returns a new tensor with the arctangent  of the elements of \code{input}.

\deqn{
    \text{out}_{i} = \tan^{-1}(\text{input}_{i})
}
}

\examples{

a = torch_randn(c(4))
a
torch_atan(a)
}
