% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_atan2}
\alias{torch_atan2}
\title{Atan2}
\arguments{
\item{input}{(Tensor) the first input tensor}

\item{other}{(Tensor) the second input tensor}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Atan2
}
\section{atan2(input, other, out=None) -> Tensor }{


Element-wise arctangent of \eqn{\text{input}_{i} / \text{other}_{i}}
with consideration of the quadrant. Returns a new tensor with the signed angles
in radians between vector \eqn{(\text{other}_{i}, \text{input}_{i})}
and vector \eqn{(1, 0)}. (Note that \eqn{\text{other}_{i}}, the second
parameter, is the x-coordinate, while \eqn{\text{input}_{i}}, the first
parameter, is the y-coordinate.)

The shapes of \code{input} and \code{other} must be
broadcastable .
}

\examples{

a = torch_randn(c(4))
a
torch_atan2(a, torch_randn(c(4)))
}
