% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_atleast_1d}
\alias{torch_atleast_1d}
\title{Atleast_1d}
\usage{
torch_atleast_1d(self, tensors)
}
\arguments{
\item{input}{(Tensor or list of Tensors)}
}
\description{
Returns a 1-dimensional view of each input tensor with zero dimensions.
Input tensors with one or more dimensions are returned as-is.
}
