% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/wrapers.R
\name{torch_atleast_2d}
\alias{torch_atleast_2d}
\title{Atleast_2d}
\usage{
torch_atleast_2d(self)
}
\arguments{
\item{self}{(Tensor or list of Tensors)}
}
\description{
Returns a 2-dimensional view of each each input tensor with zero dimensions.
Input tensors with two or more dimensions are returned as-is.
}
\examples{
if (torch_is_installed()) {

x <- torch_tensor(1.)
x
torch_atleast_2d(x)
x <- torch_randn(c(2,2))
x
torch_atleast_2d(x)
x <- torch_tensor(0.5)
y <- torch_tensor(1.)
torch_atleast_2d(list(x,y))
}
}
