% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_avg_pool1d}
\alias{torch_avg_pool1d}
\title{Avg_pool1d}
\usage{
torch_avg_pool1d(
  self,
  kernel_size,
  stride = list(),
  padding = 0L,
  ceil_mode = FALSE,
  count_include_pad = TRUE
)
}
\arguments{
\item{self}{input tensor of shape \eqn{(\mbox{minibatch} , \mbox{in\_channels} , iW)}}

\item{kernel_size}{the size of the window. Can be a single number or a tuple \verb{(kW,)}}

\item{stride}{the stride of the window. Can be a single number or a tuple \verb{(sW,)}. Default: \code{kernel_size}}

\item{padding}{implicit zero paddings on both sides of the input. Can be a single number or a tuple \verb{(padW,)}. Default: 0}

\item{ceil_mode}{when \code{TRUE}, will use \code{ceil} instead of \code{floor} to compute the output shape. Default: \code{FALSE}}

\item{count_include_pad}{when \code{TRUE}, will include the zero-padding in the averaging calculation. Default: \code{TRUE}}
}
\description{
Avg_pool1d
}
\section{avg_pool1d(input, kernel_size, stride=NULL, padding=0, ceil_mode=FALSE, count_include_pad=TRUE) -> Tensor }{


Applies a 1D average pooling over an input signal composed of several
input planes.

See \code{\link[=nn_avg_pool1d]{nn_avg_pool1d()}} for details and output shape.
}

