% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_baddbmm}
\alias{torch_baddbmm}
\title{Baddbmm}
\arguments{
\item{beta}{(Number, optional) multiplier for :attr:\code{input} (:math:\verb{\\beta})}

\item{input}{(Tensor) the tensor to be added}

\item{alpha}{(Number, optional) multiplier for :math:\verb{\\text\{batch1\} \\mathbin\{@\} \\text\{batch2\}} (:math:\verb{\\alpha})}

\item{batch1}{(Tensor) the first batch of matrices to be multiplied}

\item{batch2}{(Tensor) the second batch of matrices to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
baddbmm(beta=1, input, alpha=1, batch1, batch2, out=None) -> Tensor
}
\details{
Performs a batch matrix-matrix product of matrices in :attr:\code{batch1}
and :attr:\code{batch2}.
:attr:\code{input} is added to the final result.

:attr:\code{batch1} and :attr:\code{batch2} must be 3-D tensors each containing the same
number of matrices.

If :attr:\code{batch1} is a :math:\verb{(b \\times n \\times m)} tensor, :attr:\code{batch2} is a
:math:\verb{(b \\times m \\times p)} tensor, then :attr:\code{input} must be
:ref:\verb{broadcastable <broadcasting-semantics>} with a
:math:\verb{(b \\times n \\times p)} tensor and :attr:\code{out} will be a
:math:\verb{(b \\times n \\times p)} tensor. Both :attr:\code{alpha} and :attr:\code{beta} mean the
same as the scaling factors used in :meth:\code{torch.addbmm}.

.. math::
\text{out}_i = \beta\ \text{input}_i + \alpha\ (\text{batch1}_i \mathbin{@} \text{batch2}_i)

For inputs of type \code{FloatTensor} or \code{DoubleTensor}, arguments :attr:\code{beta} and
:attr:\code{alpha} must be real numbers, otherwise they should be integers.
}
