% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_bitwise_and}
\alias{torch_bitwise_and}
\title{Bitwise_and}
\arguments{
\item{input}{NA the first input tensor}

\item{other}{NA the second input tensor}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Bitwise_and
}
\section{bitwise_and(input, other, out=None) -> Tensor }{


Computes the bitwise AND of \code{input} and \code{other}. The input tensor must be of
integral or Boolean types. For bool tensors, it computes the logical AND.
}

