% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_bitwise_not}
\alias{torch_bitwise_not}
\title{Bitwise_not}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Bitwise_not
}
\section{bitwise_not(input, out=None) -> Tensor }{


Computes the bitwise NOT of the given input tensor. The input tensor must be of
integral or Boolean types. For bool tensors, it computes the logical NOT.
}

\examples{



}
