% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_bitwise_xor}
\alias{torch_bitwise_xor}
\title{Bitwise_xor}
\arguments{
\item{input}{NA the first input tensor}

\item{other}{NA the second input tensor}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Computes the bitwise XOR of \code{input} and \code{other}. The input tensor must be of
integral or Boolean types. For bool tensors, it computes the logical XOR.
}
\section{Signatures}{


bitwise_xor(input, other, out=None) -> Tensor
}

