% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_block_diag}
\alias{torch_block_diag}
\title{Block_diag}
\usage{
torch_block_diag(tensors)
}
\arguments{
\item{tensors}{(list of tensors) One or more tensors with 0, 1, or 2
dimensions.}
}
\description{
Create a block diagonal matrix from provided tensors.
}
\examples{
if (torch_is_installed()) {

A <- torch_tensor(rbind(c(0, 1), c(1, 0)))
B <- torch_tensor(rbind(c(3, 4, 5), c(6, 7, 8)))
C <- torch_tensor(7)
D <- torch_tensor(c(1, 2, 3))
E <- torch_tensor(rbind(4, 5, 6))
torch_block_diag(list(A, B, C, D, E))
}
}
