% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_bmm}
\alias{torch_bmm}
\title{Bmm}
\arguments{
\item{input}{(Tensor) the first batch of matrices to be multiplied}

\item{mat2}{(Tensor) the second batch of matrices to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
bmm(input, mat2, out=None) -> Tensor
}
\details{
Performs a batch matrix-matrix product of matrices stored in \code{input}
and \code{mat2}.

\code{input} and \code{mat2} must be 3-D tensors each containing
the same number of matrices.

If \code{input} is a :math:\verb{(b \\times n \\times m)} tensor, \code{mat2} is a
:math:\verb{(b \\times m \\times p)} tensor, \code{out} will be a
:math:\verb{(b \\times n \\times p)} tensor.

\deqn{
    \text{out}_i = \text{input}_i \mathbin{@} \text{mat2}_i
}
.. note:: This function does not :ref:\verb{broadcast <broadcasting-semantics>}.
For broadcasting matrix products, see \code{\link{torch_matmul}}.
}
\examples{
\dontrun{
input = torch_randn(10, 3, 4)
mat2 = torch_randn(10, 4, 5)
res = torch_bmm(input, mat2)
res.size()
}

}
