% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_broadcast_tensors}
\alias{torch_broadcast_tensors}
\title{Broadcast_tensors}
\arguments{
\item{*tensors}{NA any number of tensors of the same type}
}
\description{
Broadcasts the given tensors according to :ref:\code{broadcasting-semantics}.
}
\section{Signatures}{


broadcast_tensors(*tensors) -> List of Tensors
}

