% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_broadcast_tensors}
\alias{torch_broadcast_tensors}
\title{Broadcast_tensors}
\arguments{
\item{*tensors}{NA any number of tensors of the same type}
}
\description{
Broadcast_tensors
}
\section{broadcast_tensors(*tensors) -> List of Tensors }{


Broadcasts the given tensors according to broadcasting-semantics.
}

\examples{

\dontrun{
x = torch_arange(3)$view(1, 3)
y = torch_arange(2)$view(2, 1)
a, b = torch_broadcast_tensors(x, y)
a$size()
a
}

}
