% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_broadcast_tensors}
\alias{torch_broadcast_tensors}
\title{Broadcast_tensors}
\usage{
torch_broadcast_tensors(tensors)
}
\arguments{
\item{tensors}{a list containing any number of tensors of the same type}
}
\description{
Broadcast_tensors
}
\section{broadcast_tensors(tensors) -> List of Tensors }{


Broadcasts the given tensors according to broadcasting-semantics.
}

\examples{

x <- torch_arange(0, 3)$view(c(1, 4))
y <- torch_arange(0, 2)$view(c(3, 1))
out <- torch_broadcast_tensors(list(x, y))
out[[1]]
}
