% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_cartesian_prod}
\alias{torch_cartesian_prod}
\title{Cartesian_prod}
\arguments{
\item{*tensors}{NA any number of 1 dimensional tensors.}
}
\description{
Do cartesian product of the given sequence of tensors. The behavior is similar to
python's \code{itertools.product}.
}
\examples{
\dontrun{
a = [1, 2, 3]
b = [4, 5]
list(itertools.product(a, b))
tensor_a = torch_tensor(a)
tensor_b = torch_tensor(b)
torch_cartesian_prod(tensor_a, tensor_b)
}

}
