% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_cat}
\alias{torch_cat}
\title{Cat}
\arguments{
\item{tensors}{(sequence of Tensors) any python sequence of tensors of the same type.        Non-empty tensors provided must have the same shape, except in the        cat dimension.}

\item{dim}{(int, optional) the dimension over which the tensors are concatenated}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Cat
}
\section{cat(tensors, dim=0, out=None) -> Tensor }{


Concatenates the given sequence of \code{seq} tensors in the given dimension.
All tensors must either have the same shape (except in the concatenating
dimension) or be empty.

\code{\link{torch_cat}} can be seen as an inverse operation for \code{\link{torch.split}}
and \code{\link{torch_chunk}}.

\code{\link{torch_cat}} can be best understood via examples.
}

\examples{

\dontrun{
x = torch_randn(c(2, 3))
x
torch_cat(list(x, x, x), 0)
torch_cat(list(x, x, x), 1)
}

}
