% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_cdist}
\alias{torch_cdist}
\title{Cdist}
\arguments{
\item{x1}{(Tensor) input tensor of shape \eqn{B \times P \times M}.}

\item{x2}{(Tensor) input tensor of shape \eqn{B \times R \times M}.}

\item{p}{NA p value for the p-norm distance to calculate between each vector pair        \eqn{\in [0, \infty]}.}

\item{compute_mode}{NA 'use_mm_for_euclid_dist_if_necessary' - will use matrix multiplication approach to calculate        euclidean distance (p = 2) if P > 25 or R > 25        'use_mm_for_euclid_dist' - will always use matrix multiplication approach to calculate        euclidean distance (p = 2)        'donot_use_mm_for_euclid_dist' - will never use matrix multiplication approach to calculate        euclidean distance (p = 2)        Default: use_mm_for_euclid_dist_if_necessary.}
}
\description{
Cdist
}
\section{TEST }{


Computes batched the p-norm distance between each pair of the two collections of row vectors.
}

