% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_ceil}
\alias{torch_ceil}
\title{Ceil}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the ceil of the elements of \code{input},
the smallest integer greater than or equal to each element.
}
\details{
\deqn{
    \text{out}_{i} = \left\lceil \text{input}_{i} \right\rceil = \left\lfloor \text{input}_{i} \right\rfloor + 1
}
}
\section{Signatures}{


ceil(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(4)
a
torch_ceil(a)
}

}
