% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_celu_}
\alias{torch_celu_}
\title{Celu_}
\usage{
torch_celu_(self, alpha = 1L)
}
\arguments{
\item{self}{the input tensor}

\item{alpha}{the alpha value for the CELU formulation. Default: 1.0}
}
\description{
Celu_
}
\section{celu_(input, alpha=1.) -> Tensor }{


In-place version of \code{\link[=torch_celu]{torch_celu()}}.
}

