% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_chunk}
\alias{torch_chunk}
\title{Chunk}
\arguments{
\item{input}{(Tensor) the tensor to split}

\item{chunks}{(int) number of chunks to return}

\item{dim}{(int) dimension along which to split the tensor}
}
\description{
chunk(input, chunks, dim=0) -> List of Tensors
}
\details{
Splits a tensor into a specific number of chunks.

Last chunk will be smaller if the tensor size along the given dimension
:attr:\code{dim} is not divisible by :attr:\code{chunks}.
}
