% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_clamp}
\alias{torch_clamp}
\title{Clamp}
\arguments{
\item{min}{(Number) lower-bound of the range to be clamped to}

\item{max}{(Number) upper-bound of the range to be clamped to}

\item{input}{(Tensor) the input tensor.}

\item{value}{(Number) maximal value of each element in the output}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
clamp(input, min, max, out=None) -> Tensor

clamp(input, *, min, out=None) -> Tensor

clamp(input, *, max, out=None) -> Tensor
}
\details{
Clamp all elements in :attr:\code{input} into the range \code{[} :attr:\code{min}, :attr:\code{max} \verb{]} and return
a resulting tensor:

.. math::
y_i = \begin{cases}
\text{min} & \text{if } x_i < \text{min} \\
x_i & \text{if } \text{min} \leq x_i \leq \text{max} \\
\text{max} & \text{if } x_i > \text{max}
\end{cases}

If :attr:\code{input} is of type \code{FloatTensor} or \code{DoubleTensor}, args :attr:\code{min}
and :attr:\code{max} must be real numbers, otherwise they should be integers.

Clamps all elements in :attr:\code{input} to be larger or equal :attr:\code{min}.

If :attr:\code{input} is of type \code{FloatTensor} or \code{DoubleTensor}, :attr:\code{value}
should be a real number, otherwise it should be an integer.

Clamps all elements in :attr:\code{input} to be smaller or equal :attr:\code{max}.

If :attr:\code{input} is of type \code{FloatTensor} or \code{DoubleTensor}, :attr:\code{value}
should be a real number, otherwise it should be an integer.
}
\examples{
\dontrun{
a = torch_randn(4)
a
torch_clamp(a, min=-0.5, max=0.5)
}

\dontrun{
a = torch_randn(4)
a
torch_clamp(a, min=0.5)
}

\dontrun{
a = torch_randn(4)
a
torch_clamp(a, max=0.5)
}

}
