% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_clone}
\alias{torch_clone}
\title{Clone}
\usage{
torch_clone(self, memory_format = NULL)
}
\arguments{
\item{input}{(Tensor) the input tensor.}
}
\description{
Clone
}
\note{
This function is differentiable, so gradients will flow back from the
result of this operation to \code{input}. To create a tensor without an
autograd relationship to \code{input} see \code{Tensor$detach}.
}
\section{clone(input, *, memory_format=torch.preserve_format) -> Tensor }{


Returns a copy of \code{input}.
}

