% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_combinations}
\alias{torch_combinations}
\title{Combinations}
\arguments{
\item{input}{(Tensor) 1D vector.}

\item{r}{(int, optional) number of elements to combine}

\item{with_replacement}{(boolean, optional) whether to allow duplication in combination}
}
\description{
Compute combinations of length \eqn{r} of the given tensor. The behavior is similar to
python's \code{itertools.combinations} when \code{with_replacement} is set to \code{False}, and
\code{itertools.combinations_with_replacement} when \code{with_replacement} is set to \code{True}.
}
\section{Signatures}{


combinations(input, r=2, with_replacement=False) -> seq
}

\examples{
\dontrun{
a = c(1, 2, 3)
list(itertools$combinations(a, r=2))
list(itertools$combinations(a, r=3))
list(itertools$combinations_with_replacement(a, r=2))
tensor_a = torch_tensor(a)
torch_combinations(tensor_a)
torch_combinations(tensor_a, r=3)
torch_combinations(tensor_a, with_replacement=TRUE)
}

}
