% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_conj}
\alias{torch_conj}
\title{Conj}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Conj
}
\section{conj(input, out=None) -> Tensor }{


Computes the element-wise conjugate of the given \code{input} tensor.

\deqn{
    \mbox{out}_{i} = conj(\mbox{input}_{i})
}
}

\examples{
\dontrun{
torch_conj(torch_tensor(c(-1 + 1i, -2 + 2i, 3 - 3i)))
}
}
