% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_conj}
\alias{torch_conj}
\title{Conj}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
conj(input, out=None) -> Tensor
}
\details{
Computes the element-wise conjugate of the given :attr:\code{input} tensor.

.. math::
\text{out}\emph{{i} = conj(\text{input}}{i})
}
