% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_conv_tbc}
\alias{torch_conv_tbc}
\title{Conv_tbc}
\usage{
torch_conv_tbc(self, weight, bias, pad = 0L)
}
\arguments{
\item{self}{NA input tensor of shape \eqn{(\mbox{sequence length} \times batch \times \mbox{in\_channels})}}

\item{weight}{NA filter of shape (\eqn{\mbox{kernel width} \times \mbox{in\_channels} \times \mbox{out\_channels}})}

\item{bias}{NA bias of shape (\eqn{\mbox{out\_channels}})}

\item{pad}{NA number of timesteps to pad. Default: 0}
}
\description{
Conv_tbc
}
\section{TEST }{


Applies a 1-dimensional sequence convolution over an input sequence.
Input and output dimensions are (Time, Batch, Channels) - hence TBC.
}

