% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_cos}
\alias{torch_cos}
\title{Cos}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Cos
}
\section{cos(input, out=None) -> Tensor }{


Returns a new tensor with the cosine  of the elements of \code{input}.

\deqn{
    \mbox{out}_{i} = \cos(\mbox{input}_{i})
}
}

\examples{
\dontrun{

a = torch_randn(c(4))
a
torch_cos(a)
}
}
