% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_cos}
\alias{torch_cos}
\title{Cos}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
cos(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the cosine  of the elements of :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \cos(\text{input}}{i})
}
\examples{
\dontrun{
a = torch_randn(4)
a
torch_cos(a)
}

}
