% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_cosh}
\alias{torch_cosh}
\title{Cosh}
\usage{
torch_cosh(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Cosh
}
\section{cosh(input, out=None) -> Tensor }{


Returns a new tensor with the hyperbolic cosine  of the elements of
\code{input}.

\deqn{
    \mbox{out}_{i} = \cosh(\mbox{input}_{i})
}
}

\examples{
if (torch_is_installed()) {

a = torch_randn(c(4))
a
torch_cosh(a)
}
}
