% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_cumprod}
\alias{torch_cumprod}
\title{Cumprod}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int) the dimension to do the operation over}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{If}{NA attr:\code{dtype} before the operation}

\item{is}{NA None.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
cumprod(input, dim, out=None, dtype=None) -> Tensor
}
\details{
Returns the cumulative product of elements of :attr:\code{input} in the dimension
:attr:\code{dim}.

For example, if :attr:\code{input} is a vector of size N, the result will also be
a vector of size N, with elements.

.. math::
y_i = x_1 \times x_2\times x_3\times \dots \times x_i
}
