% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_cumsum}
\alias{torch_cumsum}
\title{Cumsum}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int) the dimension to do the operation over}

\item{dtype}{(\code{torch.dtype}, optional) the desired data type of returned tensor.        If specified, the input tensor is casted to \code{dtype} before the operation        is performed. This is useful for preventing data type overflows. Default: None.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns the cumulative sum of elements of \code{input} in the dimension
\code{dim}.
}
\details{
For example, if \code{input} is a vector of size N, the result will also be
a vector of size N, with elements.

\deqn{
    y_i = x_1 + x_2 + x_3 + \dots + x_i
}
}
\section{Signatures}{


cumsum(input, dim, out=None, dtype=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(c(10))
a
torch_cumsum(a, dim=0)
}

}
