% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_deg2rad}
\alias{torch_deg2rad}
\title{Deg2rad}
\usage{
torch_deg2rad(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Deg2rad
}
\section{deg2rad(input, *, out=None) -> Tensor }{


Returns a new tensor with each of the elements of \code{input}
converted from angles in degrees to radians.
}

\examples{
if (torch_is_installed()) {

a <- torch_tensor(rbind(c(180.0, -180.0), c(360.0, -360.0), c(90.0, -90.0)))
torch_deg2rad(a)
}
}
