% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/wrapers.R
\name{torch_dequantize}
\alias{torch_dequantize}
\title{Dequantize}
\usage{
torch_dequantize(tensor)
}
\arguments{
\item{tensor}{(Tensor) A quantized Tensor or a list oof quantized tensors}
}
\description{
Dequantize
}
\section{dequantize(tensor) -> Tensor }{


Returns an fp32 Tensor by dequantizing a quantized Tensor
}

\section{dequantize(tensors) -> sequence of Tensors }{


Given a list of quantized Tensors, dequantize them and return a list of fp32 Tensors
}

