% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_det}
\alias{torch_det}
\title{Det}
\arguments{
\item{input}{(Tensor) the input tensor of size \verb{(*, n, n)} where \code{*} is zero or more                batch dimensions.}
}
\description{
det(input) -> Tensor
}
\details{
Calculates determinant of a square matrix or batches of square matrices.

.. note::
Backward through :meth:\code{det} internally uses SVD results when :attr:\code{input} is
not invertible. In this case, double backward through :meth:\code{det} will be
unstable in when :attr:\code{input} doesn't have distinct singular values. See
:meth:\code{~torch.svd} for details.
}
