% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_digamma}
\alias{torch_digamma}
\title{Digamma}
\arguments{
\item{input}{(Tensor) the tensor to compute the digamma function on}
}
\description{
Digamma
}
\section{digamma(input, out=None) -> Tensor }{


Computes the logarithmic derivative of the gamma function on \code{input}.

\deqn{
    \psi(x) = \frac{d}{dx} \ln\left(\Gamma\left(x\right)\right) = \frac{\Gamma'(x)}{\Gamma(x)}
}
}

\examples{

\dontrun{
a = torch_tensor(c(1, 0.5))
torch_digamma(a)
}

}
