% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_dist}
\alias{torch_dist}
\title{Dist}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{other}{(Tensor) the Right-hand-side input tensor}

\item{p}{(float, optional) the norm to be computed}
}
\description{
dist(input, other, p=2) -> Tensor
}
\details{
Returns the p-norm of (\code{input} - \code{other})

The shapes of \code{input} and \code{other} must be
:ref:\verb{broadcastable <broadcasting-semantics>}.
}
\examples{
\dontrun{
x = torch_randn(4)
x
y = torch_randn(4)
y
torch_dist(x, y, 3.5)
torch_dist(x, y, 3)
torch_dist(x, y, 0)
torch_dist(x, y, 1)
}

}
