% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_dist}
\alias{torch_dist}
\title{Dist}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{other}{(Tensor) the Right-hand-side input tensor}

\item{p}{(float, optional) the norm to be computed}
}
\description{
dist(input, other, p=2) -> Tensor
}
\details{
Returns the p-norm of (:attr:\code{input} - :attr:\code{other})

The shapes of :attr:\code{input} and :attr:\code{other} must be
:ref:\verb{broadcastable <broadcasting-semantics>}.
}
