% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-new-examples.R, R/gen-namespace.R
\name{torch_divide}
\alias{torch_divide}
\title{Divide}
\usage{
torch_divide(self, other)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{other}{(Number) the number to be divided to each element of \code{input}}
}
\description{
Divide
}
\section{divide(input, other, *, out=None) -> Tensor }{


Alias for \code{\link[=torch_div]{torch_div()}}.
}

