% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_dot}
\alias{torch_dot}
\title{Dot}
\description{
Dot
}
\note{
This function does not broadcast .
}
\section{dot(input, tensor) -> Tensor }{


Computes the dot product (inner product) of two tensors.
}

\examples{
if (torch_is_installed()) {

torch_dot(torch_tensor(c(2, 3)), torch_tensor(c(2, 1)))
}
}
