% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_dot}
\alias{torch_dot}
\title{Dot}
\description{
Computes the dot product (inner product) of two tensors.
}
\details{
.. note:: This function does not :ref:\verb{broadcast <broadcasting-semantics>}.

Example::
}
\section{Signatures}{


dot(input, tensor) -> Tensor
}

\examples{
\dontrun{
torch_dot(torch_tensor([2, 3]), torch_tensor([2, 1]))
}

}
