% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_eig}
\alias{torch_eig}
\title{Eig}
\arguments{
\item{input}{(Tensor) the square matrix of shape \eqn{(n \times n)} for which the eigenvalues and eigenvectors        will be computed}

\item{eigenvectors}{(bool) \code{True} to compute both eigenvalues and eigenvectors;        otherwise, only eigenvalues will be computed}

\item{out}{(tuple, optional) the output tensors}
}
\description{
Computes the eigenvalues and eigenvectors of a real square matrix.
}
\note{
\preformatted{Since eigenvalues and eigenvectors might be complex, backward pass is supported only
for :func:`torch_symeig`
}
}
\section{Signatures}{


eig(input, eigenvectors=False, out=None) -> (Tensor, Tensor)
}

