% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_eig}
\alias{torch_eig}
\title{Eig}
\arguments{
\item{input}{(Tensor) the square matrix of shape :math:\verb{(n \\times n)} for which the eigenvalues and eigenvectors}

\item{will}{NA}

\item{eigenvectors}{(bool) \code{True} to compute both eigenvalues and eigenvectors;}

\item{otherwise, }{NA}

\item{out}{(tuple, optional) the output tensors}
}
\description{
eig(input, eigenvectors=False, out=None) -> (Tensor, Tensor)
}
\details{
Computes the eigenvalues and eigenvectors of a real square matrix.

.. note::
Since eigenvalues and eigenvectors might be complex, backward pass is supported only
for :func:\code{torch.symeig}
}
