% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_empty}
\alias{torch_empty}
\title{Empty}
\arguments{
\item{size}{(int...) a sequence of integers defining the shape of the output tensor.}

\item{Can}{NA}

\item{out}{(Tensor, optional) the output tensor.}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{layout}{NA class:\code{torch.layout}, optional): the desired layout of returned Tensor.}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if \code{None}, uses the current device for the default tensor type}

\item{NA}{func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU}

\item{for}{NA}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}

\item{pin_memory}{(bool, optional) If set, returned tensor would be allocated in}

\item{the}{NA \code{False}.}
}
\description{
empty(*size, out=None, dtype=None, layout=torch.strided, device=None, requires_grad=False, pin_memory=False) -> Tensor
}
\details{
Returns a tensor filled with uninitialized data. The shape of the tensor is
defined by the variable argument :attr:\code{size}.
}
