% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_empty_like}
\alias{torch_empty_like}
\title{Empty_like}
\arguments{
\item{input}{(Tensor) the size of \code{input} will determine size of the output tensor.}

\item{dtype}{(\code{torch.dtype}, optional) the desired data type of returned Tensor.        Default: if \code{None}, defaults to the dtype of \code{input}.}

\item{layout}{(\code{torch.layout}, optional) the desired layout of returned tensor.        Default: if \code{None}, defaults to the layout of \code{input}.}

\item{device}{(\code{torch.device}, optional) the desired device of returned tensor.        Default: if \code{None}, defaults to the device of \code{input}.}

\item{requires_grad}{(bool, optional) If autograd should record operations on the        returned tensor. Default: \code{False}.}
}
\description{
Empty_like
}
\section{empty_like(input, dtype=None, layout=None, device=None, requires_grad=False) -> Tensor }{


Returns an uninitialized tensor with the same size as \code{input}.
\code{torch_empty_like(input)} is equivalent to
\code{torch_empty(input.size(), dtype=input.dtype, layout=input.layout, device=input.device)}.
}

\examples{

\dontrun{
torch_empty(list(2,3), dtype=torch_int64())
}

}
