% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_equal}
\alias{torch_equal}
\title{Equal}
\description{
\code{True} if two tensors have the same size and elements, \code{False} otherwise.
}
\details{
Example::
}
\section{Signatures}{


equal(input, other) -> bool
}

\examples{
\dontrun{
torch_equal(torch_tensor([1, 2]), torch_tensor([1, 2]))
}

}
