% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_equal}
\alias{torch_equal}
\title{Equal}
\usage{
torch_equal(self, other)
}
\arguments{
\item{self}{the input tensor}

\item{other}{the other input tensor}
}
\description{
Equal
}
\section{equal(input, other) -> bool }{


\code{TRUE} if two tensors have the same size and elements, \code{FALSE} otherwise.
}

\examples{
if (torch_is_installed()) {

torch_equal(torch_tensor(c(1, 2)), torch_tensor(c(1, 2)))
}
}
