% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_equal}
\alias{torch_equal}
\title{Equal}
\description{
equal(input, other) -> bool
}
\details{
\code{True} if two tensors have the same size and elements, \code{False} otherwise.

Example::
}
