% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_equal}
\alias{torch_equal}
\title{Equal}
\description{
Equal
}
\section{equal(input, other) -> bool }{


\code{True} if two tensors have the same size and elements, \code{False} otherwise.
}

\examples{
if (torch_is_installed()) {

torch_equal(torch_tensor(c(1, 2)), torch_tensor(c(1, 2)))
}
}
