% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_erfc}
\alias{torch_erfc}
\title{Erfc}
\usage{
torch_erfc(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Erfc
}
\section{erfc(input, out=None) -> Tensor }{


Computes the complementary error function of each element of \code{input}.
The complementary error function is defined as follows:

\deqn{
    \mathrm{erfc}(x) = 1 - \frac{2}{\sqrt{\pi}} \int_{0}^{x} e^{-t^2} dt
}
}

\examples{
if (torch_is_installed()) {

torch_erfc(torch_tensor(c(0, -1., 10.)))
}
}
