% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_exp}
\alias{torch_exp}
\title{Exp}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
exp(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the exponential of the elements
of the input tensor :attr:\code{input}.

.. math::
y_{i} = e^{x_{i}}
}
\examples{
\dontrun{
torch_exp(torch_tensor([0, math.log(2.)]))
}

}
