% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_exp}
\alias{torch_exp}
\title{Exp}
\usage{
torch_exp(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Exp
}
\section{exp(input, out=None) -> Tensor }{


Returns a new tensor with the exponential of the elements
of the input tensor \code{input}.

\deqn{
    y_{i} = e^{x_{i}}
}
}

\examples{
if (torch_is_installed()) {

torch_exp(torch_tensor(c(0, log(2))))
}
}
