% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_exp2}
\alias{torch_exp2}
\title{Exp2}
\usage{
torch_exp2(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Exp2
}
\section{exp2(input, *, out=None) -> Tensor }{


Computes the base two exponential function of \code{input}.

\deqn{
    y_{i} = 2^{x_{i}}
}
}

\examples{

torch_exp2(torch_tensor(c(0, log2(2.), 3, 4)))
}
