% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_expm1}
\alias{torch_expm1}
\title{Expm1}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
expm1(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the exponential of the elements minus 1
of :attr:\code{input}.

.. math::
y_{i} = e^{x_{i}} - 1
}
\examples{
\dontrun{
torch_expm1(torch_tensor([0, math.log(2.)]))
}

}
