% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_expm1}
\alias{torch_expm1}
\title{Expm1}
\usage{
torch_expm1(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Expm1
}
\section{expm1(input, out=None) -> Tensor }{


Returns a new tensor with the exponential of the elements minus 1
of \code{input}.

\deqn{
    y_{i} = e^{x_{i}} - 1
}
}

\examples{
if (torch_is_installed()) {

torch_expm1(torch_tensor(c(0, log(2))))
}
}
