% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_expm1}
\alias{torch_expm1}
\title{Expm1}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Expm1
}
\section{expm1(input, out=None) -> Tensor }{


Returns a new tensor with the exponential of the elements minus 1
of \code{input}.

\deqn{
    y_{i} = e^{x_{i}} - 1
}
}

\examples{

\dontrun{
torch_expm1list(torch_tensor([0, math$log(2.)]))
}

}
