% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_flip}
\alias{torch_flip}
\title{Flip}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dims}{(a list or tuple) axis to flip on}
}
\description{
Reverse the order of a n-D tensor along given axis in dims.
}
\section{Signatures}{


flip(input, dims) -> Tensor
}

\examples{
\dontrun{
x = torch_arange(8).view(2, 2, 2)
x
torch_flip(x, [0, 1])
}

}
