% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_flip}
\alias{torch_flip}
\title{Flip}
\usage{
torch_flip(self, dims)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{dims}{(a list or tuple) axis to flip on}
}
\description{
Flip
}
\section{flip(input, dims) -> Tensor }{


Reverse the order of a n-D tensor along given axis in dims.
}

\examples{
if (torch_is_installed()) {

x = torch_arange(0, 8)$view(c(2, 2, 2))
x
torch_flip(x, c(1, 2))
}
}
