% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_flipud}
\alias{torch_flipud}
\title{Flipud}
\usage{
torch_flipud(self)
}
\arguments{
\item{input}{(Tensor) Must be at least 1-dimensional.}
}
\description{
Flipud
}
\note{
Equivalent to \code{input[-1,]}. Requires the array to be at least 1-D.
}
\section{flipud(input) -> Tensor }{


Flip array in the up/down direction, returning a new tensor.

Flip the entries in each column in the up/down direction.
Rows are preserved, but appear in a different order than before.
}

