% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_floor}
\alias{torch_floor}
\title{Floor}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the floor of the elements of \code{input},
the largest integer less than or equal to each element.
}
\details{
\deqn{
    \text{out}_{i} = \left\lfloor \text{input}_{i} \right\rfloor
}
}
\section{Signatures}{


floor(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(4)
a
torch_floor(a)
}

}
