% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_floor}
\alias{torch_floor}
\title{Floor}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
floor(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the floor of the elements of :attr:\code{input},
the largest integer less than or equal to each element.

.. math::
\text{out}\emph{{i} = \left\lfloor \text{input}}{i} \right\rfloor
}
\examples{
\dontrun{
a = torch_randn(4)
a
torch_floor(a)
}

}
