% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_floor}
\alias{torch_floor}
\title{Floor}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Floor
}
\section{floor(input, out=None) -> Tensor }{


Returns a new tensor with the floor of the elements of \code{input},
the largest integer less than or equal to each element.

\deqn{
    \mbox{out}_{i} = \left\lfloor \mbox{input}_{i} \right\rfloor
}
}

\examples{
if (torch_is_installed()) {

a = torch_randn(c(4))
a
torch_floor(a)
}
}
