% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_frac}
\alias{torch_frac}
\title{Frac}
\description{
frac(input, out=None) -> Tensor
}
\details{
Computes the fractional portion of each element in :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \text{input}}{i} - \left\lfloor |\text{input}\emph{{i}| \right\rfloor * \operatorname{sgn}(\text{input}}{i})

Example::
}
