% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_frac}
\alias{torch_frac}
\title{Frac}
\description{
Frac
}
\section{frac(input, out=None) -> Tensor }{


Computes the fractional portion of each element in \code{input}.

\deqn{
    \mbox{out}_{i} = \mbox{input}_{i} - \left\lfloor |\mbox{input}_{i}| \right\rfloor * \mbox{sgn}(\mbox{input}_{i})
}
}

\examples{

torch_frac(torch_tensor(c(1, 2.5, -3.2)))
}
