% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_full_like}
\alias{torch_full_like}
\title{Full_like}
\arguments{
\item{input}{(Tensor) the size of :attr:\code{input} will determine size of the output tensor.}

\item{fill_value}{NA the number to fill the output tensor with.}

\item{dtype}{(:class:\code{torch.dtype}, optional) the desired data type of returned Tensor.        Default: if \code{None}, defaults to the dtype of :attr:\code{input}.}

\item{layout}{(:class:\code{torch.layout}, optional) the desired layout of returned tensor.        Default: if \code{None}, defaults to the layout of :attr:\code{input}.}

\item{device}{(:class:\code{torch.device}, optional) the desired device of returned tensor.        Default: if \code{None}, defaults to the device of :attr:\code{input}.}

\item{requires_grad}{(bool, optional) If autograd should record operations on the        returned tensor. Default: \code{False}.}
}
\description{
full_like(input, fill_value, out=None, dtype=None, layout=torch.strided, device=None, requires_grad=False) -> Tensor
}
\details{
Returns a tensor with the same size as :attr:\code{input} filled with :attr:\code{fill_value}.
\code{torch.full_like(input, fill_value)} is equivalent to
\code{torch.full(input.size(), fill_value, dtype=input.dtype, layout=input.layout, device=input.device)}.
}
