% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_gather}
\alias{torch_gather}
\title{Gather}
\arguments{
\item{input}{(Tensor) the source tensor}

\item{dim}{(int) the axis along which to index}

\item{index}{(LongTensor) the indices of elements to gather}

\item{out}{(Tensor, optional) the destination tensor}

\item{sparse_grad}{(bool,optional) If \code{True}, gradient w.r.t. :attr:\code{input} will be a sparse tensor.}
}
\description{
gather(input, dim, index, out=None, sparse_grad=False) -> Tensor
}
\details{
Gathers values along an axis specified by \code{dim}.

For a 3-D tensor the output is specified by::\preformatted{out[i][j][k] = input[index[i][j][k]][j][k]  # if dim == 0
out[i][j][k] = input[i][index[i][j][k]][k]  # if dim == 1
out[i][j][k] = input[i][j][index[i][j][k]]  # if dim == 2
}

If :attr:\code{input} is an n-dimensional tensor with size
:math:\verb{(x_0, x_1..., x_\{i-1\}, x_i, x_\{i+1\}, ..., x_\{n-1\})}
and \code{dim = i}, then :attr:\code{index} must be an :math:\code{n}-dimensional tensor with
size :math:\verb{(x_0, x_1, ..., x_\{i-1\}, y, x_\{i+1\}, ..., x_\{n-1\})} where :math:\verb{y \\geq 1}
and :attr:\code{out} will have the same size as :attr:\code{index}.
}
