% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_geqrf}
\alias{torch_geqrf}
\title{Geqrf}
\arguments{
\item{input}{(Tensor) the input matrix}

\item{out}{(tuple, optional) the output tuple of (Tensor, Tensor)}
}
\description{
Geqrf
}
\section{geqrf(input, out=None) -> (Tensor, Tensor) }{


This is a low-level function for calling LAPACK directly. This function
returns a namedtuple (a, tau) as defined in \verb{LAPACK documentation for geqrf}_ .

You'll generally want to use \code{\link{torch_qr}} instead.

Computes a QR decomposition of \code{input}, but without constructing
\eqn{Q} and \eqn{R} as explicit separate matrices.

Rather, this directly calls the underlying LAPACK function \code{?geqrf}
which produces a sequence of 'elementary reflectors'.

See \verb{LAPACK documentation for geqrf}_ for further details.
}

\examples{


}
