% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_imag}
\alias{torch_imag}
\title{Imag}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
imag(input, out=None) -> Tensor
}
\details{
Computes the element-wise imag value of the given :attr:\code{input} tensor.

.. math::
\text{out}\emph{{i} = imag(\text{input}}{i})
}
\examples{
\dontrun{
torch_imag(torch_tensor([-1 + 1j, -2 + 2j, 3 - 3j]))
}

}
