% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_index_select}
\alias{torch_index_select}
\title{Index_select}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int) the dimension in which we index}

\item{index}{(LongTensor) the 1-D tensor containing the indices to index}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
index_select(input, dim, index, out=None) -> Tensor
}
\details{
Returns a new tensor which indexes the :attr:\code{input} tensor along dimension
:attr:\code{dim} using the entries in :attr:\code{index} which is a \code{LongTensor}.

The returned tensor has the same number of dimensions as the original tensor
(:attr:\code{input}).  The :attr:\code{dim}\ th dimension has the same size as the length
of :attr:\code{index}; other dimensions have the same size as in the original tensor.

.. note:: The returned tensor does \strong{not} use the same storage as the original
tensor.  If :attr:\code{out} has a different shape than expected, we
silently change it to the correct shape, reallocating the underlying
storage if necessary.
}
