% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{torch_install_path}
\alias{torch_install_path}
\title{A simple exported version of install_path
Returns the torch installation path.}
\usage{
torch_install_path()
}
\description{
A simple exported version of install_path
Returns the torch installation path.
}
