% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_inverse}
\alias{torch_inverse}
\title{Inverse}
\arguments{
\item{input}{(Tensor) the input tensor of size \eqn{(*, n, n)} where \code{*} is zero or more                    batch dimensions}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Inverse
}
\note{
\preformatted{Irrespective of the original strides, the returned tensors will be
transposed, i.e. with strides like `input.contiguous().transpose(-2, -1).stride()`
}
}
\section{inverse(input, out=None) -> Tensor }{


Takes the inverse of the square matrix \code{input}. \code{input} can be batches
of 2D square tensors, in which case this function would return a tensor composed of
individual inverses.
}

\examples{

\dontrun{
x = torch_rand(4, 4)
y = torch_inverse(x)
z = torch_mm(x, y)
z
torch_max(torch_abs(z - torch_eye(4))) # Max non-zero
# Batched inverse example
x = torch_randn(c(2, 3, 4, 4))
y = torch_inverse(x)
z = torch_matmul(x, y)
torch_max(torch_abs(z - torch_eye(4)$expand_as(x))) # Max non-zero
}

}
