% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_is_complex}
\alias{torch_is_complex}
\title{Is_complex}
\usage{
torch_is_complex(self)
}
\arguments{
\item{self}{(Tensor) the PyTorch tensor to test}
}
\description{
Is_complex
}
\section{is_complex(input) -> (bool) }{


Returns TRUE if the data type of \code{input} is a complex data type i.e.,
one of \code{torch_complex64}, and \code{torch.complex128}.
}

