% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_is_floating_point}
\alias{torch_is_floating_point}
\title{Is_floating_point}
\arguments{
\item{input}{(Tensor) the PyTorch tensor to test}
}
\description{
is_floating_point(input) -> (bool)
}
\details{
Returns True if the data type of :attr:\code{input} is a floating point data type i.e.,
one of \code{torch.float64}, \code{torch.float32} and \code{torch.float16}.
}
