% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_isfinite}
\alias{torch_isfinite}
\title{Isfinite}
\arguments{
\item{tensor}{(Tensor) A tensor to check}
}
\description{
Returns a new tensor with boolean elements representing if each element is \code{Finite} or not.
}
\examples{
\dontrun{
torch_isfinite(torch_tensor([1, float('inf'), 2, float('-inf'), float('nan')]))
}

}
