% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_isfinite}
\alias{torch_isfinite}
\title{Isfinite}
\arguments{
\item{tensor}{(Tensor) A tensor to check}
}
\description{
Isfinite
}
\section{TEST }{


Returns a new tensor with boolean elements representing if each element is \code{Finite} or not.
}

\examples{

torch_isfinite(torch_tensor(c(1, Inf, 2, -Inf, NaN)))
}
