% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_isfinite}
\alias{torch_isfinite}
\title{Isfinite}
\description{
Returns a new tensor with boolean elements representing if each element is \code{Finite} or not.
}
\details{
\preformatted{Arguments:
    tensor (Tensor): A tensor to check

Returns:
}
}
