% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_isfinite}
\alias{torch_isfinite}
\title{Isfinite}
\description{
Returns a new tensor with boolean elements representing if each element is \code{Finite} or not.
}
\details{
\preformatted{Arguments:
    tensor (Tensor): A tensor to check

Returns:
    Tensor: ``A torch.Tensor with dtype torch.bool`` containing a True at each location of finite elements and False otherwise

Example::

    >>> torch.isfinite(torch.tensor([1, float('inf'), 2, float('-inf'), float('nan')]))
    tensor([True,  False,  True,  False,  False])
}
}
