% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_isnan}
\alias{torch_isnan}
\title{Isnan}
\arguments{
\item{input}{(Tensor) A tensor to check}
}
\description{
Returns a new tensor with boolean elements representing if each element is \code{NaN} or not.
}
\examples{
\dontrun{
torch_isnan(torch_tensor([1, float('nan'), 2]))
}

}
