% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_isnan}
\alias{torch_isnan}
\title{Isnan}
\arguments{
\item{input}{(Tensor) A tensor to check}
}
\description{
Isnan
}
\section{TEST }{


Returns a new tensor with boolean elements representing if each element is \code{NaN} or not.
}

\examples{
if (torch_is_installed()) {

torch_isnan(torch_tensor(c(1, NaN, 2)))
}
}
