% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_isreal}
\alias{torch_isreal}
\title{Isreal}
\usage{
torch_isreal(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Isreal
}
\section{isreal(input) -> Tensor }{


Returns a new tensor with boolean elements representing if each element of \code{input} is real-valued or not.
All real-valued types are considered real. Complex values are considered real when their imaginary part is 0.
}

\examples{
if (torch_is_installed()) {
if (FALSE) {
torch_isreal(torch_tensor(c(1, 1+1i, 2+0i)))
}
}
}
