% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_kthvalue}
\alias{torch_kthvalue}
\title{Kthvalue}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{k}{(int) k for the k-th smallest element}

\item{dim}{(int, optional) the dimension to find the kth value along}

\item{keepdim}{(bool) whether the output tensor has :attr:\code{dim} retained or not.}

\item{out}{(tuple, optional) the output tuple of (Tensor, LongTensor)                           can be optionally given to be used as output buffers}
}
\description{
kthvalue(input, k, dim=None, keepdim=False, out=None) -> (Tensor, LongTensor)
}
\details{
Returns a namedtuple \verb{(values, indices)} where \code{values} is the :attr:\code{k} th
smallest element of each row of the :attr:\code{input} tensor in the given dimension
:attr:\code{dim}. And \code{indices} is the index location of each element found.

If :attr:\code{dim} is not given, the last dimension of the \code{input} is chosen.

If :attr:\code{keepdim} is \code{True}, both the :attr:\code{values} and :attr:\code{indices} tensors
are the same size as :attr:\code{input}, except in the dimension :attr:\code{dim} where
they are of size 1. Otherwise, :attr:\code{dim} is squeezed
(see :func:\code{torch.squeeze}), resulting in both the :attr:\code{values} and
:attr:\code{indices} tensors having 1 fewer dimension than the :attr:\code{input} tensor.
}
