% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_lcm}
\alias{torch_lcm}
\title{Lcm}
\usage{
torch_lcm(self, other)
}
\arguments{
\item{other}{(Tensor) the second input tensor}

\item{input}{(Tensor) the input tensor.}
}
\description{
Lcm
}
\note{
This defines \eqn{lcm(0, 0) = 0} and \eqn{lcm(0, a) = 0}.
}
\section{lcm(input, other, *, out=None) -> Tensor }{


Computes the element-wise least common multiple (LCM) of \code{input} and \code{other}.

Both \code{input} and \code{other} must have integer types.
}

