% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_le}
\alias{torch_le}
\title{Le}
\arguments{
\item{input}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}

\item{out}{(Tensor, optional) the output tensor that must be a \code{BoolTensor}}
}
\description{
Computes \eqn{\text{input} \leq \text{other}} element-wise.
}
\details{
The second argument can be a number or a tensor whose shape is
broadcastable  with the first argument.
}
\section{Signatures}{


le(input, other, out=None) -> Tensor
}

\examples{
\dontrun{
torch_le(torch_tensor([[1, 2], [3, 4]]), torch_tensor([[1, 1], [4, 4]]))
}

}
