% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_le}
\alias{torch_le}
\title{Le}
\arguments{
\item{input}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}

\item{out}{(Tensor, optional) the output tensor that must be a \code{BoolTensor}}
}
\description{
le(input, other, out=None) -> Tensor
}
\details{
Computes :math:\verb{\\text\{input\} \\leq \\text\{other\}} element-wise.

The second argument can be a number or a tensor whose shape is
:ref:\verb{broadcastable <broadcasting-semantics>} with the first argument.
}
