% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_lerp}
\alias{torch_lerp}
\title{Lerp}
\arguments{
\item{input}{(Tensor) the tensor with the starting points}

\item{end}{(Tensor) the tensor with the ending points}

\item{weight}{(float or tensor) the weight for the interpolation formula}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
lerp(input, end, weight, out=None)
}
\details{
Does a linear interpolation of two tensors :attr:\code{start} (given by :attr:\code{input}) and :attr:\code{end} based
on a scalar or tensor :attr:\code{weight} and returns the resulting :attr:\code{out} tensor.

.. math::
\text{out}_i = \text{start}_i + \text{weight}_i \times (\text{end}_i - \text{start}_i)

The shapes of :attr:\code{start} and :attr:\code{end} must be
:ref:\verb{broadcastable <broadcasting-semantics>}. If :attr:\code{weight} is a tensor, then
the shapes of :attr:\code{weight}, :attr:\code{start}, and :attr:\code{end} must be :ref:\verb{broadcastable <broadcasting-semantics>}.
}
\examples{
\dontrun{
start = torch_arange(1., 5.)
end = torch_empty(4).fill_(10)
start
end
torch_lerp(start, end, 0.5)
torch_lerp(start, end, torch_full_like(start, 0.5))
}

}
