% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_lgamma}
\alias{torch_lgamma}
\title{Lgamma}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
lgamma(input, out=None) -> Tensor
}
\details{
Computes the logarithm of the gamma function on :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \log \Gamma(\text{input}}{i})
}
