% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_lgamma}
\alias{torch_lgamma}
\title{Lgamma}
\usage{
torch_lgamma(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Lgamma
}
\section{lgamma(input, out=NULL) -> Tensor }{


Computes the logarithm of the gamma function on \code{input}.

\deqn{
    \mbox{out}_{i} = \log \Gamma(\mbox{input}_{i})
}
}

\examples{

a <- torch_arange(0.5, 2, 0.5)
torch_lgamma(a)
}
