% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{torch_load}
\alias{torch_load}
\title{Loads a saved object}
\usage{
torch_load(path, device = "cpu")
}
\arguments{
\item{path}{a path to the saved object}

\item{device}{a device to load tensors to. By default we load to the \code{cpu} but you can also
load them to any \code{cuda} device.}
}
\description{
Loads a saved object
}
\seealso{
Other torch_save: 
\code{\link{torch_save}()}
}
\concept{serialization}
\concept{torch_save}
