% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_log}
\alias{torch_log}
\title{Log}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the natural logarithm of the elements
of \code{input}.
}
\details{
\deqn{
    y_{i} = \log_{e} (x_{i})
}
}
\section{Signatures}{


log(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(c(5))
a
torch_log(a)
}

}
