% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_log}
\alias{torch_log}
\title{Log}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
log(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the natural logarithm of the elements
of :attr:\code{input}.

.. math::
y_{i} = \log_{e} (x_{i})
}
