% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_log1p}
\alias{torch_log1p}
\title{Log1p}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
log1p(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the natural logarithm of (1 + \code{input}).

\deqn{
    y_i = \log_{e} (x_i + 1)
}
.. note:: This function is more accurate than \code{\link{torch_log}} for small
values of \code{input}
}
\examples{
\dontrun{
a = torch_randn(5)
a
torch_log1p(a)
}

}
