% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_log1p}
\alias{torch_log1p}
\title{Log1p}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Log1p
}
\note{
This function is more accurate than \code{\link{torch_log}} for small
values of \code{input}
}
\section{log1p(input, out=None) -> Tensor }{


Returns a new tensor with the natural logarithm of (1 + \code{input}).

\deqn{
    y_i = \log_{e} (x_i + 1)
}
}

\examples{
if (torch_is_installed()) {

a = torch_randn(c(5))
a
torch_log1p(a)
}
}
