% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_logcumsumexp}
\alias{torch_logcumsumexp}
\title{Logcumsumexp}
\usage{
torch_logcumsumexp(self, dim)
}
\arguments{
\item{dim}{(int) the dimension to do the operation over}

\item{input}{(Tensor) the input tensor.}
}
\description{
Logcumsumexp
}
\section{logcumsumexp(input, dim, *, out=None) -> Tensor }{


Returns the logarithm of the cumulative summation of the exponentiation of
elements of \code{input} in the dimension \code{dim}.

For summation index \eqn{j} given by \code{dim} and other indices \eqn{i}, the result is

\deqn{
        \mbox{logcumsumexp}(x)_{ij} = \log \sum\limits_{j=0}^{i} \exp(x_{ij})
}
}

