% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_logdet}
\alias{torch_logdet}
\title{Logdet}
\arguments{
\item{input}{(Tensor) the input tensor of size \verb{(*, n, n)} where \code{*} is zero or more                batch dimensions.}
}
\description{
Calculates log determinant of a square matrix or batches of square matrices.
}
\note{
\preformatted{Result is ``-inf`` if `input` has zero log determinant, and is ``nan`` if
`input` has negative determinant.
}

\preformatted{Backward through :meth:`logdet` internally uses SVD results when `input`
is not invertible. In this case, double backward through :meth:`logdet` will
be unstable in when `input` doesn't have distinct singular values. See
:meth:`~torch.svd` for details.
}
}
\section{Signatures}{


logdet(input) -> Tensor
}

\examples{
\dontrun{
A = torch_randn(3, 3)
torch_det(A)
torch_logdet(A)
A
A.det()
A.det().log()
}

}
