% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_logical_not}
\alias{torch_logical_not}
\title{Logical_not}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
logical_not(input, out=None) -> Tensor
}
\details{
Computes the element-wise logical NOT of the given input tensor. If not specified, the output tensor will have the bool
dtype. If the input tensor is not a bool tensor, zeros are treated as \code{False} and non-zeros are treated as \code{True}.
}
