% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_logical_not}
\alias{torch_logical_not}
\title{Logical_not}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Logical_not
}
\section{logical_not(input, out=NULL) -> Tensor }{


Computes the element-wise logical NOT of the given input tensor. If not specified, the output tensor will have the bool
dtype. If the input tensor is not a bool tensor, zeros are treated as \code{FALSE} and non-zeros are treated as \code{TRUE}.
}

\examples{
if (torch_is_installed()) {

torch_logical_not(torch_tensor(c(TRUE, FALSE)))
torch_logical_not(torch_tensor(c(0, 1, -10), dtype=torch_int8()))
torch_logical_not(torch_tensor(c(0., 1.5, -10.), dtype=torch_double()))
}
}
