% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_logical_or}
\alias{torch_logical_or}
\title{Logical_or}
\usage{
torch_logical_or(self, other)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{other}{(Tensor) the tensor to compute OR with}
}
\description{
Logical_or
}
\section{logical_or(input, other, out=NULL) -> Tensor }{


Computes the element-wise logical OR of the given input tensors. Zeros are treated as \code{FALSE} and nonzeros are
treated as \code{TRUE}.
}

\examples{

torch_logical_or(torch_tensor(c(TRUE, FALSE, TRUE)), torch_tensor(c(TRUE, FALSE, FALSE)))
a = torch_tensor(c(0, 1, 10, 0), dtype=torch_int8())
b = torch_tensor(c(4, 0, 1, 0), dtype=torch_int8())
torch_logical_or(a, b)
\dontrun{
torch_logical_or(a$double(), b$double())
torch_logical_or(a$double(), b)
torch_logical_or(a, b, out=torch_empty(4, dtype=torch_bool()))
}
}
