% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_lt}
\alias{torch_lt}
\title{Lt}
\usage{
torch_lt(self, other)
}
\arguments{
\item{self}{(Tensor) the tensor to compare}

\item{other}{(Tensor or float) the tensor or value to compare}
}
\description{
Lt
}
\section{lt(input, other, out=NULL) -> Tensor }{


Computes \eqn{\mbox{input} < \mbox{other}} element-wise.

The second argument can be a number or a tensor whose shape is
broadcastable  with the first argument.
}

\examples{

torch_lt(torch_tensor(matrix(1:4, ncol = 2, byrow=TRUE)), 
         torch_tensor(matrix(c(1,1,4,4), ncol = 2, byrow=TRUE)))
}
