% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generator.R
\name{torch_manual_seed}
\alias{torch_manual_seed}
\title{Sets the seed for generating random numbers.}
\usage{
torch_manual_seed(seed)
}
\arguments{
\item{seed}{integer seed.}
}
\description{
Sets the seed for generating random numbers.
}
