% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_masked_select}
\alias{torch_masked_select}
\title{Masked_select}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{mask}{(ByteTensor) the tensor containing the binary mask to index with}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
masked_select(input, mask, out=None) -> Tensor
}
\details{
Returns a new 1-D tensor which indexes the :attr:\code{input} tensor according to
the boolean mask :attr:\code{mask} which is a \code{BoolTensor}.

The shapes of the :attr:\code{mask} tensor and the :attr:\code{input} tensor don't need
to match, but they must be :ref:\verb{broadcastable <broadcasting-semantics>}.

.. note:: The returned tensor does \strong{not} use the same storage
as the original tensor
}
\examples{
\dontrun{
x = torch_randn(3, 4)
x
mask = x.ge(0.5)
mask
torch_masked_select(x, mask)
}

}
