% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_matrix_exp}
\alias{torch_matrix_exp}
\title{Matrix_exp}
\usage{
torch_matrix_exp(self)
}
\arguments{
\item{input}{(Tensor) the input tensor.}
}
\description{
Matrix_exp
}
\section{matrix_power(input) -> Tensor }{


Returns the matrix exponential. Supports batched input.
For a matrix \code{A}, the matrix exponential is defined as

\deqn{
    \exp^A = \sum_{k=0}^\infty A^k / k!.
}

The implementation is based on:
Bader, P.; Blanes, S.; Casas, F.
Computing the Matrix Exponential with an Optimized Taylor Polynomial Approximation.
Mathematics 2019, 7, 1174.
}

