% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_median}
\alias{torch_median}
\title{Median}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int) the dimension to reduce.}

\item{keepdim}{(bool) whether the output tensor has :attr:\code{dim} retained or not.}

\item{values}{(Tensor, optional) the output tensor}

\item{indices}{(Tensor, optional) the output index tensor}
}
\description{
median(input) -> Tensor

median(input, dim=-1, keepdim=False, values=None, indices=None) -> (Tensor, LongTensor)
}
\details{
Returns the median value of all elements in the :attr:\code{input} tensor.

Returns a namedtuple \verb{(values, indices)} where \code{values} is the median
value of each row of the :attr:\code{input} tensor in the given dimension
:attr:\code{dim}. And \code{indices} is the index location of each median value found.

By default, :attr:\code{dim} is the last dimension of the :attr:\code{input} tensor.

If :attr:\code{keepdim} is \code{True}, the output tensors are of the same size
as :attr:\code{input} except in the dimension :attr:\code{dim} where they are of size 1.
Otherwise, :attr:\code{dim} is squeezed (see :func:\code{torch.squeeze}), resulting in
the outputs tensor having 1 fewer dimension than :attr:\code{input}.
}
