% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_meshgrid}
\alias{torch_meshgrid}
\title{Meshgrid}
\usage{
torch_meshgrid(tensors)
}
\arguments{
\item{tensors}{(list of Tensor) list of scalars or 1 dimensional tensors. Scalars will be
treated (1,).}
}
\description{
Meshgrid
}
\section{TEST }{


Take \eqn{N} tensors, each of which can be either scalar or 1-dimensional
vector, and create \eqn{N} N-dimensional grids, where the \eqn{i} \code{th} grid is defined by
expanding the \eqn{i} \code{th} input over dimensions defined by other inputs.
}

\examples{
if (torch_is_installed()) {

x = torch_tensor(c(1, 2, 3))
y = torch_tensor(c(4, 5, 6))
out = torch_meshgrid(list(x, y))
out
}
}
