% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_meshgrid}
\alias{torch_meshgrid}
\title{Meshgrid}
\description{
Take :math:\code{N} tensors, each of which can be either scalar or 1-dimensional
vector, and create :math:\code{N} N-dimensional grids, where the :math:\code{i} :sup:\code{th} grid is defined by
expanding the :math:\code{i} :sup:\code{th} input over dimensions defined by other inputs.
}
\details{
\preformatted{Args:
}
}
