% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_meshgrid}
\alias{torch_meshgrid}
\title{Meshgrid}
\arguments{
\item{tensors}{(list of Tensor) list of scalars or 1 dimensional tensors. Scalars will be}

\item{treated}{(1,)}
}
\description{
Take \eqn{N} tensors, each of which can be either scalar or 1-dimensional
vector, and create \eqn{N} N-dimensional grids, where the \eqn{i} :sup:\code{th} grid is defined by
expanding the \eqn{i} :sup:\code{th} input over dimensions defined by other inputs.
}
\examples{
\dontrun{
x = torch_tensor([1, 2, 3])
y = torch_tensor([4, 5, 6])
grid_x, grid_y = torch_meshgrid(x, y)
grid_x
grid_y
}

}
