% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_min}
\alias{torch_min}
\title{Min}
\arguments{
\item{{input}}{NA}

\item{{dim}}{NA}

\item{{keepdim}}{NA}

\item{input}{(Tensor) the input tensor.}

\item{other}{(Tensor) the second input tensor}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
min(input) -> Tensor

min(input, dim, keepdim=False, out=None) -> (Tensor, LongTensor)

min(input, other, out=None) -> Tensor
}
\details{
Returns the minimum value of all elements in the :attr:\code{input} tensor.

Returns a namedtuple \verb{(values, indices)} where \code{values} is the minimum
value of each row of the :attr:\code{input} tensor in the given dimension
:attr:\code{dim}. And \code{indices} is the index location of each minimum value found
(argmin).

If :attr:\code{keepdim} is \code{True}, the output tensors are of the same size as
:attr:\code{input} except in the dimension :attr:\code{dim} where they are of size 1.
Otherwise, :attr:\code{dim} is squeezed (see :func:\code{torch.squeeze}), resulting in
the output tensors having 1 fewer dimension than :attr:\code{input}.

Each element of the tensor :attr:\code{input} is compared with the corresponding
element of the tensor :attr:\code{other} and an element-wise minimum is taken.
The resulting tensor is returned.

The shapes of :attr:\code{input} and :attr:\code{other} don't need to match,
but they must be :ref:\verb{broadcastable <broadcasting-semantics>}.

.. math::
\text{out}_i = \min(\text{tensor}_i, \text{other}_i)

.. note:: When the shapes do not match, the shape of the returned output tensor
follows the :ref:\verb{broadcasting rules <broadcasting-semantics>}.
}
\examples{
\dontrun{
a = torch_randn(1, 3)
a
torch_min(a)
}

\dontrun{
a = torch_randn(4, 4)
a
torch_min(a, 1)
}

\dontrun{
a = torch_randn(4)
a
b = torch_randn(4)
b
torch_min(a, b)
}

}
