% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_mm}
\alias{torch_mm}
\title{Mm}
\arguments{
\item{input}{(Tensor) the first matrix to be multiplied}

\item{mat2}{(Tensor) the second matrix to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Performs a matrix multiplication of the matrices \code{input} and \code{mat2}.
}
\details{
If \code{input} is a \eqn{(n \times m)} tensor, \code{mat2} is a
\eqn{(m \times p)} tensor, \code{out} will be a \eqn{(n \times p)} tensor.
}
\note{
This function does not broadcast .
For broadcasting matrix products, see \code{\link{torch_matmul}}.
}
\section{Signatures}{


mm(input, mat2, out=None) -> Tensor
}

\examples{
\dontrun{
mat1 = torch_randn(2, 3)
mat2 = torch_randn(3, 3)
torch_mm(mat1, mat2)
}

}
