% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_mm}
\alias{torch_mm}
\title{Mm}
\usage{
torch_mm(self, mat2, out_dtype)
}
\arguments{
\item{self}{(Tensor) the first matrix to be multiplied}

\item{mat2}{(Tensor) the second matrix to be multiplied}

\item{out_dtype}{(torch_dtype, optional) the output dtype}
}
\description{
Mm
}
\note{
This function does not broadcast .
For broadcasting matrix products, see \code{\link{torch_matmul}}.
}
\section{mm(input, mat2, out=NULL) -> Tensor }{


Performs a matrix multiplication of the matrices \code{input} and \code{mat2}.

If \code{input} is a \eqn{(n \times m)} tensor, \code{mat2} is a
\eqn{(m \times p)} tensor, \code{out} will be a \eqn{(n \times p)} tensor.
}

\examples{
if (torch_is_installed()) {

mat1 = torch_randn(c(2, 3))
mat2 = torch_randn(c(3, 3))
torch_mm(mat1, mat2)
}
}
