% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_mode}
\alias{torch_mode}
\title{Mode}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int) the dimension to reduce.}

\item{keepdim}{(bool) whether the output tensor has \code{dim} retained or not.}

\item{values}{(Tensor, optional) the output tensor}

\item{indices}{(Tensor, optional) the output index tensor}
}
\description{
Returns a namedtuple \verb{(values, indices)} where \code{values} is the mode
value of each row of the \code{input} tensor in the given dimension
\code{dim}, i.e. a value which appears most often
in that row, and \code{indices} is the index location of each mode value found.
}
\details{
By default, \code{dim} is the last dimension of the \code{input} tensor.

If \code{keepdim} is \code{True}, the output tensors are of the same size as
\code{input} except in the dimension \code{dim} where they are of size 1.
Otherwise, \code{dim} is squeezed (see \code{\link{torch_squeeze}}), resulting
in the output tensors having 1 fewer dimension than \code{input}.
}
\note{
This function is not defined for \code{torch_cuda.Tensor} yet.
}
\section{Signatures}{


mode(input, dim=-1, keepdim=False, values=None, indices=None) -> (Tensor, LongTensor)
}

\examples{
\dontrun{
a = torch_randint(10, (5,))
a
b = a + (torch_randn(50, 1) * 5).long()
torch_mode(b, 0)
}

}
