% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_movedim}
\alias{torch_movedim}
\title{Movedim}
\usage{
torch_movedim(self, source, destination)
}
\arguments{
\item{source}{(int or tuple of ints) Original positions of the dims to move. These must be unique.}

\item{destination}{(int or tuple of ints) Destination positions for each of the original dims. These must also be unique.}

\item{input}{(Tensor) the input tensor.}
}
\description{
Movedim
}
\section{movedim(input, source, destination) -> Tensor }{


Moves the dimension(s) of \code{input} at the position(s) in \code{source}
to the position(s) in \code{destination}.

Other dimensions of \code{input} that are not explicitly moved remain in
their original order and appear at the positions not specified in \code{destination}.
}

