% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_mul}
\alias{torch_mul}
\title{Mul}
\arguments{
\item{{input}}{NA}

\item{value}{(Number) the number to be multiplied to each element of :attr:\code{input}}

\item{{out}}{NA}

\item{input}{(Tensor) the first multiplicand tensor}

\item{other}{(Tensor) the second multiplicand tensor}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
mul(input, other, out=None)

mul(input, other, out=None)
}
\details{
Multiplies each element of the input :attr:\code{input} with the scalar
:attr:\code{other} and returns a new resulting tensor.

.. math::
\text{out}_i = \text{other} \times \text{input}_i

If :attr:\code{input} is of type \code{FloatTensor} or \code{DoubleTensor}, :attr:\code{other}
should be a real number, otherwise it should be an integer

Each element of the tensor :attr:\code{input} is multiplied by the corresponding
element of the Tensor :attr:\code{other}. The resulting tensor is returned.

The shapes of :attr:\code{input} and :attr:\code{other} must be
:ref:\verb{broadcastable <broadcasting-semantics>}.

.. math::
\text{out}_i = \text{input}_i \times \text{other}_i
}
\examples{
\dontrun{
a = torch_randn(3)
a
torch_mul(a, 100)
}

\dontrun{
a = torch_randn(4, 1)
a
b = torch_randn(1, 4)
b
torch_mul(a, b)
}

}
