% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_multinomial}
\alias{torch_multinomial}
\title{Multinomial}
\arguments{
\item{input}{(Tensor) the input tensor containing probabilities}

\item{num_samples}{(int) number of samples to draw}

\item{replacement}{(bool, optional) whether to draw with replacement or not}

\item{generator}{NA class:\code{torch.Generator}, optional): a pseudorandom number generator for sampling}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
multinomial(input, num_samples, replacement=False, *, generator=None, out=None) -> LongTensor
}
\details{
Returns a tensor where each row contains :attr:\code{num_samples} indices sampled
from the multinomial probability distribution located in the corresponding row
of tensor :attr:\code{input}.

.. note::
The rows of :attr:\code{input} do not need to sum to one (in which case we use
the values as weights), but must be non-negative, finite and have
a non-zero sum.

Indices are ordered from left to right according to when each was sampled
(first samples are placed in first column).

If :attr:\code{input} is a vector, :attr:\code{out} is a vector of size :attr:\code{num_samples}.

If :attr:\code{input} is a matrix with \code{m} rows, :attr:\code{out} is an matrix of shape
:math:\verb{(m \\times \\text\{num\\_samples\})}.

If replacement is \code{True}, samples are drawn with replacement.

If not, they are drawn without replacement, which means that when a
sample index is drawn for a row, it cannot be drawn again for that row.

.. note::
When drawn without replacement, :attr:\code{num_samples} must be lower than
number of non-zero elements in :attr:\code{input} (or the min number of non-zero
elements in each row of :attr:\code{input} if it is a matrix).
}
