% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_mv}
\alias{torch_mv}
\title{Mv}
\arguments{
\item{input}{(Tensor) matrix to be multiplied}

\item{vec}{(Tensor) vector to be multiplied}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
mv(input, vec, out=None) -> Tensor
}
\details{
Performs a matrix-vector product of the matrix :attr:\code{input} and the vector
:attr:\code{vec}.

If :attr:\code{input} is a :math:\verb{(n \\times m)} tensor, :attr:\code{vec} is a 1-D tensor of
size :math:\code{m}, :attr:\code{out} will be 1-D of size :math:\code{n}.

.. note:: This function does not :ref:\verb{broadcast <broadcasting-semantics>}.
}
\examples{
\dontrun{
mat = torch_randn(2, 3)
vec = torch_randn(3)
torch_mv(mat, vec)
}

}
