% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_neg}
\alias{torch_neg}
\title{Neg}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
neg(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the negative of the elements of \code{input}.

\deqn{
    \text{out} = -1 \times \text{input}
}
}
\examples{
\dontrun{
a = torch_randn(5)
a
torch_neg(a)
}

}
