% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_neg}
\alias{torch_neg}
\title{Neg}
\usage{
torch_neg(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Neg
}
\section{neg(input, out=NULL) -> Tensor }{


Returns a new tensor with the negative of the elements of \code{input}.

\deqn{
    \mbox{out} = -1 \times \mbox{input}
}
}

\examples{
if (torch_is_installed()) {

a = torch_randn(c(5))
a
torch_neg(a)
}
}
