% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_normal}
\alias{torch_normal}
\title{Normal}
\arguments{
\item{generator}{NA class:\code{torch.Generator}, optional): a pseudorandom number generator for sampling}

\item{mean}{(float) the mean for all distributions}

\item{std}{(float) the standard deviation for all distributions}

\item{size}{(int...) a sequence of integers defining the shape of the output tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
normal(mean, std, *, generator=None, out=None) -> Tensor

normal(mean=0.0, std, out=None) -> Tensor

normal(mean, std=1.0, out=None) -> Tensor

normal(mean, std, size, *, out=None) -> Tensor
}
\details{
Returns a tensor of random numbers drawn from separate normal distributions
whose mean and standard deviation are given.

The :attr:\code{mean} is a tensor with the mean of
each output element's normal distribution

The :attr:\code{std} is a tensor with the standard deviation of
each output element's normal distribution

The shapes of :attr:\code{mean} and :attr:\code{std} don't need to match, but the
total number of elements in each tensor need to be the same.

.. note:: When the shapes do not match, the shape of :attr:\code{mean}
is used as the shape for the returned output tensor

Similar to the function above, but the means are shared among all drawn
elements.

Similar to the function above, but the standard-deviations are shared among
all drawn elements.

Similar to the function above, but the means and standard deviations are shared
among all drawn elements. The resulting tensor has size given by :attr:\code{size}.
}
