% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_ones_like}
\alias{torch_ones_like}
\title{Ones_like}
\arguments{
\item{input}{(Tensor) the size of \code{input} will determine size of the output tensor.}

\item{dtype}{(\code{torch.dtype}, optional) the desired data type of returned Tensor.        Default: if \code{None}, defaults to the dtype of \code{input}.}

\item{layout}{(\code{torch.layout}, optional) the desired layout of returned tensor.        Default: if \code{None}, defaults to the layout of \code{input}.}

\item{device}{(\code{torch.device}, optional) the desired device of returned tensor.        Default: if \code{None}, defaults to the device of \code{input}.}

\item{requires_grad}{(bool, optional) If autograd should record operations on the        returned tensor. Default: \code{False}.}
}
\description{
Returns a tensor filled with the scalar value \code{1}, with the same size as
\code{input}. \code{torch_ones_like(input)} is equivalent to
\code{torch_ones(input.size(), dtype=input.dtype, layout=input.layout, device=input.device)}.
}
\section{Warning}{

As of 0.4, this function does not support an \code{out} keyword. As an alternative,
the old \code{torch_ones_like(input, out=output)} is equivalent to
\code{torch_ones(input.size(), out=output)}.
}

\section{Signatures}{


ones_like(input, dtype=None, layout=None, device=None, requires_grad=False) -> Tensor
}

\examples{
\dontrun{
input = torch_empty(2, 3)
torch_ones_like(input)
}

}
