% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_orgqr}
\alias{torch_orgqr}
\title{Orgqr}
\usage{
torch_orgqr(self, input2)
}
\arguments{
\item{self}{(Tensor) the \code{a} from \code{\link{torch_geqrf}}.}

\item{input2}{(Tensor) the \code{tau} from \code{\link{torch_geqrf}}.}
}
\description{
Orgqr
}
\section{orgqr(input, input2) -> Tensor }{


Computes the orthogonal matrix \code{Q} of a QR factorization, from the \verb{(input, input2)}
tuple returned by \code{\link{torch_geqrf}}.

This directly calls the underlying LAPACK function \code{?orgqr}.
See \verb{LAPACK documentation for orgqr}_ for further details.
}

