% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_ormqr}
\alias{torch_ormqr}
\title{Ormqr}
\usage{
torch_ormqr(self, input2, input3, left = TRUE, transpose = FALSE)
}
\arguments{
\item{self}{(Tensor) the \code{a} from \code{\link{torch_geqrf}}.}

\item{self2}{(Tensor) the \code{tau} from \code{\link{torch_geqrf}}.}

\item{self3}{(Tensor) the matrix to be multiplied.}
}
\description{
Ormqr
}
\section{ormqr(input, input2, input3, left=True, transpose=False) -> Tensor }{


Multiplies \code{mat} (given by \code{input3}) by the orthogonal \code{Q} matrix of the QR factorization
formed by \code{\link{torch_geqrf}} that is represented by \verb{(a, tau)} (given by (\code{input}, \code{input2})).

This directly calls the underlying LAPACK function \code{?ormqr}.
See \verb{LAPACK documentation for ormqr}_ for further details.
}

