% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_pdist}
\alias{torch_pdist}
\title{Pdist}
\arguments{
\item{input}{NA input tensor of shape :math:\verb{N \\times M}.}

\item{p}{NA p value for the p-norm distance to calculate between each vector pair        :math:\verb{\\in [0, \\infty]}.}
}
\description{
pdist(input, p=2) -> Tensor
}
\details{
Computes the p-norm distance between every pair of row vectors in the input.
This is identical to the upper triangular portion, excluding the diagonal, of
\verb{torch.norm(input[:, None] - input, dim=2, p=p)}. This function will be faster
if the rows are contiguous.

If input has shape :math:\verb{N \\times M} then the output will have shape
:math:\verb{\\frac\{1\}\{2\} N (N - 1)}.

This function is equivalent to \code{scipy.spatial.distance.pdist(input, 'minkowski', p=p)} if :math:\verb{p \\in (0, \\infty)}. When :math:\code{p = 0} it is
equivalent to \code{scipy.spatial.distance.pdist(input, 'hamming') * M}.
When :math:\verb{p = \\infty}, the closest scipy function is
\verb{scipy.spatial.distance.pdist(xn, lambda x, y: np.abs(x - y).max())}.
}
