% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_pinverse}
\alias{torch_pinverse}
\title{Pinverse}
\arguments{
\item{input}{(Tensor) The input tensor of size :math:\verb{(*, m, n)} where :math:\code{*} is zero or more batch dimensions}

\item{rcond}{(float) A floating point value to determine the cutoff for small singular values.}

\item{Default}{NA 1e-15}
}
\description{
pinverse(input, rcond=1e-15) -> Tensor
}
\details{
Calculates the pseudo-inverse (also known as the Moore-Penrose inverse) of a 2D tensor.
Please look at \verb{Moore-Penrose inverse}_ for more details

.. note::
This method is implemented using the Singular Value Decomposition.

.. note::
The pseudo-inverse is not necessarily a continuous function in the elements of the matrix \verb{[1]}\emph{.
Therefore, derivatives are not always existent, and exist for a constant rank only \verb{[2]}}.
However, this method is backprop-able due to the implementation by using SVD results, and
could be unstable. Double-backward will also be unstable due to the usage of SVD internally.
See :meth:\code{~torch.svd} for more details.

Arguments:
input (Tensor): The input tensor of size :math:\verb{(*, m, n)} where :math:\code{*} is zero or more batch dimensions
rcond (float): A floating point value to determine the cutoff for small singular values.
Default: 1e-15

Returns:
The pseudo-inverse of :attr:\code{input} of dimensions :math:\verb{(*, n, m)}

Example::
}
