% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_pixel_shuffle}
\alias{torch_pixel_shuffle}
\title{Pixel_shuffle}
\arguments{
\item{input}{(Tensor) the input tensor}

\item{upscale_factor}{(int) factor to increase spatial resolution by}
}
\description{
Rearranges elements in a tensor of shape \eqn{(*, C \times r^2, H, W)} to a
tensor of shape \eqn{(*, C, H \times r, W \times r)}.
}
\details{
See \code{~torch.nn.PixelShuffle} for details.
}
\section{Signatures}{


Rearranges elements in a tensor of shape :math:\verb{(*, C \\times r^2, H, W)} to a
}

