% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_prod}
\alias{torch_prod}
\title{Prod}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int) the dimension to reduce.}

\item{keepdim}{(bool) whether the output tensor has :attr:\code{dim} retained or not.}

\item{dtype}{(:class:\code{torch.dtype}, optional) the desired data type of returned tensor.        If specified, the input tensor is casted to :attr:\code{dtype} before the operation        is performed. This is useful for preventing data type overflows. Default: None.}
}
\description{
prod(input, dtype=None) -> Tensor

prod(input, dim, keepdim=False, dtype=None) -> Tensor
}
\details{
Returns the product of all elements in the :attr:\code{input} tensor.

Returns the product of each row of the :attr:\code{input} tensor in the given
dimension :attr:\code{dim}.

If :attr:\code{keepdim} is \code{True}, the output tensor is of the same size
as :attr:\code{input} except in the dimension :attr:\code{dim} where it is of size 1.
Otherwise, :attr:\code{dim} is squeezed (see :func:\code{torch.squeeze}), resulting in
the output tensor having 1 fewer dimension than :attr:\code{input}.
}
\examples{
\dontrun{
a = torch_randn(1, 3)
a
torch_prod(a)
}

\dontrun{
a = torch_randn(4, 2)
a
torch_prod(a, 1)
}

}
