% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_quantize_per_tensor}
\alias{torch_quantize_per_tensor}
\title{Quantize_per_tensor}
\arguments{
\item{input}{(Tensor) float tensor to quantize}

\item{scale}{(float) scale to apply in quantization formula}

\item{zero_point}{(int) offset in integer value that maps to float zero}

\item{dtype}{NA class:\code{torch.dtype}): the desired data type of returned tensor.}

\item{Has}{NA \code{torch.quint8}, \code{torch.qint8}, \code{torch.qint32}}
}
\description{
quantize_per_tensor(input, scale, zero_point, dtype) -> Tensor
}
\details{
Converts a float tensor to quantized tensor with given scale and zero point.

Arguments:
input (Tensor): float tensor to quantize
scale (float): scale to apply in quantization formula
zero_point (int): offset in integer value that maps to float zero
dtype (:class:\code{torch.dtype}): the desired data type of returned tensor.
Has to be one of the quantized dtypes: \code{torch.quint8}, \code{torch.qint8}, \code{torch.qint32}

Returns:
Tensor: A newly quantized tensor

Example::
}
