% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_randint_like}
\alias{torch_randint_like}
\title{Randint_like}
\arguments{
\item{input}{(Tensor) the size of :attr:\code{input} will determine size of the output tensor.}

\item{low}{(int, optional) Lowest integer to be drawn from the distribution. Default: 0.}

\item{high}{(int) One above the highest integer to be drawn from the distribution.}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned Tensor.}

\item{layout}{NA class:\code{torch.layout}, optional): the desired layout of returned tensor.}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if \code{None}, defaults to the device of :attr:\code{input}.}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}
}
\description{
randint_like(input, low=0, high, dtype=None, layout=torch.strided, device=None, requires_grad=False) -> Tensor
}
\details{
Returns a tensor with the same shape as Tensor :attr:\code{input} filled with
random integers generated uniformly between :attr:\code{low} (inclusive) and
:attr:\code{high} (exclusive).

.. note:
With the global dtype default (\code{torch.float32}), this function returns
a tensor with dtype \code{torch.int64}.
}
