% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_randn_like}
\alias{torch_randn_like}
\title{Randn_like}
\arguments{
\item{input}{(Tensor) the size of :attr:\code{input} will determine size of the output tensor.}

\item{dtype}{(:class:\code{torch.dtype}, optional) the desired data type of returned Tensor.        Default: if \code{None}, defaults to the dtype of :attr:\code{input}.}

\item{layout}{(:class:\code{torch.layout}, optional) the desired layout of returned tensor.        Default: if \code{None}, defaults to the layout of :attr:\code{input}.}

\item{device}{(:class:\code{torch.device}, optional) the desired device of returned tensor.        Default: if \code{None}, defaults to the device of :attr:\code{input}.}

\item{requires_grad}{(bool, optional) If autograd should record operations on the        returned tensor. Default: \code{False}.}
}
\description{
randn_like(input, dtype=None, layout=None, device=None, requires_grad=False) -> Tensor
}
\details{
Returns a tensor with the same size as :attr:\code{input} that is filled with
random numbers from a normal distribution with mean 0 and variance 1.
\code{torch.randn_like(input)} is equivalent to
\code{torch.randn(input.size(), dtype=input.dtype, layout=input.layout, device=input.device)}.
}
