% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_range}
\alias{torch_range}
\title{Range}
\arguments{
\item{start}{(float) the starting value for the set of points. Default: \code{0}.}

\item{end}{(float) the ending value for the set of points}

\item{step}{(float) the gap between each pair of adjacent points. Default: \code{1}.}

\item{out}{(Tensor, optional) the output tensor.}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{arguments.}{NA}

\item{`dtype`}{NA}

\item{be}{NA}

\item{layout}{NA class:\code{torch.layout}, optional): the desired layout of returned Tensor.}

\item{device}{NA class:\code{torch.device}, optional): the desired device of returned tensor.}

\item{Default}{NA if \code{None}, uses the current device for the default tensor type}

\item{NA}{func:\code{torch.set_default_tensor_type}). :attr:\code{device} will be the CPU}

\item{for}{NA}

\item{requires_grad}{(bool, optional) If autograd should record operations on the}

\item{returned}{NA \code{False}.}
}
\description{
range(start=0, end, step=1, out=None, dtype=None, layout=torch.strided, device=None, requires_grad=False) -> Tensor
}
\details{
Returns a 1-D tensor of size :math:\verb{\\left\\lfloor \\frac\{\\text\{end\} - \\text\{start\}\}\{\\text\{step\}\} \\right\\rfloor + 1}
with values from :attr:\code{start} to :attr:\code{end} with step :attr:\code{step}. Step is
the gap between two values in the tensor.

.. math::
\text{out}_{i+1} = \text{out}_i + \text{step}.

.. warning::
This function is deprecated in favor of :func:\code{torch.arange}.
}
