% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_real}
\alias{torch_real}
\title{Real}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
real(input, out=None) -> Tensor
}
\details{
Computes the element-wise real value of the given :attr:\code{input} tensor.

.. math::
\text{out}\emph{{i} = real(\text{input}}{i})
}
