% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_real}
\alias{torch_real}
\title{Real}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Real
}
\section{real(input, out=None) -> Tensor }{


Returns the real part of the \code{input} tensor. If
\code{input} is a real (non-complex) tensor, this function just
returns it.
}

\section{Warning}{

Not yet implemented for complex tensors.

\deqn{
    \text{out}_{i} = real(\text{input}_{i})
}
}

\examples{
\dontrun{
torch_real(torch_tensor(c(-1 + 1i, -2 + 2i, 3 - 3i)))
}
}
