% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_real}
\alias{torch_real}
\title{Real}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Real
}
\section{real(input, out=None) -> Tensor }{


Computes the element-wise real value of the given \code{input} tensor.

\deqn{
    \text{out}_{i} = real(\text{input}_{i})
}
}

\examples{

\dontrun{
torch_real(torch_tensor(c(-1 + 1i, -2 + 2i, 3 - 3i)))
}

}
