% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_reciprocal}
\alias{torch_reciprocal}
\title{Reciprocal}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
reciprocal(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the reciprocal of the elements of :attr:\code{input}

.. math::
\text{out}\emph{{i} = \frac{1}{\text{input}}{i}}
}
