% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_reciprocal}
\alias{torch_reciprocal}
\title{Reciprocal}
\usage{
torch_reciprocal(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Reciprocal
}
\section{reciprocal(input, out=NULL) -> Tensor }{


Returns a new tensor with the reciprocal of the elements of \code{input}

\deqn{
    \mbox{out}_{i} = \frac{1}{\mbox{input}_{i}}
}
}

\examples{

a <- torch_randn(c(4))
a
torch_reciprocal(a)
}
