% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_reciprocal}
\alias{torch_reciprocal}
\title{Reciprocal}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the reciprocal of the elements of \code{input}
}
\details{
\deqn{
    \text{out}_{i} = \frac{1}{\text{input}_{i}}
}
}
\section{Signatures}{


reciprocal(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(c(4))
a
torch_reciprocal(a)
}

}
