% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_relu_}
\alias{torch_relu_}
\title{Relu_}
\usage{
torch_relu_(self)
}
\description{
Relu_
}
\section{relu_(input) -> Tensor }{


In-place version of \code{torch_relu}.
}

