% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_remainder}
\alias{torch_remainder}
\title{Remainder}
\arguments{
\item{input}{(Tensor) the dividend}

\item{other}{(Tensor or float) the divisor that may be either a number or a                               Tensor of the same shape as the dividend}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Remainder
}
\section{remainder(input, other, out=None) -> Tensor }{


Computes the element-wise remainder of division.

The divisor and dividend may contain both for integer and floating point
numbers. The remainder has the same sign as the divisor.

When \code{other} is a tensor, the shapes of \code{input} and
\code{other} must be broadcastable .
}

\examples{

torch_remainder(torch_tensor(c(-3., -2, -1, 1, 2, 3)), 2)
torch_remainder(torch_tensor(c(1., 2, 3, 4, 5)), 1.5)
}
