% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_result_type}
\alias{torch_result_type}
\title{Result_type}
\arguments{
\item{tensor1}{(Tensor or Number) an input tensor or number}

\item{tensor2}{(Tensor or Number) an input tensor or number}
}
\description{
Result_type
}
\section{result_type(tensor1, tensor2) -> dtype }{


Returns the \code{torch_dtype} that would result from performing an arithmetic
operation on the provided input tensors. See type promotion documentation
for more information on the type promotion logic.
}

\examples{
if (torch_is_installed()) {

torch_result_type(tensor = torch_tensor(c(1, 2), dtype=torch_int()), 1.0)
}
}
