% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_rot90}
\alias{torch_rot90}
\title{Rot90}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{k}{(int) number of times to rotate}

\item{dims}{(a list or tuple) axis to rotate}
}
\description{
Rot90
}
\section{rot90(input, k, dims) -> Tensor }{


Rotate a n-D tensor by 90 degrees in the plane specified by dims axis.
Rotation direction is from the first towards the second axis if k > 0, and from the second towards the first for k < 0.
}

\examples{

\dontrun{
x = torch_arange(4)$view(2, 2)
x
torch_rot90list(x, 1, c(0, 1))
x = torch_arange(8)$view(2, 2, 2)
x
torch_rot90list(x, 1, c(1, 2))
}

}
