% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_round}
\alias{torch_round}
\title{Round}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
round(input, out=None) -> Tensor
}
\details{
Returns a new tensor with each of the elements of :attr:\code{input} rounded
to the closest integer.
}
