% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_rrelu_}
\alias{torch_rrelu_}
\title{Rrelu_}
\description{
Rrelu_
}
\section{rrelu_(input, lower=1./8, upper=1./3, training=False) -> Tensor }{


In-place version of \code{torch_rrelu}.
}

