% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_rrelu_}
\alias{torch_rrelu_}
\title{Rrelu_}
\usage{
torch_rrelu_(
  self,
  lower = 0.125,
  upper = 0.333333333333333,
  training = FALSE,
  generator = NULL
)
}
\arguments{
\item{self}{the input tensor}

\item{lower}{lower bound of the uniform distribution. Default: 1/8}

\item{upper}{upper bound of the uniform distribution. Default: 1/3}

\item{training}{bool wether it's a training pass. DEfault: FALSE}

\item{generator}{random number generator}
}
\description{
Rrelu_
}
\section{rrelu_(input, lower=1./8, upper=1./3, training=False) -> Tensor }{


In-place version of \code{torch_rrelu}.
}

