% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_rsqrt}
\alias{torch_rsqrt}
\title{Rsqrt}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
rsqrt(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the reciprocal of the square-root of each of
the elements of :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \frac{1}{\sqrt{\text{input}}{i}}}
}
\examples{
\dontrun{
a = torch_randn(4)
a
torch_rsqrt(a)
}

}
