% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_rsqrt}
\alias{torch_rsqrt}
\title{Rsqrt}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Rsqrt
}
\section{rsqrt(input, out=None) -> Tensor }{


Returns a new tensor with the reciprocal of the square-root of each of
the elements of \code{input}.

\deqn{
    \mbox{out}_{i} = \frac{1}{\sqrt{\mbox{input}_{i}}}
}
}

\examples{

a = torch_randn(c(4))
a
torch_rsqrt(a)
}
