% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{torch_save}
\alias{torch_save}
\title{Saves an object to a disk file}
\usage{
torch_save(obj, path, ...)
}
\arguments{
\item{obj}{the saved object}

\item{path}{a path to save the object}
}
\description{
Saves an object to a disk file
}
