% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{torch_serialize}
\alias{torch_serialize}
\title{Serialize a torch object returning a raw object}
\usage{
torch_serialize(obj, ...)
}
\arguments{
\item{obj}{the saved object}

\item{...}{Additional arguments passed to \code{\link[=torch_save]{torch_save()}}. \code{obj} and \code{path} are
not accepted as they are set by \code{\link[=torch_serialize]{torch_serialize()}}.}
}
\value{
A raw vector containing the serialized object. Can be reloaded using
\code{\link[=torch_load]{torch_load()}}.
}
\description{
It's just a wraper around \code{\link[=torch_save]{torch_save()}}.
}
\seealso{
Other torch_save: 
\code{\link{torch_load}()},
\code{\link{torch_save}()}
}
\concept{serialization}
\concept{torch_save}
