% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_sigmoid}
\alias{torch_sigmoid}
\title{Sigmoid}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Sigmoid
}
\section{sigmoid(input, out=None) -> Tensor }{


Returns a new tensor with the sigmoid of the elements of \code{input}.

\deqn{
    \mbox{out}_{i} = \frac{1}{1 + e^{-\mbox{input}_{i}}}
}
}

\examples{
\dontrun{

a = torch_randn(c(4))
a
torch_sigmoid(a)
}
}
