% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_sign}
\alias{torch_sign}
\title{Sign}
\usage{
torch_sign(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Sign
}
\section{sign(input, out=None) -> Tensor }{


Returns a new tensor with the signs of the elements of \code{input}.

\deqn{
    \mbox{out}_{i} = \mbox{sgn}(\mbox{input}_{i})
}
}

\examples{
if (torch_is_installed()) {

a = torch_tensor(c(0.7, -1.2, 0., 2.3))
a
torch_sign(a)
}
}
