% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_sign}
\alias{torch_sign}
\title{Sign}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the signs of the elements of \code{input}.
}
\details{
\deqn{
    \text{out}_{i} = \operatorname{sgn}(\text{input}_{i})
}
}
\section{Signatures}{


sign(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_tensor([0.7, -1.2, 0., 2.3])
a
torch_sign(a)
}

}
