% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_signbit}
\alias{torch_signbit}
\title{Signbit}
\usage{
torch_signbit(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Signbit
}
\section{signbit(input, *, out=None) -> Tensor }{


Tests if each element of \code{input} has its sign bit set (is less than zero) or not.
}

\examples{
if (torch_is_installed()) {

a <- torch_tensor(c(0.7, -1.2, 0., 2.3))
torch_signbit(a)
}
}
