% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_sin}
\alias{torch_sin}
\title{Sin}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Sin
}
\section{sin(input, out=None) -> Tensor }{


Returns a new tensor with the sine of the elements of \code{input}.

\deqn{
    \mbox{out}_{i} = \sin(\mbox{input}_{i})
}
}

\examples{

a = torch_randn(c(4))
a
torch_sin(a)
}
