% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_sinh}
\alias{torch_sinh}
\title{Sinh}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the hyperbolic sine of the elements of
\code{input}.
}
\details{
\deqn{
    \text{out}_{i} = \sinh(\text{input}_{i})
}
}
\section{Signatures}{


sinh(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(4)
a
torch_sinh(a)
}

}
