% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_sinh}
\alias{torch_sinh}
\title{Sinh}
\usage{
torch_sinh(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Sinh
}
\section{sinh(input, out=None) -> Tensor }{


Returns a new tensor with the hyperbolic sine of the elements of
\code{input}.

\deqn{
    \mbox{out}_{i} = \sinh(\mbox{input}_{i})
}
}

\examples{
if (torch_is_installed()) {

a = torch_randn(c(4))
a
torch_sinh(a)
}
}
