% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_sinh}
\alias{torch_sinh}
\title{Sinh}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
sinh(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the hyperbolic sine of the elements of
:attr:\code{input}.

.. math::
\text{out}\emph{{i} = \sinh(\text{input}}{i})
}
\examples{
\dontrun{
a = torch_randn(4)
a
torch_sinh(a)
}

}
