% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_slogdet}
\alias{torch_slogdet}
\title{Slogdet}
\arguments{
\item{input}{(Tensor) the input tensor of size \verb{(*, n, n)} where \code{*} is zero or more}

\item{batch}{NA}
}
\description{
slogdet(input) -> (Tensor, Tensor)
}
\details{
Calculates the sign and log absolute value of the determinant(s) of a square matrix or batches of square matrices.

.. note::
If \code{input} has zero determinant, this returns \verb{(0, -inf)}.

.. note::
Backward through :meth:\code{slogdet} internally uses SVD results when :attr:\code{input}
is not invertible. In this case, double backward through :meth:\code{slogdet}
will be unstable in when :attr:\code{input} doesn't have distinct singular values.
See :meth:\code{~torch.svd} for details.

Arguments:
input (Tensor): the input tensor of size \verb{(*, n, n)} where \code{*} is zero or more
batch dimensions.

Returns:
A namedtuple (sign, logabsdet) containing the sign of the determinant, and the log
value of the absolute determinant.

Example::
}
