% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_slogdet}
\alias{torch_slogdet}
\title{Slogdet}
\arguments{
\item{input}{(Tensor) the input tensor of size \verb{(*, n, n)} where \code{*} is zero or more                batch dimensions.}
}
\description{
Calculates the sign and log absolute value of the determinant(s) of a square matrix or batches of square matrices.
}
\note{
\preformatted{If ``input`` has zero determinant, this returns ``(0, -inf)``.
}

\preformatted{Backward through :meth:`slogdet` internally uses SVD results when `input`
is not invertible. In this case, double backward through :meth:`slogdet`
will be unstable in when `input` doesn't have distinct singular values.
See :meth:`~torch.svd` for details.
}
}
\section{Signatures}{


slogdet(input) -> (Tensor, Tensor)
}

\examples{
\dontrun{
A = torch_randn(3, 3)
A
torch_det(A)
torch_logdet(A)
torch_slogdet(A)
}

}
