% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_slogdet}
\alias{torch_slogdet}
\title{Slogdet}
\arguments{
\item{input}{(Tensor) the input tensor of size \verb{(*, n, n)} where \code{*} is zero or more                batch dimensions.}
}
\description{
slogdet(input) -> (Tensor, Tensor)
}
\details{
Calculates the sign and log absolute value of the determinant(s) of a square matrix or batches of square matrices.

.. note::
If \code{input} has zero determinant, this returns \verb{(0, -inf)}.

.. note::
Backward through :meth:\code{slogdet} internally uses SVD results when :attr:\code{input}
is not invertible. In this case, double backward through :meth:\code{slogdet}
will be unstable in when :attr:\code{input} doesn't have distinct singular values.
See :meth:\code{~torch.svd} for details.
}
\examples{
\dontrun{
A = torch_randn(3, 3)
A
torch_det(A)
torch_logdet(A)
torch_slogdet(A)
}

}
