% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_solve}
\alias{torch_solve}
\title{Solve}
\arguments{
\item{input}{(Tensor) input matrix :math:\code{B} of size :math:\verb{(*, m, k)} , where :math:\code{*}                is zero or more batch dimensions.}

\item{A}{(Tensor) input square matrix of size :math:\verb{(*, m, m)}, where                :math:\code{*} is zero or more batch dimensions.}

\item{out}{((Tensor, Tensor) optional output tuple.}
}
\description{
This function returns the solution to the system of linear
equations represented by :math:\code{AX = B} and the LU factorization of
A, in order as a namedtuple \verb{solution, LU}.
}
\details{
\code{LU} contains \code{L} and \code{U} factors for LU factorization of \code{A}.

\code{torch_solve(B, A)} can take in 2D inputs \verb{B, A} or inputs that are
batches of 2D matrices. If the inputs are batches, then returns
batched outputs \verb{solution, LU}.
}
\note{
\preformatted{Irrespective of the original strides, the returned matrices
`solution` and `LU` will be transposed, i.e. with strides like
`B.contiguous().transpose(-1, -2).stride()` and
`A.contiguous().transpose(-1, -2).stride()` respectively.
}
}
\section{Signatures}{


torch.solve(input, A, out=None) -> (Tensor, Tensor)
}

\examples{
\dontrun{
A = torch_tensor([[6.80, -2.11,  5.66,  5.97,  8.23],
B = torch_tensor([[4.02,  6.19, -8.22, -7.57, -3.03],
X, LU = torch_solve(B, A)
torch_dist(B, torch_mm(A, X))
# Batched solver example
A = torch_randn(2, 3, 1, 4, 4)
B = torch_randn(2, 3, 1, 4, 6)
X, LU = torch_solve(B, A)
torch_dist(B, A.matmul(X))
}

}
