% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_split}
\alias{torch_split}
\title{Split}
\arguments{
\item{tensor}{(Tensor) tensor to split.}

\item{split_size_or_sections}{(int) size of a single chunk or        list of sizes for each chunk}

\item{dim}{(int) dimension along which to split the tensor.}
}
\description{
Splits the tensor into chunks.
}
\details{
\preformatted{If `split_size_or_sections` is an integer type, then `tensor` will
be split into equally sized chunks (if possible). Last chunk will be smaller if
the tensor size along the given dimension `dim` is not divisible by
`split_size`.

If `split_size_or_sections` is a list, then `tensor` will be split
into ``len(split_size_or_sections)`` chunks with sizes in `dim` according
to `split_size_or_sections`.
}
}
