% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_split}
\alias{torch_split}
\title{Split}
\usage{
torch_split(self, split_size, dim = 1L)
}
\arguments{
\item{self}{(Tensor) tensor to split.}

\item{split_size}{(int) size of a single chunk or        list of sizes for each chunk}

\item{dim}{(int) dimension along which to split the tensor.}
}
\description{
Split
}
\section{TEST }{


Splits the tensor into chunks. Each chunk is a view of the original tensor.\preformatted{If `split_size_or_sections` is an integer type, then `tensor` will
be split into equally sized chunks (if possible). Last chunk will be smaller if
the tensor size along the given dimension `dim` is not divisible by
`split_size`.

If `split_size_or_sections` is a list, then `tensor` will be split
into `len(split_size_or_sections)` chunks with sizes in `dim` according
to `split_size_or_sections`.
}
}

