% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_sqrt}
\alias{torch_sqrt}
\title{Sqrt}
\usage{
torch_sqrt(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Sqrt
}
\section{sqrt(input, out=NULL) -> Tensor }{


Returns a new tensor with the square-root of the elements of \code{input}.

\deqn{
    \mbox{out}_{i} = \sqrt{\mbox{input}_{i}}
}
}

\examples{
if (torch_is_installed()) {

a = torch_randn(c(4))
a
torch_sqrt(a)
}
}
