% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_sqrt}
\alias{torch_sqrt}
\title{Sqrt}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Sqrt
}
\section{sqrt(input, out=None) -> Tensor }{


Returns a new tensor with the square-root of the elements of \code{input}.

\deqn{
    \text{out}_{i} = \sqrt{\text{input}_{i}}
}
}

\examples{

a = torch_randn(c(4))
a
torch_sqrt(a)
}
