% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_sqrt}
\alias{torch_sqrt}
\title{Sqrt}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
sqrt(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the square-root of the elements of :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \sqrt{\text{input}}{i}}
}
