% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_square}
\alias{torch_square}
\title{Square}
\usage{
torch_square(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Square
}
\section{square(input, out=NULL) -> Tensor }{


Returns a new tensor with the square of the elements of \code{input}.
}

\examples{

a <- torch_randn(c(4))
a
torch_square(a)
}
