% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_stack}
\alias{torch_stack}
\title{Stack}
\usage{
torch_stack(x)
}
\arguments{
\item{tensors}{(sequence of Tensors) sequence of tensors to concatenate}

\item{dim}{(int) dimension to insert. Has to be between 0 and the number        of dimensions of concatenated tensors (inclusive)}
}
\description{
Stack
}
\section{stack(tensors, dim=0, out=NULL) -> Tensor }{


Concatenates sequence of tensors along a new dimension.

All tensors need to be of the same size.
}

