% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-new-docs.R, R/gen-namespace.R
\name{torch_subtract}
\alias{torch_subtract}
\title{Subtract}
\usage{
torch_subtract(self, other, alpha = 1L)
}
\arguments{
\item{other}{(Tensor or Scalar) the tensor or scalar to subtract from \code{input}}
}
\description{
Subtract
}
\section{subtract(input, other, *, alpha=1, out=None) -> Tensor }{


Alias for \code{\link[=torch_sub]{torch_sub()}}.
}

