% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_sum}
\alias{torch_sum}
\title{Sum}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{dim}{(int or tuple of ints) the dimension or dimensions to reduce.}

\item{keepdim}{(bool) whether the output tensor has :attr:\code{dim} retained or not.}

\item{dtype}{NA class:\code{torch.dtype}, optional): the desired data type of returned tensor.}

\item{If}{NA attr:\code{dtype} before the operation}

\item{is}{NA None.}
}
\description{
sum(input, dtype=None) -> Tensor

sum(input, dim, keepdim=False, dtype=None) -> Tensor
}
\details{
Returns the sum of all elements in the :attr:\code{input} tensor.

Returns the sum of each row of the :attr:\code{input} tensor in the given
dimension :attr:\code{dim}. If :attr:\code{dim} is a list of dimensions,
reduce over all of them.

If :attr:\code{keepdim} is \code{True}, the output tensor is of the same size
as :attr:\code{input} except in the dimension(s) :attr:\code{dim} where it is of size 1.
Otherwise, :attr:\code{dim} is squeezed (see :func:\code{torch.squeeze}), resulting in the
output tensor having 1 (or \code{len(dim)}) fewer dimension(s).
}
