% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R
\name{torch_t}
\alias{torch_t}
\title{T}
\arguments{
\item{input}{(Tensor) the input tensor.}
}
\description{
T
}
\section{t(input) -> Tensor }{


Expects \code{input} to be <= 2-D tensor and transposes dimensions 0
and 1.

0-D and 1-D tensors are returned as is. When input is a 2-D tensor this
is equivalent to \code{transpose(input, 0, 1)}.
}

\examples{

x = torch_randn(c(2,3))
x
torch_t(x)
x = torch_randn(c(3))
x
torch_t(x)
x = torch_randn(c(2, 3))
x
torch_t(x)
}
