% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_take}
\alias{torch_take}
\title{Take}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{indices}{(LongTensor) the indices into tensor}
}
\description{
take(input, index) -> Tensor
}
\details{
Returns a new tensor with the elements of :attr:\code{input} at the given indices.
The input tensor is treated as if it were viewed as a 1-D tensor. The result
takes the same shape as the indices.
}
