% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_tan}
\alias{torch_tan}
\title{Tan}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
tan(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the tangent of the elements of :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \tan(\text{input}}{i})
}
