% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_tan}
\alias{torch_tan}
\title{Tan}
\usage{
torch_tan(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Tan
}
\section{tan(input, out=NULL) -> Tensor }{


Returns a new tensor with the tangent of the elements of \code{input}.

\deqn{
    \mbox{out}_{i} = \tan(\mbox{input}_{i})
}
}

\examples{
if (torch_is_installed()) {

a = torch_randn(c(4))
a
torch_tan(a)
}
}
