% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_tanh}
\alias{torch_tanh}
\title{Tanh}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
tanh(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the hyperbolic tangent of the elements
of :attr:\code{input}.

.. math::
\text{out}\emph{{i} = \tanh(\text{input}}{i})
}
