% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_tanh}
\alias{torch_tanh}
\title{Tanh}
\usage{
torch_tanh(self)
}
\arguments{
\item{self}{(Tensor) the input tensor.}
}
\description{
Tanh
}
\section{tanh(input, out=NULL) -> Tensor }{


Returns a new tensor with the hyperbolic tangent of the elements
of \code{input}.

\deqn{
    \mbox{out}_{i} = \tanh(\mbox{input}_{i})
}
}

\examples{

a <- torch_randn(c(4))
a
torch_tanh(a)
}
