% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_tensordot}
\alias{torch_tensordot}
\title{Tensordot}
\arguments{
\item{a}{(Tensor) Left tensor to contract}

\item{b}{(Tensor) Right tensor to contract}

\item{dims}{(int or tuple of two lists of integers) number of dimensions to     contract or explicit lists of dimensions for \code{a} and     \code{b} respectively}
}
\description{
Tensordot
}
\section{TEST }{


Returns a contraction of a and b over multiple dimensions.\preformatted{`tensordot` implements a generalized matrix product.
}
}

\examples{



}
