% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_tensordot}
\alias{torch_tensordot}
\title{Tensordot}
\description{
Returns a contraction of a and b over multiple dimensions.
}
\details{
\preformatted{:attr:`tensordot` implements a generalized matrix product.

Args:
  a (Tensor): Left tensor to contract
}
}
