% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_topk}
\alias{torch_topk}
\title{Topk}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{k}{(int) the k in "top-k"}

\item{dim}{(int, optional) the dimension to sort along}

\item{largest}{(bool, optional) controls whether to return largest or           smallest elements}

\item{sorted}{(bool, optional) controls whether to return the elements           in sorted order}

\item{out}{(tuple, optional) the output tuple of (Tensor, LongTensor) that can be        optionally given to be used as output buffers}
}
\description{
topk(input, k, dim=None, largest=True, sorted=True, out=None) -> (Tensor, LongTensor)
}
\details{
Returns the :attr:\code{k} largest elements of the given :attr:\code{input} tensor along
a given dimension.

If :attr:\code{dim} is not given, the last dimension of the \code{input} is chosen.

If :attr:\code{largest} is \code{False} then the \code{k} smallest elements are returned.

A namedtuple of \verb{(values, indices)} is returned, where the \code{indices} are the indices
of the elements in the original \code{input} tensor.

The boolean option :attr:\code{sorted} if \code{True}, will make sure that the returned
\code{k} elements are themselves sorted
}
