% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_triangular_solve}
\alias{torch_triangular_solve}
\title{Triangular_solve}
\arguments{
\item{input}{(Tensor) multiple right-hand sides of size :math:\verb{(*, m, k)} where                :math:\code{*} is zero of more batch dimensions (:math:\code{b})}

\item{A}{(Tensor) the input triangular coefficient matrix of size :math:\verb{(*, m, m)}                where :math:\code{*} is zero or more batch dimensions}

\item{upper}{(bool, optional) whether to solve the upper-triangular system        of equations (default) or the lower-triangular system of equations. Default: \code{True}.}

\item{transpose}{(bool, optional) whether :math:\code{A} should be transposed before        being sent into the solver. Default: \code{False}.}

\item{unitriangular}{(bool, optional) whether :math:\code{A} is unit triangular.        If True, the diagonal elements of :math:\code{A} are assumed to be        1 and not referenced from :math:\code{A}. Default: \code{False}.}
}
\description{
Solves a system of equations with a triangular coefficient matrix \eqn{A}
and multiple right-hand sides \eqn{b}.
}
\details{
In particular, solves \eqn{AX = b} and assumes \eqn{A} is upper-triangular
with the default keyword arguments.

\code{torch_triangular_solve(b, A)} can take in 2D inputs \verb{b, A} or inputs that are
batches of 2D matrices. If the inputs are batches, then returns
batched outputs \code{X}
}
\section{Signatures}{


triangular_solve(input, A, upper=True, transpose=False, unitriangular=False) -> (Tensor, Tensor)
}

