% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_trunc}
\alias{torch_trunc}
\title{Trunc}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
trunc(input, out=None) -> Tensor
}
\details{
Returns a new tensor with the truncated integer values of
the elements of :attr:\code{input}.
}
