% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_trunc}
\alias{torch_trunc}
\title{Trunc}
\arguments{
\item{input}{(Tensor) the input tensor.}

\item{out}{(Tensor, optional) the output tensor.}
}
\description{
Returns a new tensor with the truncated integer values of
the elements of \code{input}.
}
\section{Signatures}{


trunc(input, out=None) -> Tensor
}

\examples{
\dontrun{
a = torch_randn(c(4))
a
torch_trunc(a)
}

}
