% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R,
%   R/gen-namespace-examples.R, R/gen-namespace.R
\name{torch_unbind}
\alias{torch_unbind}
\title{Unbind}
\usage{
torch_unbind(self, dim = 1L)
}
\arguments{
\item{self}{(Tensor) the tensor to unbind}

\item{dim}{(int) dimension to remove}
}
\description{
Unbind
}
\section{unbind(input, dim=0) -> seq }{


Removes a tensor dimension.

Returns a tuple of all slices along a given dimension, already without it.
}

\examples{
if (torch_is_installed()) {

torch_unbind(torch_tensor(matrix(1:9, ncol = 3, byrow=TRUE)))
}
}
