% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_unique_consecutive}
\alias{torch_unique_consecutive}
\title{Unique_consecutive}
\description{
Eliminates all but the first element from every consecutive group of equivalent elements.
}
\details{
\preformatted{.. note:: This function is different from :func:`torch.unique` in the sense that this function
    only eliminates consecutive duplicate values. This semantics is similar to `std::unique`
    in C++.
}
}
