% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-namespace-docs.R
\name{torch_where}
\alias{torch_where}
\title{Where}
\arguments{
\item{condition}{(BoolTensor) When True (nonzero), yield x, otherwise yield y}

\item{x}{(Tensor) values selected at indices where :attr:\code{condition} is \code{True}}

\item{y}{(Tensor) values selected at indices where :attr:\code{condition} is \code{False}}
}
\description{
Return a tensor of elements selected from either \code{x} or \code{y}, depending on \code{condition}.

\code{torch_where(condition)} is identical to
\code{torch_nonzero(condition, as_tuple=True)}.
}
\details{
The operation is defined as:

\deqn{
    \text{out}_i = \begin{cases}
        \text{x}_i & \text{if } \text{condition}_i \\
        \text{y}_i & \text{otherwise} \\
    \end{cases}
}
}
\note{
\preformatted{The tensors `condition`, `x`, `y` must be :ref:`broadcastable <broadcasting-semantics>`.
}

\preformatted{See also :func:`torch_nonzero`.
}
}
\section{Signatures}{


where(condition, x, y) -> Tensor



where(condition) -> tuple of LongTensor
}

\examples{
\dontrun{
x = torch_randn(3, 2)
y = torch_ones(3, 2)
x
torch_where(x > 0, x, y)
}

}
